Fixes a few ferries being unloadable [MDB IGNORE] (#20362)
* Fixes a few ferries being unloadable (#74504)

## About The Pull Request

Centcom's ferry port was built for the bare minimum, and so did not
allow for the absolute MAXIMUM

This meant shuttles like meat or lighthouse were undockable. s dumb.

Changing the bounds (particularly the width) does mean that shuttles
could in theory clip the edge of centcom, but just like, don't 4head

## Why It's Good For The Game

If we're gonna have ferries you should be able to load them

## Changelog
:cl:
admin: All centcom ferries are now loadable. Yes this WASbroken
/:cl:

* Fixes a few ferries being unloadable

---------

Co-authored-by: LemonInTheDark <58055496+LemonInTheDark@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>