Allows PKAs to be attached to mining MODsuits [MDB IGNORE] (#20460)
* Allows PKAs to be attached to mining MODsuits (#74601)

## About The Pull Request

like the title says, this PR allows miners to attach their funny little
proto-kinetic accelerator to their hard-earned modsuit

## Why It's Good For The Game

this is probably an oversight, but if not, it's a really weird design
choice. if you can attach a crusher to it, there's no reason you
wouldn't be able to attach a PKA. a very dumb inconvenience imo

## Changelog

:cl:
balance: PKAs can now be attached to mining MODsuits.
/:cl:

* Allows PKAs to be attached to mining MODsuits

---------

Co-authored-by: Vile Beggar <edmir995@gmail.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>