Material Container screentips [MDB IGNORE] (#20472)
* Material Container screentips (#74626)

## About The Pull Request

The material_container component now also grants whatever it is added to
a simple screentip of "Insert" on items that it can accept, if it is a
material container that can accept held items.

## Why It's Good For The Game

Now when a traitor feeds their 13 TC revolver into the autolathe they
will see a screentip for a few blissful seconds before pressing left
click anyway and sending an ahelp begging for their revolver back.

## Changelog

:cl:
qol: material components (such as autolathes) have contextual screentips
if you can put an item inside of it
/:cl:

* Material Container screentips

---------

Co-authored-by: Sol N <116288367+flowercuco@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>