Makes uprooting and digging out plants use tool_behaviour instead of just typepath of shovel. [MDB IGNORE] (#20494)
* Makes uprooting and digging out plants use tool_behaviour instead of just typepath of shovel. (#74630)

## About The Pull Request
Closes https://github.com/tgstation/tgstation/issues/74564.

You can now uproot plants with any item that has TOOL_SHOVEL flag
instead of just item/shovel. So you can use trench tool for this
purpose.
And the same with digging out plants. It was originally limited to just
spade shovel, which is rather strange considering that there is bone
shovel recipe and one of it's ingridients is spade.

Changed icon of omnitool's shovel because spade won't be the only shovel
used for uprooting plants.

Also removed a bunch of code and it didn't affect anything in my tests
so I'm assuming it was unused and completely useless.
Maybe i'm not right and it affects something tho.
## Why It's Good For The Game
Trench tool can be used as a normal shovel.
## Changelog
:cl:
fix: You can now uproot and dig out plants with any shovel and not just
spade.
/:cl:

* Makes uprooting and digging out plants use tool_behaviour instead of just typepath of shovel.

---------

Co-authored-by: Helg2 <93882977+Helg2@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>