Removes spectral saxophone from lavaland ruin [MDB IGNORE] (#20549)
* Removes spectral saxophone from lavaland ruin (#74717)

## About The Pull Request

The spectral saxophone is an item which you can use to turn another
person into a skeleton as many times as you want, with flavour text
telling them to go and convert other people into skeletons (although
they are not an antagonist, so this isn't binding).

It spawns during halloween, or _very rarely_ in maintenance at random.

Admins have been upset for the past week about how this item has
suddenly appeared in multiple rounds despite its supposed rarity,
causing disruption they did not particularly like, the culprit is this
recently revamped lavaland ruin which has one mapped into it, making it
available much more often.

I have replaced it with a normal saxophone and a commemorative skull
hat.

Thanks @ necromanceranne for spotting it.

## Why It's Good For The Game

This ruin appears quite frequently and this item isn't supposed to be so
frequently available.

## Changelog

:cl:
fix: Removes the spectral trombone from the lavaland pizza party ruin.
/:cl:

* Removes spectral saxophone from lavaland ruin

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>