Text that appears when you have lost or destroyed your uplink now has information on replacement uplinks [MDB IGNORE] (#20596)
* Text that appears when you have lost or destroyed your uplink now has information on replacement uplinks (#74679)

## About The Pull Request

![image](https://user-images.githubusercontent.com/116288367/231466073-d4613595-a180-4288-94f0-6f7993820a48.png)
the text that shows up if you mulch your uplink or if you got bugged and
didnt spawn with one now tells you how to get a replacement uplink
instead of covering up information that you need to get your replacement
uplink.

fixes #74637

## Why It's Good For The Game

If your uplink has been destroyed in some way that is when this
information is most relevant, and should be highlighted and put on the
forfront.

## Changelog

:cl:
qol: text that appears when you dont have an uplink in the traitor panel
now tells you the codes for your replacement uplink
/:cl:

* Text that appears when you have lost or destroyed your uplink now has information on replacement uplinks

---------

Co-authored-by: Sol N <116288367+flowercuco@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>