Added nukeops victory state for failing nuke but hijacking the shuttle [MDB IGNORE] (#20591)
* Added nukeops victory state for failing nuke but hijacking the shuttle (#74639)

## About The Pull Request

Added a win state for failing to nuke the ship, but managing to hijack
the emergency shuttle, with disk securing or not as small alternatives.

## Why It's Good For The Game

I almost saw this happen earlier today and it amused me enough to make a
PR on the off-chance it actually happens sometime ingame, like the
in-code victory state for the crew managing to nuke the syndie base.

## Changelog

:cl:
add: Added a nukeops victory state for failing to nuke the station, but
somehow hijacking the shuttle.
/:cl:

---------

Co-authored-by: san7890 <the@ san7890.com>

* Added nukeops victory state for failing nuke but hijacking the shuttle

---------

Co-authored-by: carlarctg <53100513+carlarctg@users.noreply.github.com>
Co-authored-by: san7890 <the@ san7890.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>