Cyborg Rechargers now restock with metal and glass from the ore silo, and no longer grant the materials for free. [MDB IGNORE] (#20613)
* Cyborg Rechargers now restock with metal and glass from the ore silo, and no longer grant the materials for free. (#74770)

## About The Pull Request
- Cyborg chargers, officially "Recharging Stations", now have an ore
silo link, set by default and linkable via multitool as one would
expect.
- To restock a cyborg's internal metal and glass tools, the charger will
now pull from the silo, until either the cyborg is fully restocked, or
the silo is empty.
- Restocking is started by right-clicking the silo while a borg is in
the charger, and the examine text mentions this. The borg can
right-click their own charger, of course.
- Other cyborg resources, such as medical borg reagents, are untouched,
and will refill automatically as before.
## Why It's Good For The Game
Engineer borgs, specifically, can just print iron (usually in the form
of rods), and then obtain more for free in the charging stations. It's a
bit dumb, and though I'm sure every borg player has taken advantage of
it at some point, it's time to retire this technique. If the station is
low on metal, the solution should be a mining borg, not an engineering
one.

The charging stations do not auto-restock metal and glass, because the
borg player may not wish to pull materials from the silo if they know
the station is running low. Other resources that do not require mats
will auto-resupply, as there is no reason to not do so.

Restocking speed is 1/8th the max storable material (3750u) per charge
cycle. Balloon alerts are in place for each restock cycle to give visual
feedback, and it also plays the same sound effect as when you load ammo
into a mech.

![image](https://user-images.githubusercontent.com/37497534/232305414-fcdec239-4720-4df8-b064-8e195bc374fe.png)
## Changelog
:cl:
balance: Cyborg Rechargers now restock with metal and glass from the ore
silo, and no longer grant the materials for free.
/:cl:

* Cyborg Rechargers now restock with metal and glass from the ore silo, and no longer grant the materials for free.

---------

Co-authored-by: zxaber <37497534+zxaber@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>