Bio hoods now only have FOV if it covers your face. [MDB IGNORE] (#20626)
* Bio hoods now only have FOV if it covers your face. (#74766)

## About The Pull Request

Bio hoods now only give you an FOV if it hides your identity. The only
one that doesn't currently, is the plague doctor's hat.

The original PR that added the FOV argued that it had to be added
because it was replacing gas masks as ways to hide your identity, but
the plague doctor helmet doesn't, so I think it was unintentional that
it was added to these as well.

## Why It's Good For The Game

The plague doctor's helmet is considered a "Bio suit" but it doesn't
hide your face, so it doesn't make sense that it gives you an FOV. It
moreso just makes the costume less favored by Clown/Mime players and
encourages it to solely be used as "just another biosuit".

## Changelog

:cl:
fix: Plague doctor hats no longer give you an FOV.
/:cl:

* Bio hoods now only have FOV if it covers your face.

---------

Co-authored-by: John Willard <53777086+JohnFulpWillard@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>