Fix med analyzer sometimes cutting off the top of the health report [MDB IGNORE] (#20648)
* Fix med analyzer sometimes cutting off the top of the health report (#74833)

This text should be concatenated to the health report, not replace it
## About The Pull Request
Fixes a simple bug/oversight in the health analyzer code

## Why It's Good For The Game
It makes a thing work properly.

## Changelog
:cl:iain0
fix: Fixes an error in health analyzers which would cut off the top of
the health scan if the player was deaf.
/:cl:

* Fix med analyzer sometimes cutting off the top of the health report

---------

Co-authored-by: Iain Price <github@predestined.net>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>