Refactors genetic damage component to be a status effect [MDB IGNORE] (#20671)
* Refactors genetic damage component to be a status effect (#74799)

## About The Pull Request

The genetic damage component looked like a status effect, swam like a
status effect, quacked like a status effect, but wasn't a status effect.

Irradiated component is also guilty of this, but it has the excuse of
also getting applied to items. This one only applies to mobs though,
so...

## Why It's Good For The Game

Easier to maintain code, that's about it.

## Changelog

Not player facing.

* Refactors genetic damage component to be a status effect

---------

Co-authored-by: ChungusGamer666 <82850673+ChungusGamer666@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>