Fixes silliness involving soup pots clearing ingredients list [MDB IGNORE] (#20700)
* Fixes silliness involving soup pots clearing ingredients list  (#74882)

## About The Pull Request

Fixes #74866

The issue is misleading, it never deleted the disk. The disk remained in
the pot's contents but was no longer tracked as an ingredient, leaving
it stuck

Foolishly I had a case for indestructible things when soup reactions
finished, where it wouldn't delete stuff like the disk.
But I never removed it from the pot. I just nulled the ingredients list.
Oops.

So pots will drop ingredients that fail to cook into anything when the
reaction finishes.

Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>