Actually anti glow [MDB IGNORE] (#20709)
* Actually anti glow (#74733)

## About The Pull Request
Now that I've realized I can fix things that's bugging me, a personal
pet peeve, anti-glow weren't actually doing as advertised, instead just
making it a light grey.

This is anti-glow in it's current state

![image](https://user-images.githubusercontent.com/126404225/232171576-48584926-a47f-49c7-b0f3-449687bb1d47.png)

And this is anti-glow as it probably should be.

![image](https://user-images.githubusercontent.com/126404225/232171605-70fa6da1-841e-438f-a302-911517b95a52.png)
## Why It's Good For The Game
Anti-glow does as advertised.
## Changelog
:cl:
fix: Anti-glow actually bringing some darkness instead of just a light
glow.
/:cl:

* Actually anti glow

---------

Co-authored-by: Dawnseer <126404225+Dawnseer@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>