Fixes yawns not propagating [MDB IGNORE] (#20714)
* Fixes yawns not propagating (#74909)

## About The Pull Request
#62639 was supposed to make it so yawning had a chance to make other
nearby people yawn, depending on RNG and how close they were to you. It
was also meant as a way to sniff out if someone had maybe recently
examined you, by logic that watching someone yawn usually makes you
yawn.

I eff'd up and used a wrong variable, making it so no one would ever
propagate a yawn, because it would check when the original yawner last
yawned (obviously which was immediately) instead of the person being
queried for a yawn.
## Why It's Good For The Game
Makes something that got merged forever ago actually work
## Changelog
:cl: Ryll/Shaps
fix: Yawning will now have a chance of making people near you yawn as
well, and is guaranteed to make them yawn if they examined you in the
last 2 seconds. It was supposed to already do that, but now it'll
actually do that
/:cl:

* Fixes yawns not propagating

---------

Co-authored-by: Ryll Ryll <3589655+Ryll-Ryll@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>