Turn TactiCOOL turtlenecks into TactiCAL using a washing machine and a telecrystal [MDB IGNORE] (#20745)
* Turn TactiCOOL turtlenecks into TactiCAL using a washing machine and a telecrystal (#74895)

## About The Pull Request

What is says in the PR title. Take a tacticool turtleneck or
turtleskirt, throw it into washing machine with a telecrystal, and hey
presto, tactical turtleneck.

## Why It's Good For The Game

I noticed some folk were confused how they were meant to get a tactical
turtleneck for a Regal Condor. This provides tots a method of getting
one without going to space, at the expense of 1 tc.

## Changelog
:cl:
add: You can make a tactical turtleneck or turtleskirt by putting a
telecrystal and a tactiCOOL varient of the turtleneck into a washing
machine. Watch it change before your eyes!
/:cl:

* Turn TactiCOOL turtlenecks into TactiCAL using a washing machine and a telecrystal

---------

Co-authored-by: necromanceranne <40847847+necromanceranne@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>