Fixes the biogenerator displaying item costs that were being divided by the machine's efficiency twice [MDB IGNORE] (#20744)
* Fixes the biogenerator displaying item costs that were being divided by the machine's efficiency twice (#74939)

## About The Pull Request
It was really bugging me that I would see a price being displayed in my
biogenerator, only for it to then cost me far more than what was
displayed. Looking into it, I realized that it was simply an issue with
it dividing the cost by the efficiency in the TGUI, when that math was
already done DM-side.

## Why It's Good For The Game
Being charged the cost you see on the machine is good.

## Changelog

:cl: GoldenAlpharex
fix: Biogenerators now display the accurate price of their products
after having been upgraded, as they no longer visually apply the
efficiency discount twice.
/:cl:

* Fixes the biogenerator displaying item costs that were being divided by the machine's efficiency twice

---------

Co-authored-by: GoldenAlpharex <58045821+GoldenAlpharex@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>