Fixes medisim combatants falling to the lower z-level on Icebox [MDB IGNORE] (#20760)
* Fixes medisim combatants falling to the lower z-level on Icebox (#74947)

## About The Pull Request

This adds a subtype of the chasm turf, which skips over attempting any
z-level stuff and skips to the Part Where It Kills You.

This replaces the chasms in the medisim shuttle, because knights kept
falling down the chasms and escaping containment.
## Why It's Good For The Game

Closes #56695.
## Changelog
:cl: Rhials
fix: Falling down one of the stage hazard chasms on the medisim shuttle
will now properly kill you on Icebox, instead of dropping you into the
z-level below.
/:cl:

* Fixes medisim combatants falling to the lower z-level on Icebox

---------

Co-authored-by: Rhials <Datguy33456@gmail.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>