Fixes crafting tools not taking into account subtypes [MDB IGNORE] (#20807)
* Fixes crafting tools not taking into account subtypes (#74971)

## About The Pull Request
Does what #74968 intended to do, it fixes crafting tools checking if the
required item is a subtype of the available items, instead of the other
way around. What this meant is for example

![image](https://user-images.githubusercontent.com/23585223/234320690-aa0b2c2d-a8c6-447b-be32-3554fa8310c8.png)
this wouldnt work, even though cheap lighters are a subtype of lighters,
as /obj/item/lighter/greyscale
but if i instead spawned the base /obj/item, which /obj/item/lighter is
a subtype of

![image](https://user-images.githubusercontent.com/23585223/234320959-bd388520-7d2b-4028-abab-a7638ba4adda.png)
it would work. obviously this is funnily broken

## Why It's Good For The Game
Bug gone!

## Changelog
:cl:
fix: Fixes crafting tools not taking into account subtypes, i.e. you can
craft a filet migrawr with any lighter.
/:cl:

* Fixes crafting tools not taking into account subtypes

---------

Co-authored-by: Fikou <23585223+Fikou@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>