Crashedship Space Ruin - Replaces the away mission areas with ruin ones [MDB IGNORE] (#20838)
* Crashedship Space Ruin - Replaces the away mission areas with ruin ones (#74991)

## About The Pull Request

A little bit of an odd annoyance i've had with this particular ruin,
mostly since due to it having away mission areas it teleblocked you from
attempting to rebuild and use the one there. Replaced the away mission
areas with ruin has_grav ones.

![image](https://user-images.githubusercontent.com/22140677/234639172-5e4898fe-bd34-431c-b918-ee3d34a42764.png)

![image](https://user-images.githubusercontent.com/22140677/234639182-3f3b3191-0719-4094-99b3-a559f99fe72a.png)

Area regions are the same with the exception of the away ones that were
in space just being converted to space like normal ruins are

## Why It's Good For The Game

Let's spessmen utlize the teleporter that was left behind, giving the
option that really should be there if the area wasnt being silly.

## Changelog
:cl:Zergspower
fix: Crashedship Ruin - You can now use the teleport
/:cl:

* Crashedship Space Ruin - Replaces the away mission areas with ruin ones

---------

Co-authored-by: Zergspower <Griffinj88@yahoo.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>