Fixes the Chef not having CQC outside the kitchen on Tramstation [MDB IGNORE] (#20920)
* Fixes the Chef not having CQC outside the kitchen on Tramstation (#75101)

The additional areas were broken in #68084 because the area around the
bar was changed from /area/station/service/kitchen/diner to
/area/station/commons/lounge. However, the cqc addtional areas was never
updated to reflect this change in the area.

It might be a bit big, but compared to how big the area is on other maps
I think it's fine.

Additionally, removes additional cqc areas being defined as
/area/station/service/kitchen/diner for icebox for a minor code
improvement, it was pointless since the diner is a subtype of kitchen

* Fixes the Chef not having CQC outside the kitchen on Tramstation

---------

Co-authored-by: BlueMemesauce <47338680+BlueMemesauce@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>