Fixes Mech Charger Console to Recharge Port Sync [MDB IGNORE] (#20936)
* Fixes Mech Charger Console to Recharge Port Sync (#75127)

## About The Pull Request
Found a bug while working on another project, Kapu on the Discord told
me why it was broken. For most maps, this seems to be handled by the
fallback code by chance, where it checks within one more tile away, but
for Tram, that caused the following;

![image](https://user-images.githubusercontent.com/37497534/235795350-4e2326af-37ac-4f29-b195-b286eeafac03.png)

This is a webedit, but only because I didn't want to stop mid-project
and swap branches. It has been tested on local.
## Why It's Good For The Game
Bugfix.
## Changelog
:cl:
fix: Mech chargers on Tram (and possibly elsewhere) no longer sync
oddly.
/:cl:

* Fixes Mech Charger Console to Recharge Port Sync

---------

Co-authored-by: zxaber <37497534+zxaber@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>