Adds another EVA suit to the psyker shuttle [MDB IGNORE] (#20996)
* [NO GBP] Adds another EVA suit to the psyker shuttle (#75207)

## About The Pull Request

I forgot to give the Psyker Seer their own suit storage/suit.

I didn't know where to throw it in, so I just expanded the alcove I
threw the Seer into. This was accomplished by moving the entire bathroom
one tile to the right.

They also get a pile of garbage and ants in their little alcove, to make
it feel more homey.

old:

![image](https://user-images.githubusercontent.com/28870487/236648256-6659d780-fa38-40f5-a98b-74de445e3bb6.png)

new:

![image](https://user-images.githubusercontent.com/28870487/236648254-2cf5eb1e-47e3-4756-9178-8ef933a94565.png)

This changes nothing else, other than making the armory/bathroom hallway
a bit more spacious as a side-effect.
## Why It's Good For The Game

The Seer might not be a psyker, but that doesn't mean they aren't part
of the family. They deserve a suit too!
## Changelog
:cl:
fix: The psyker shuttle now has a suit for the psyker seer.
/:cl:

* [NO GBP] Adds another EVA suit to the psyker shuttle

---------

Co-authored-by: Rhials <Datguy33456@gmail.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>