Outfit manager now searches more thoroughly for the user's ID (#75310) (#21071)
Outfit manager now searches more thoroughly for the user's ID (#75310)

The outfit manager will now search more thoroughly for a user's ID when
attempting to change its trim. This prevents a runtime when attempting
this on a user with a PDA/Wallet/non-ID item in their ID slot.

The instance causing problems in my case was the nuclear operative
outfit -- Attempting to make a player a nuclear operative from the
player panel while they are wearing a wallet would fuck up the
outfitting process.

If you are wearing a wallet it will modify the trim of whichever ID is
in the wallet's first slot.

Co-authored-by: Rhials <Datguy33456@gmail.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>