Grilling items now only smoke when the griddle is on [MDB IGNORE] (#21097)
* Grilling items now only smoke when the griddle is on (#75265)

## About The Pull Request

Prior to this change, if you put an item on the griddle, it immediately
began to smoke, regardless of whether the griddle was on or not. Now
smoke will only appear when the griddle is turned on.

## Why It's Good For The Game
Graphic QoL for chefs

https://user-images.githubusercontent.com/10997188/236855447-8912a689-bec3-4cba-a6f3-45c428d7af29.mp4

## Changelog
:cl:
fix: fixed griddle code so that the smoke over the grilling items
appears only when the griddle is on
/:cl:

* Grilling items now only smoke when the griddle is on

---------

Co-authored-by: Vallat <vox.crit2013@gmail.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>