Golem mineral functionality tweaks/fixes [MDB IGNORE] (#21121)
* [no gbp] Golem mineral functionality tweaks/fixes (#75343)

## About The Pull Request

Fixes a couple of things people have pointed out about golem
transformations.

- Diamond golems now stop being invisible when they attack or throw
something, they also turn invisible a bit slower.
- Using a bluespace knot takes 2 seconds instead of 3 seconds.
- Bananium Golems only slip you if they are lying down.

In order to achieve that last one I refactored the slipperiness
component to take an optional extra callback, and then killed a subtype
of it which could be replaced with passing in a callback. I tested it
and it seems to work the same as it used to.

These are largely how things were supposed to work and I just overlooked
them.
I am sure this won't be the last PR of a similar vein while people try
these out, provided that I actually hear anything they are saying about
it.

## Why It's Good For The Game

Diamond golems shouldn't be able to attack you while remaining invisible
and untargetable even if it is funny.
Clown golems aren't supposed to be able to slip you by swapping places
with you even if it is funny.
The bluespace hand was basically just worse than using the crystal and
not eating it, maybe still needs another buff after this one.

## Changelog

:cl:
fix: Diamond Golems can no longer attack or throw things and remain
invisible.
fix: Bananium Golems are only slippery if you actually tread on them
(aka: while they are resting).
balance: Golem bluespace teleportation is slightly quicker.
/:cl:

---------

Co-authored-by: san7890 <the@ san7890.com>

* [no gbp] Golem mineral functionality tweaks/fixes

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>
Co-authored-by: san7890 <the@ san7890.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>