Warn admins about active Internet sounds [MDB IGNORE] (#21187)
* Warn admins about active Internet sounds (#75400)

## About The Pull Request

For various reasons an admin may not be aware there is currently an
Internet sound playing. (Usually TGUI related or just recently logged
in.)

This provides a warning and optional override if they try to start an
Internet sound while one is already running.

https://github.com/tgstation/tgstation/assets/83487515/af4197c9-bf9d-43e4-a407-a556b88a6d30

## Why It's Good For The Game

Less "HEY I WAS ALREADY PLAYING SOMETHING!"

## Changelog

:cl: LT3
admin: Admins are now warned if they are going to override a currently
playing admin midi
/:cl:

* Warn admins about active Internet sounds

---------

Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>