Fix very common smoothing groups runtime by deferring icon smoothing until all atom initializations are complete [MDB IGNORE] (#21229)
* Fix very common smoothing groups runtime by deferring icon smoothing until all atom initializations are complete (#75444)

## About The Pull Request
Fixes the "bad index" runtime that shows up on 80% of rounds.

![image](https://github.com/tgstation/tgstation/assets/35135081/8b8a7b6a-90b6-4c07-8e3b-47f31e4864c5)

The issue is that during dynamic maploading, the initialization of
objects is split over several ticks. However, before the initialization
is complete, an object's smoothing groups is still in its string form
(as per [the smoothing group
optimizations](https://github.com/tgstation/tgstation/pull/71989)).
Thus, code later down the line attempts to read from this string, and
errors.

It would be expensive to check all nearby neighbors every time in any
place in this code, so instead we defer icon smoothing if there are any
initializations currently in progress. Because this only happens for
dynamic loading, the stuff being loaded dynamically is usually small,
and icon smoothing already has a somewhat visible delay for shuttles,
this delay doesn't matter.

## Changelog

:cl:
fix: Fixed an issue where objects on something that loaded dynamically,
like shuttles, would not be smoothed.
/:cl:

* Fix very common smoothing groups runtime by deferring icon smoothing until all atom initializations are complete

---------

Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>