Runtimes in mapping helpers [MDB IGNORE] (#21288)
* [NO GBP] Runtimes in mapping helpers (#75473)

## About The Pull Request

There were some attempts to get an area of an object that is confirmed
to be null in the condition, which resulted in runtimes when the helper
couldn't find an object.

Also ensures that the window spawner places window before the helper
tries to find it.

And updates damaged window integrity roll thresholds to guarantee
cracks.

## Why It's Good For The Game

No runtimes

## Changelog

:cl:
fix: fixed possible issues with apc, airalarm and damaged
machinery/windows helpers
/:cl:

* [NO GBP] Runtimes in mapping helpers

---------

Co-authored-by: Andrew <mt.forspam@gmail.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>