Adds missing wrappers to request consoles messages [MDB IGNORE] (#21290)
* [NO GBP] Adds missing wrappers to request consoles  messages (#75503)

## About The Pull Request

Request consoles messages and announcement were not trimmed and
sanitized properly. This PR fixes this, sorry about that.

Sadly, does not solve #75494. Is there a proc that trims excess
whitespace _inside_ messages? There are many places that needs this.

## Why It's Good For The Game
Whitespace after and before texts should be trimmed, and HTML stuff
should be encoded and stripped properly for safety.

## Changelog

:cl:
fix: fixes request console messages not being trimmed and encoded
properly
/:cl:

* [NO GBP] Adds missing wrappers to request consoles  messages

---------

Co-authored-by: Profakos <profakos@gmail.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>