Fixes glove offsets not working with negative numbers [MDB IGNORE] (#21434)
* Fixes glove offsets not working with negative numbers (#75638)

## About The Pull Request

This was changed in the Golem rework -
https://github.com/tgstation/tgstation/pull/74197

If you have a species that have arms that go longer than normal, you can
set the offsets of in-hands to negative y value to lower it. This isn't
an option for gloves however, because it starts at 0 and only increases
if the value is larger.
This fixes it by letting it set a negative number, and comparing your
other hands' offsets to that.

## Why It's Good For The Game

Fixes species that have arms that go longer than normal.

## Changelog

:cl:
fix: Species that have longer arms than normal can now properly set
their glove offset to go low and meet their hand.
/:cl:

* Fixes glove offsets not working with negative numbers

---------

Co-authored-by: John Willard <53777086+JohnFulpWillard@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>