Newly created areas are no longer cult permitted [MDB IGNORE] (#21503)
* Newly created areas are no longer cult permitted (#75589)

## About The Pull Request

This is a continuation of
https://github.com/tgstation/tgstation/pull/73757 because I messed up
and didn't bother removing this as well.
The reasoning for it is the same as that, so this is basically a copy
paste of it.

## Why It's Good For The Game

One of the drawbacks of Cult is that they get stronger throughout a
round, at the cost of slowly becoming easier to spot due to their base
locations, so being to completely bypass the negative part makes it
unfair to fight cult if there's a single player between them that can
take the aux base's mats and make a 4x4 rom in space or on icemoon.

Finally closes https://github.com/tgstation/tgstation/issues/47747

## Changelog

:cl:
balance: Newly created areas in Icemoon and Space are no longer valid
Cult areas for Runes and such.
/:cl:

* Newly created areas are no longer cult permitted

---------

Co-authored-by: John Willard <53777086+JohnFulpWillard@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>