Makes warehouses have lights turned off from roundstart. Also adds mapping helper for turning off lightswitch. [MDB IGNORE] (#21487)
* Makes warehouses have lights turned off from roundstart. Also adds mapping helper for turning off lightswitch. (#75666)

## About The Pull Request
Warehouse is roundstart closed with shutters and doesn't meant to spawn
in, as far as i know. Logic is simple: why do you have turned on lights
in a room that is barely used.

Basically every warehouse with a lightswitch now has lights turned off.
Northstar has all the lights broken in it, so i don't know if i should
add one to it.

I added a lightswitch on Birdshot so it looks like this:

![image](https://github.com/tgstation/tgstation/assets/93882977/8fc8dc44-517e-4491-8d97-98ef411d18e5)
And a lightswitch to Tram's warehouse:

![image](https://github.com/tgstation/tgstation/assets/93882977/c3f43629-ea8a-4f62-830c-498b0ed55121)
Also added some lights on Metastation so it looks like this:

![image](https://github.com/tgstation/tgstation/assets/93882977/754d1970-0888-4564-b99c-42d93c237357)

![image](https://github.com/tgstation/tgstation/assets/93882977/3f18d30f-2d7f-4f3b-9845-9da61d1f6e77)

I made lights to turn off with mapping helper. Suggestions accepted.

I also forgot to make a fridge free accessed on the Beach, so i'm
changing it with this pr.
## Why It's Good For The Game
Less electricity spending for no4raisin.
## Changelog
:cl:
add: [Birdshot], [Tram] Added a lightswitch to warehouse.
add: [Birdshot], [Deltastation], [Metastation], [Tram] and [Icebox]
warehouses now have lights off roundstart.
fix: [Metastation] Warehouse now has lights. Finally.
/:cl:

* Makes warehouses have lights turned off from roundstart. Also adds mapping helper for turning off lightswitch.

---------

Co-authored-by: Helg2 <93882977+Helg2@users.noreply.github.com>
Co-authored-by: Jolly-66 <70232195+Jolly-66@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>