Stops factory blobs from hard deleting. [MDB IGNORE] (#21523)
* Stops factory blobs from hard deleting. (#75676)

## About The Pull Request
The code for factory blobs and blob structures is kinda messy and ridden
with some unreacheable code and unused variables. My plan here is stop
factory blobs from hard-deleting because of uncleared references, but I
took a little liberty to cut some old code that looks more of a
hindrance than anything.

## Why It's Good For The Game
This should fix #74845.

## Changelog

No tangible change for the player.

* Stops factory blobs from hard deleting.

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>