Modular Shield Generator Fixes [MDB IGNORE] (#21527)
* [NO GBP] Modular Shield Generator Fixes  (#75705)

## About The Pull Request
This pr is a small fix for 2 things.

1. It will now stop you from being able to pulse the wire on an
unpowered generator to activate it, this bug wasnt gamebreaking because
regeneration is off when the generator has no power but it still
shouldnt be functional at all.
2. The external setting on the shield generator now just blacklists
constructed floor turfs to make it compatible with outdoor areas such as
lavaland or an asteroid.

## Why It's Good For The Game
The inability to use the external function on maps such as icebox meant
that you would be stuck with the alternative which can block internal
passageways within whatever structure you built the shield generator in,
this fixes that and makes the shield generator more compatible with
future maps that take place on a planet or asteroid.
## Changelog
:cl:
fix: You can no longer pulse the wire on an unpowered shield generator
to activate shields
fix: External only setting on modular shield generators now properly
uses terrestrial turfs such as lavawastes.
/:cl:

* [NO GBP] Modular Shield Generator Fixes

---------

Co-authored-by: moocowswag <62126254+moocowswag@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>