Makes syndi-mod in Tactical NukeOps Metagame Syndi-kit unrestricted with access. [MDB IGNORE] (#21541)
* Makes syndi-mod in Tactical NukeOps Metagame Syndi-kit unrestricted with access. (#75696)

## About The Pull Request
Basically someone forgot to make it unrestricted or add agent's card to
the kit so it would be usable when mods were added.
## Why It's Good For The Game
1 more thing that works as intended?
## Changelog
:cl:
fix: Syndicate mod-suit in Tactical NukeOps Metagame Syndi-kit (the one
with bulldog shotgun) now doesn't require any access, so you don't need
to buy agent card to simply use it.
/:cl:

* Makes syndi-mod in Tactical NukeOps Metagame Syndi-kit unrestricted with access.

---------

Co-authored-by: Helg2 <93882977+Helg2@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>