Refactors Ant into Basic Mob [MDB IGNORE] (#21601)
* Refactors Ant into Basic Mob (#75662)

## About The Pull Request

On the tin. No new fancy AI wheels or anything, just a simple port.
## Why It's Good For The Game

Knocks another one off the list, just really light stuff. They're a bit
smarter now too, I think the intention was them for them to be pet-like
as well (according to the code). Should be really rather easy to give
them the pet-like behaviors and elements if someone really wants to in
the future, just sorta paving the way for more work to be done to make
mobs more intricate/interesting.
## Changelog
:cl:
refactor: Giant ants are now more capable of distinguishing friend and
foe.
/:cl:

* Refactors Ant into Basic Mob

---------

Co-authored-by: san7890 <the@san7890.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>