Makes transparent kudzu transparent on spawn, not on vine growth [MDB IGNORE] (#21752)
* Makes transparent kudzu transparent on spawn, not on vine growth (#75942)

## About The Pull Request
Makes transparent kudzu transparent on spawn, not on vine growth.
Currently, they only become transparent after growing in place and not
after mutating, as it should. Timid allows this, as timid is low to the
ground - why shouldn't transparent vines?

## Why It's Good For The Game
Transparent vines should be transparent. Transparent is currently the
only mutation that can't be immediately identified by it's change in
appearance; this puts it back in line with the rest. It also blocks
sight until growth, which makes its viability for positive kudzu
questionable when Timid does everything it can but better. Closes #72660

## Changelog
:cl: Licks-The-Crystal
qol: Makes transparent kudzu transparent (and won't block sight)
/:cl:

* Makes transparent kudzu transparent on spawn, not on vine growth

* update modular

---------

Co-authored-by: zeckle/licks-the-crystal <79835169+mikederkan@users.noreply.github.com>
Co-authored-by: Tom <8881105+tf-4@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>