Mecha fixes: plasma generator, internal damage report & repair [MDB IGNORE] (#21713)
* Mecha fixes: plasma generator, internal damage report & repair (#75704)

Cleans up the plasma generator to actually work & make more sense:
* was not accepting any fuel, made that work; will also accept subtypes
of fuel e.g. a mapped-in stack of 5 plasma
* was not displaying the amount of fuel left in mecha stats UI, was
broken in #75052
* plasma pollution was removed a long time ago in #42908, scrubs mention
of that

Also adds some fluff text to the internal damage tracker to the short
circuit entry, as it was lacking a "system nominal" one. Also added a
`to_chat` to fixing navigation internal damage, as it was the only one
with no message for some reason.

* Mecha fixes: plasma generator, internal damage report & repair

---------

Co-authored-by: Sealed101 <cool.bullseye@yandex.ru>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>