Fixes coin values [MDB IGNORE] (#21889)
* Fixes coin values (#76066)

## About The Pull Request

This brings coins values back from being severely undervalued after
#75437. Coins are supposed to be made of and worth 400 units of mats,
but the numbers weren't updated and so they were made of significantly
less than that.

## Why It's Good For The Game

Brings coins back to what they're supposed to be worth.
## Changelog
:cl:
fix: After destroying unfathomable quantities of excess coins, their
value is back to what it's supposed to be!
/:cl:

* Fixes coin values

---------

Co-authored-by: Nick <42454181+Momo8289@users.noreply.github.com>