The Forbidden Tuna Can [MDB IGNORE] (#21878)
* The Forbidden Tuna Can (#76064)

## About The Pull Request

This PR was made to replace the well known uglymine sprite with
something that doesn't look like a coder-sprite filler and instead looks
like an actual landmine. Simple as that, there's not really anymore meat
to this PR.

![Screenshot 2023-06-14
214904](https://github.com/tgstation/tgstation/assets/609886/b5f89661-b28f-4339-904a-db6f5c3bd5ed)
## Why It's Good For The Game

Replaces an older, depreciated sprite with something that has needed an
upgrade for quite a long time now.
## Changelog
:cl:
image: Added a new Land Mine sprite
/:cl:

* The Forbidden Tuna Can

* update modular

---------

Co-authored-by: Youtubeboy139 <jared.klaas@gmail.com>
Co-authored-by: Tom <8881105+tf-4@users.noreply.github.com>