Fixes a runtime when attacking non-carbon mobs with a cuffsnapper [MDB IGNORE] (#21968)
* Fixes a runtime when attacking non-carbon mobs with a cuffsnapper  (#76190)

## About The Pull Request

The cuffsnapper component now checks to make sure the target is a carbon
before running the rest of the cuffsnapping process on them.

target.handcuffed is checked (handcuffed is defined on carbon), but
nowhere is it actually asserted that the target is a carbon, so a
runtime would occur when attacking non-carbon mobs.
## Why It's Good For The Game

This runtime fix was brought to you by https://runtimes.moth.fans
## Changelog
:cl:
fix: attacking non-carbon mobs with a cuffsnapping object will no longer
runtime.
/:cl:

* Fixes a runtime when attacking non-carbon mobs with a cuffsnapper

---------

Co-authored-by: Rhials <28870487+Rhials@users.noreply.github.com>