Changes food storage to use a right-click instead of a left-click, adds context hints [MDB IGNORE] (#21979)
* Changes food storage to use a right-click instead of a left-click, adds context hints (#76110)

What it says on the tin. Currently, if you left-click a cake or bread
with anything small enough (including a knife) you store it inside and
have to remove it to actually use it. Now it uses a right-click which
makes significantly more sense because you're almost always going to be
trying to cut the piece of food and not embedding an object into it and
left-click actions should ideally be "what you're most likely to be
doing with the object 99% of the time".

Also adds context tips for slicing, embedding, and removing embedded
items. It'll always give a prompt to remove an item even if there isn't
one, so there's no concern about it giving away that something is
inside.

* Changes food storage to use a right-click instead of a left-click, adds context hints

---------

Co-authored-by: TheVekter <TheVekter@users.noreply.github.com>