Consistency fixes for one way exits on doors [MDB IGNORE] (#21989)
* Consistency fixes for one way exits on doors (#76175)

## About The Pull Request

Adds one way accesses on science hallway doors on Delta, Icebox, and
Birdshot to make them consistent with other maps.

Also fixed a science maintenance that was set to service maintenance,
which didn't make any difference in testing but it kind of bugged me
seeing it in the mapper.

## Why It's Good For The Game

Avoids people getting trapped in science hallways, and I think
consistency in behavior across all maps for something like this is a
good thing.

## Changelog

:cl:
fix: Science department hallways now have unrestricted one-way exits on
Delta, Birdshot, and Icebox, to make these maps consistent with the
others.
/:cl:

* Consistency fixes for one way exits on doors

---------

Co-authored-by: YehnBeep <86855173+YehnBeep@users.noreply.github.com>
Co-authored-by: Jolly-66 <70232195+Jolly-66@users.noreply.github.com>