Fixes nebula vomiting being in the normal symptom pool [MDB IGNORE] (#22217)
* Fixes nebula vomiting being in the normal symptom pool (#76489)

## About The Pull Request
Fixes #76488.

To my knowledge, these are supposed to be bespoke to the floor
contamination PR, not actual symptoms you can roll in a normal virus. If
@ MTandi corrects me then I'll close this.

## Why It's Good For The Game
Thing not meant to happen happened, this make not happen.

## Changelog
:cl: Vekter
fix: Removed nebula vomiting from the normal symptom pool. It should no
longer appear on regular viruses.
/:cl:

* Fixes nebula vomiting being in the normal symptom pool

---------

Co-authored-by: Vekter <TheVekter@users.noreply.github.com>