Adds Jaywalkers [MDB IGNORE] (#22235)
* Adds Jaywalkers (#76474)

## About The Pull Request

Adds Jaywalking as a prisoner crime
Also moved Mass Murder so it's alphabetically ordered.

## Why It's Good For The Game

I was pretty surprised such a serious crime was somehow overlooked.

But more seriously, I feel it fits the tone of this game to have
something a little silly there. Hopefully the warden and sec won't be
too hard on people put away for this crime.

## Changelog

:cl:
add: Prisoners can now be jaywalkers
qol: reordered a crime so it appears on the choice list alphabetically.
/:cl:

* Adds Jaywalkers

---------

Co-authored-by: YehnBeep <86855173+YehnBeep@users.noreply.github.com>