Illiterate people can send trams to random stations [MDB IGNORE] (#22252)
* Illiterate people can send trams to random stations (#76524)

## About The Pull Request

As with the mining shuttle console, people who can't read can fumble
around on the tram console until they manage to select a station at
random.
You may argue that the menu is largely pictorial anyway, but that's why
it only takes them 5 seconds.

## Why It's Good For The Game

I think the "pressing stuff at random" interaction with the mining
shuttle is cute and I think that being able to utilise the tram map's
central feature in a somewhat aimless way is fun.

## Changelog

:cl:
add: People who cannot read can interact with the tram console to send
it to a random station.
/:cl:

* Illiterate people can send trams to random stations

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>