Knocks the lights out of Tinea Luxor [MDB IGNORE] (#22060)
* [s] Knocks the lights out of Tinea Luxor (#76295)

## About The Pull Request
So, we've a problem where the lights from Tinea Luxor can be staked up
indefinitely, because someone thought it was a brilliant idea to use
expose_mob rather than on_mob_metabolize.

## Why It's Good For The Game
This will close #76263. Aaaand yes, this is a webedit PR.

## Changelog

:cl:
fix: Fixes the lights from Tinea Luxor being stackable to the point of
crashing the game for others.
/:cl:

* [s] Knocks the lights out of Tinea Luxor

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>