AddComponent() now throws a crash message when the comp type arg is not a component. [MDB IGNORE] (#22178)
* AddComponent() now throws a crash message when the comp type arg is not a component. (#76221)

## About The Pull Request
HEY! Did you know that the AddComponent() proc can complete without
throwing errors when the component type argument is actually not a type
or instance of a component? Me neither (until now), but I've recently
discovered that while working on #76219 (Make sure to merge that one
first).

Also adds early return checks to the AddElement and RemoveElement procs.

## Why It's Good For The Game
Silent bugs are awful.

## Changelog
N/A

* AddComponent() now throws a crash message when the comp type arg is not a component.

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>