Fix Too Slowing people with high fives [MDB IGNORE] (#22086)
* Fix Too Slowing people with high fives  (#76277)

## About The Pull Request

At some point with the refactors to offering it was made so that
dropping the item stops the offer, unfortunately too slowing people with
high fives relied on this behavior (dropping not stopping the offer).

Restores that behavior with a bit more code tweaking.

## Why It's Good For The Game

How can I be too slow?

## Changelog

:cl: Melbert
fix: You can once again "too slow" someone with a high five
/:cl:

* Fix Too Slowing people with high fives

* d'oh!

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>
Co-authored-by: Pinta <68373373+softcerv@users.noreply.github.com>