fixes icebox genetics having a science headset instead of a medisci headset [MDB IGNORE] (#22503)
* fixes icebox genetics having a science headset instead of a medisci headset (#76857)

## About The Pull Request

currently of all the maps in rotation other than birdshot icebox is the
only map wherein the genetics office is lacking a medisci headset for
reasons unknown, unlike birdshot however icebox is meant to be a full
station that was properly put together whereas the theme for birdshot
would lend it some credibility to NOT having the medisci headset.

## Why It's Good For The Game

consistency issues are annoying, and as a genetics player myself this is
one i know has probably irked some other people as well.

## Changelog

:cl:
fix: replaces icebox genetics' regular science headset with a medisci
headset, making it more in line with other genetics offices
/:cl:

* fixes icebox genetics having a science headset instead of a medisci headset

---------

Co-authored-by: Donglesplonge <120208006+Donglesplonge@users.noreply.github.com>