fixes some weird shading on the plasma tank sprite [MDB IGNORE] (#22619)
* fixes some weird shading on the plasma tank sprite (#76988)

## About The Pull Request
a little bit ago, i resprited a lot of basic tanks. well, the plasmatank
i had to do from the ground up, and it wasn't perfect by any means. so
ryerice kindly did some shading corrections and gave it a cool new
symbol...

![image](https://github.com/tgstation/tgstation/assets/110322848/e8290882-dd33-4b8d-8df8-617f355dd1b8)
and it's a definite step up!
## Why It's Good For The Game
my plasmatank resprite was fairly mid, so hopefully this brings it a
little bit up to snuff.
## Changelog
:cl: CoiledLamb, RyeRice
image: cleans up shading on plasmatank sprite, gives it a cool new
symbol.
/:cl:

* fixes some weird shading on the plasma tank sprite

---------

Co-authored-by: Lamb <110322848+CoiledLamb@users.noreply.github.com>