Assorted changes to mafia map closet contents [MDB IGNORE] (#22669)
* Assorted changes to mafia map closet contents (#77008)

## About The Pull Request

This changes the contents of some lockers on various mafia maps.

**Lavaland Excursion:** The spare mining uniforms in each locker have
been replaced with a pickaxe and shovel.

**Snowdin:** The rooms containing closets now have alternate (mining)
winter coats instead of the standard ones you already spawn with.

**Mothership Zeta:** All closets now contain a gizmo. Wow!

**Syndicate Megastation:** The closets now contain one random piece of
syndie clothing, instead of another tacticool turtleneck.
## Why It's Good For The Game

Some of the gear in the closets kind of became redundant when the map
outfits were added and I just forgot to remove them.

Gives some more stuff for people to fidget with during mafia downtime.
For the maps not touched, they either already had closet contents or
were intentionally left empty.

I'm all but certain the items provided in the closets have no way of
endangering the controlled environment of a mafia game, so that
shouldn't be an issue.
## Changelog
:cl: Rhials
qol: Modifies the contents of some Mafia lockers. Go check 'em out!
/:cl:

* Assorted changes to mafia map closet contents

---------

Co-authored-by: Rhials <28870487+Rhials@users.noreply.github.com>