Adds a negative station trait: Vending products shortage. [MDB IGNORE] (#22746)
* Adds a negative station trait: Vending products shortage. (#77125)

## About The Pull Request
Adds in a mean station trait that randomly lowers the availability of
all vending products from vending machines on the station.

## Why It's Good For The Game
Kicking aside the simple concept of "The more the merrier (until it's a
feature creep)" for a moment, vending (re)supplies are scarcely needed
on a generic round basis. The addition of this new trait could lead to
situations where more vending supplies are wanted/ordered once in a
while.

Beside, this is a pretty evil negative trait notwithstanding its
simplicity, unlike many other relatively innocuous ones.

## Changelog

:cl:
add: Added a 'Vending products shortage' station trait, that randomly
lowers the availability of all vending products from vending machines on
the station, with a 1/20 chance of the vending machine itself being
tilted.
/:cl:

* Adds a negative station trait: Vending products shortage.

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>