Puts a cap to the amount of stickers that can be sticked to an atom. [MDB IGNORE] (#22752)
* Puts a cap to the amount of stickers that can be sticked to an atom. (#77160)

## About The Pull Request
The lack of a limit to the amount of stickers that can be attached on an
atom can lead overlays-related issues, as shown in #76987, likely by
going past the maximum number of overlays an atom can safely have.

The cap will be of 12 stickers per atom, an honest amount if you ask me.

Oh yeah, I've also taken the opportunity to improve the code a smidge I
guess.

## Why It's Good For The Game
This will fix #76987.

## Changelog

:cl:
fix: Put a cap to the amount of stickers that can be sticked to an atom
(12) to prevent icon-related issues.
/:cl:

---------

Co-authored-by: san7890 <the@ san7890.com>

* Puts a cap to the amount of stickers that can be sticked to an atom.

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Co-authored-by: san7890 <the@ san7890.com>