fixes the logic for mob size-related maptext height. [MDB IGNORE] (#22755)
* [NO GBP] fixes the logic for mob size-related maptext height. (#77171)

## About The Pull Request
I've made a hiccup in the logic at the last moment and forgot that the
offset should be half of the size difference from trom the default
identity, not its entirety, this is because scaling is equally
distributed vertically and horizontally, we only have to deal with the
upper half of it.

## Why It's Good For The Game
Fixing a human error.

## Changelog
N/A, it's a barely noticeable whoopsie all in all.

* [NO GBP] fixes the logic for mob size-related maptext height.

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>