Fixing BoH bombing being doable with just one BoH. [MDB IGNORE] (#22839)
* Fixing BoH bombing being doable with just one BoH. (#77243)

## About The Pull Request
BoH-bombing is supposed to require two bags to be achievable.

## Why It's Good For The Game
Fixing an easy and amusing bug, except being the one fixing it fills me
with an ounce of remorse, so you could say it's not all that easy. This
will fix #77204.

## Changelog

:cl:
fix: BoH bombing now correctly requires two bags to be done.
/:cl:

* Fixing BoH bombing being doable with just one BoH.

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>