Fixes formatting for derelict and syndicate drone job names [MDB IGNORE] (#22858)
* Fixes formatting for derelict and syndicate drone job names (#77303)

## About The Pull Request
These were using the generic "ghost role" job names instead of the
defines that it's supposed to. This should fix it.

Note: This touches code I'm not intimately familiar with, but I did test
this and it works fine.

## Why It's Good For The Game
Fixes #76922

## Changelog
:cl: Vekter
fix: Fixes formatting for derelict and syndicate drone job names. They
should no longer tell you they're "the Ghost Role" when spawning as
them.
/:cl:

* Fixes formatting for derelict and syndicate drone job names

---------

Co-authored-by: Vekter <TheVekter@users.noreply.github.com>