Restores some mechs being able to pivot step (move and turn in the same keypress) and diagonal movement [MDB IGNORE] (#22884)
* Restores some mechs being able to pivot step (move and turn in the same keypress) and diagonal movement (#77345)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Most mechs couldn't move diagonally. This restores that for all but the
Clarke (which is a vehicle on treads)

The Clarke and Odysseus once again have pivot step, lost during the
refactor for unknown reasons. It was literally just right there, it was
an easy fix. Clarkes once again move like a tank!

## Why It's Good For The Game

So the way that mech movement works, by blocking diagonal movement, what
instead happens is that the mech eats the move command and goes on
movement cooldown. By pressing two buttons rather than one, you come to
a full stop and no longer can move. I cannot think of a more hostile
player-facing feature than this. It utterly gutted the usability of
mechs for no real apparent benefit. As such, I'm just removing it on all
but the most relevant mech, the Clarke. This should make it much more
approachable to pilot mechs and not feel like the game is malfunctioning
half the time.

Clarkes and Odysseus mechs had a really cool feature. They turned and
moved in the same button press. This made them very nimble mechs and fun
to utilize. They lost this feature in the refactor for, once again, not
much apparent reason since this was literally a line change to restore.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: Restores a lost feature, Clarkes and Odysseus mechs once again can
pivot on a dime and step in the same button press.
qol: Mechs universally have diagonal movement as to avoid unresponsive
movement, as block diagonal movement was having unintended side effects
and was actually intended for pivot step mechs.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

* Restores some mechs being able to pivot step (move and turn in the same keypress) and diagonal movement

---------

Co-authored-by: necromanceranne <40847847+necromanceranne@users.noreply.github.com>