Fixes former blobs being able to potentially respawn when having their mind transferred to a new body [MDB IGNORE] (#22915)
* Fixes former blobs being able to potentially respawn when having their mind transferred to a new body (#77242)

## About The Pull Request
This PR fixes an oversight allowing blobs to "respawn" after having
their mind placed in a new body, this is caused by the blob antag datum
being left present in the mind after the blob is killed.
## Why It's Good For The Game
This is an upstream PR to deal with an issue that occurs due to a
downstream mechanic
https://github.com/Skyrat-SS13/Skyrat-tg/issues/22772. That being said,
in any kind of scenario where we don't make a new mind, this glitch can
occur and end up creating issues.

here's a example of me getting the glitch to work by ctrl+shift+clicking
my ghost after they were killed as a blob

![image](https://github.com/tgstation/tgstation/assets/68373373/d9d5b8c9-c255-4909-89a4-d8602f3979be)

## Changelog
:cl:
fix: former blobs are no longer able to respawn after having their mind
transferred to a new body.
/:cl:

* Fixes former blobs being able to potentially respawn when having their mind transferred to a new body

---------

Co-authored-by: Pinta <68373373+softcerv@users.noreply.github.com>