Adds Unit Tests for Radio Messages [MDB IGNORE] (#22950)
* Adds Unit Tests for Radio Messages (#77392)

## About The Pull Request

In light of the issues caused by #77350, it would be good to add unit
tests for radio messages so that this doesn't happen again.

I have tested the unit tests, they will pass with the current code but
correctly fail if the aforementioned bug is still present.
## Why It's Good For The Game

Unit tests are good.
## Changelog
No player-facing changes, just unit tests.

* Adds Unit Tests for Radio Messages

---------

Co-authored-by: GPeckman <21979502+GPeckman@users.noreply.github.com>