Adds the ability to jam cigarettes into the filter slot of a gas mask [MDB IGNORE] (#22967)
* Adds the ability to jam cigarettes into the filter slot of a gas mask (#76881)

## About The Pull Request
This PR makes it so that when a gas mask of any type (normal, explorer,
mime, etc) has isn't holding any filters, and isn't covering the mouth,
you can jam a cigarette in there and smoke it like normal. Cigs placed
in a mask can be lit by using the lighting implement on the mask as you
would a normal cig. The gas mask can't be adjusted while holding a cig
(if the gas mask can be adjusted at all). Whether or not internals can
be used depends on the mask and its adjusted state.

![image](https://github.com/tgstation/tgstation/assets/42454181/a1d00139-eef3-479a-b8f2-b12b531f6833)
A mime enjoying their favourite robusto cigars
## Why It's Good For The Game
First off, I just find the idea of replacing an air filter in a gas
mask, something meant to protect you from harmful gasses and fumes, with
a cancer-inducing burning stick of chemicals, to be really funny.

Secondly, and the reason that was the initial inspiration for this PR,
is it allows mimes and clowns to smoke without removing their masks.
This opens the door for mimes and clowns to use smoking as part of their
character gimmick, without losing a pretty big part of their identity.
## Changelog
:cl:
add: Cigarettes can now be placed in and smoked through gas masks, so
long as the mask has no filters installed and is not covering the mouth.
/:cl:

---------

Co-authored-by: san7890 <the@ san7890.com>

* Adds the ability to jam cigarettes into the filter slot of a gas mask

---------

Co-authored-by: Nick <42454181+Momo8289@users.noreply.github.com>
Co-authored-by: san7890 <the@ san7890.com>