Fixes mime book consuming itself when it isn't supposed to [MDB IGNORE] (#23032)
* Fixes mime book consuming itself when it isn't supposed to (#77513)

## About The Pull Request

``on_start_reading`` now has a return value, which the mime book will
use to early return out if an option isn't selected.
I also tested fikou's report of it eating its use while in the dark and
this seems to fix that as well.

## Why It's Good For The Game

Closes https://github.com/tgstation/tgstation/issues/77441
Fixes cancelling out of the mime book giving you the spells and removing
its use.

## Changelog

:cl:
fix: Mime spell books don't eat itself when used in the dark or
cancelled out of.
/:cl:

* Fixes mime book consuming itself when it isn't supposed to

---------

Co-authored-by: John Willard <53777086+JohnFulpWillard@users.noreply.github.com>