Harvesting podpeople seeds now requires explicit approval [MDB IGNORE] (#23041)
* Harvesting podpeople seeds now requires explicit approval (#77526)

## About The Pull Request

Podpeople seeds have a popup saying "This tray has no soul, are you sure
you want to harvest right now?", but it will still harvest if you x out
of the tab instead of clicking cancel. This flips it so you have to
explicitely say yes for it to go through.

## Why It's Good For The Game

When you X out of a tab, you did not mean to say yes.

## Changelog

:cl:
fix: X'ing out of the podpeople no soul prompt will no longer harvest
the seeds.
/:cl:

* Harvesting podpeople seeds now requires explicit approval

---------

Co-authored-by: John Willard <53777086+JohnFulpWillard@users.noreply.github.com>