Bumps gateway access down to Common access [MDB IGNORE] (#23042)
* Bumps gateway access down to Common access (#77534)

## About The Pull Request

Moves Gateway from requiring a Silver ID to a regular one.

## Why It's Good For The Game

Gateway is a Command access, but when it is enabled it is either locked
behind requiring a Silver ID, or opening the door/shutters to public
access.
We should not expect Command members (the only ones who have silver IDs)
to abandon the station to go explore the gateway, and we also should not
expect it to always be public access as it ruins the point of an
ID-locked area.

This lets the HoP give out Gateway access at the cost of a trim slot,
rather than force the entire area to be put on emergency access/open
shutters, which is a lame obligation for exploration content.

## Changelog

:cl:
qol: Gateway access no longer requires a Silver ID.
/:cl:

* Bumps gateway access down to Common access

---------

Co-authored-by: John Willard <53777086+JohnFulpWillard@users.noreply.github.com>