The chef can no longer close the HoP's shutter in Meta [MDB IGNORE] (#23055)
* The chef can no longer close the HoP's shutter in Meta (#77548)

## About The Pull Request

The HoP's shutter had "Kitchen" as the access, meaning in addition to
the HoP the Chef could also close these shutters. Only god knows how
long this has been a thing for.

## Why It's Good For The Game

Fixes an oversight in access

## Changelog
:cl:
fix: The chef can no longer close the shutters in Meta's HoP office.
/:cl:

* The chef can no longer close the HoP's shutter in Meta

---------

Co-authored-by: Fazzie <84548101+EuSouAFazer@users.noreply.github.com>