Fixes missing seating at the All American Diner ruin [MDB IGNORE] (#23086)
* Fixes missing seating at the All American Diner ruin (#77589)

## About The Pull Request

Fixes a missing sofa on All American Diner

## Why It's Good For The Game

Visiting the ruin recently I noticed a section of the seating was
missing; at first I'd thought maybe something had happened to it but
then I noticed it was missing across several rounds and opening up the
map editor, it seems it got deleted somehow.

So, this is just a tiny fix that adds it back.

## Changelog

:cl:
fix: fixed missing seating at the All American Diner
/:cl:

* Fixes missing seating at the All American Diner ruin

---------

Co-authored-by: YehnBeep <86855173+YehnBeep@users.noreply.github.com>