Held items are now dropped when their associated arm is dismembered [MDB IGNORE] (#23101)
* Held items are now dropped when their associated arm is dismembered (#77613)

## About The Pull Request

What it says on the tin. The parent of
`obj/item/bodypart/arm/drop_limb()` would remove the arm from the
owner's hand_bodyparts list, which would then prevent the child proc
from dropping the item held in the now missing hand.

Now we call the parent afterwards and it works just fine.
## Why It's Good For The Game

Addresses the issue with the GodHand described in #76899, although it's
not limited to items you can't ordinarily drop. It's a bit more
noticeable with them, because there's no risk of dropping the item in
shock if the limb is being hit with a weapon, but you can verify using
amputation shears that any ordinary item would also stay stuck on the
arm.
## Changelog
:cl:
fix: items no longer stay in your hands after their respective arm is
dismembered
/:cl:

* Held items are now dropped when their associated arm is dismembered

---------

Co-authored-by: A miscellaneous Fern <80640114+FernandoJ8@users.noreply.github.com>