Edits a proc from dynamic.dm to use the actual cumulative formula of the Cauchy-Lorentz distribution. [MDB IGNORE] (#23151)
* Edits a proc from dynamic.dm to use the actual cumulative formula of the Cauchy-Lorentz distribution. (#77460)

Alternative Title: I'm bad at maths but thank goodness graphing
calculators exist.

## About The Pull Request
So, it was reported on #77329 that `dynamic/proc/lorentz_to_amount` is
kinda "jagged", as you can see in the issue report yourselves. There,
they also suggested a better version with the actual segments connected.
However, as I internally shouted "It's copypasting time!", My gears came
to a halt. I promptly realized, looking at the graph, how it looked like
an approximation of a comulative distribution.
Fuddled, I wondered if we could replace that switch() call and its
indented crude algebra with the actual macro of the lorentz cumulative
distribution, which I did, not without the use of softwares to alleviate
my incapacity at performing such complex mathematical functions off the
top of my head, and with some peculiarities and improvements attached to
this PR, such as, instead of clamping, multiplying the value of the
distribution by the maximum threat.

## Why It's Good For The Game
This will fix #77329. The formula was tested a graphing calculator,
though it may work differently than the original proc, a testmerge is
reccomended.

## Changelog

:cl:
fix: Fixed some jank with the dynamic threat curve.
/:cl:

---------

Co-authored-by: san7890 <the@ san7890.com>

* Edits a proc from dynamic.dm to use the actual cumulative formula of the Cauchy-Lorentz distribution.

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Co-authored-by: san7890 <the@ san7890.com>