Add pickaxe to basic tools techweb (#72938) (#23155)
Add pickaxe to basic tools techweb (#72938)

Adds the pickaxe to basic tools techweb that can be printed the cargo
lathe or autolathe.

This is a basic mining tool and should be available at roundstart.

It wasn't possible to create pickaxes, despite it being a T1 item. It
should be as easy to make as a shovel since these usually are mapped
together. I was tempted to put this into mining technology, but that
already has advanced drills that are T2 versions of the pickaxe.

:cl:
qol: Add pickaxe to basic tools techweb
/:cl:

Co-authored-by: Tim <timothymtorres@gmail.com>