Fixing a minor mistake with update_transform() [MDB IGNORE] (#23176)
* [NO GBP] Fixing a minor mistake with update_transform() (#77685)

## About The Pull Request
Well, while the translation for lying/standing from #77619 is supposed
to work using the decimal place of whatever value the
`body_position_pixel_y_offset` variable should be while the mob is in
standing position (basically, `get_pixel_y_offset_standing()`), for a
fleeting moment idiocy took the better of me, and instead I just used
`body_position_pixel_y_offset` (which is of a different value when the
mob is lying down), resulting in resized mobs slooowly ascending (if
bigger) or descending (if smaller) each time they stood up.

Also taking a few seconds to add some `if` checks to avoid running these
operations for the wide majority of mobs, which have a default size of 1
anyway.

~~I hope I won't have to make further such PRs for a while.~~

## Why It's Good For The Game
Read above.

## Changelog
N/A

* [NO GBP] Fixing a minor mistake with update_transform()

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>