Makes satchels craftable + small map fix [MDB IGNORE] (#23226)
* Makes satchels craftable + small map fix (#77702)

## About The Pull Request

This makes satchels craftable using the exact same recipe as backpacks.

Also because it seems too small to be worth making a whole separate PR
for (though I can if it'd be preferred), also included is a fix for a
light inside a wall on Delta Station.

![image](https://github.com/tgstation/tgstation/assets/86855173/c72abe7e-6fd4-4fb0-a15a-a5bcf19f41e2)

## Why It's Good For The Game

I think it's a bit of a fashion crime that we can make backpacks but not
satchels.

## Changelog

:cl:
qol: Satchels can now be crafted using the same recipe (4 cloth) as
backpacks.
fix: Fixed a light inside a wall on Delta Station.
/:cl:

* Makes satchels craftable + small map fix

---------

Co-authored-by: YehnBeep <86855173+YehnBeep@users.noreply.github.com>