Fixes beakers disappearing when switching in chem heater [MDB IGNORE] (#23229)
* Fixes beakers disappearing when switching in chem heater (#77770)

## About The Pull Request
Fixes #77762

The function to replace beakers in the Chem Heater unregistered the
signal from the beaker after removing it, causing it to break and the
beaker to disappear. This fixes that by unregistering the signal first.
## Why It's Good For The Game
Beakers don't disappear when you try to swap them in the Chem Heater
anymore.
## Changelog
:cl:
fix: fixed beakers disappearing when swapped in a chem heater
/:cl:

* Fixes beakers disappearing when switching in chem heater

---------

Co-authored-by: Ical <86125936+Ical92@users.noreply.github.com>