Adds new sprites for opened fireaxe and mech removal crowbar cabinets [MDB IGNORE] (#23234)
* Adds new sprites for opened fireaxe and mech removal crowbar cabinets (#77772)

## About The Pull Request
Adds some new sprites for when the fireaxe and mech removal crowbar wall
cabinets are opened.
Old (it's just a line):

![oldcabinets](https://github.com/tgstation/tgstation/assets/58376695/34a21d06-971d-47b9-b9f4-73021531f542)
New:

![fireaxe](https://github.com/tgstation/tgstation/assets/58376695/5ef7b5ae-f0b3-40e8-822f-a7cde186cd8c)

![mechremoval](https://github.com/tgstation/tgstation/assets/58376695/017a0e23-75a4-4796-a2eb-07291f84a0ac)
## Why It's Good For The Game
The old sprite just seemed awkward, and you could barely tell it even
had the cover on it when opened. Makes the cabinets nicer looking when
opened and shows it from a better perspective.
## Changelog
:cl:
image: adds new sprites for opened fireaxe/mech crowbar cabinets.
/:cl:

---------

Co-authored-by: san7890 <the@ san7890.com>

* Adds new sprites for opened fireaxe and mech removal crowbar cabinets

---------

Co-authored-by: die_amond <58376695+dieamond13@users.noreply.github.com>
Co-authored-by: san7890 <the@ san7890.com>