Sets the dynamic human appearance via copy_overlays [MDB IGNORE] (#23249)
* Sets the dynamic human appearance via copy_overlays (#77794)

Set_dynamic_human_appearance set the new appearance directly on the
target mob. This has caused the mob's name to become Unknown, and its
description to become empty, as it inherited the name and description of
the appearance's source, which was a default spawned dummy human.

This PR makes it so that it nulls the icon of the target, and then
copies the appearance via copy_overlays.

Thank you Fikou for pointing out this simple solution, I almost did
something complicated with signals.

* Sets the dynamic human appearance via copy_overlays

---------

Co-authored-by: Profakos <profakos@gmail.com>