Fixes RCD'ing lattices in open space [MDB IGNORE] (#23306)
* Fixes RCD'ing lattices in open space (#77895)

## About The Pull Request

If you RCD a lattice with floor/wall mode on icebox, it won't actually
place down the floor because the open space is not "space" - this
switches the check to openspace, allowing you to fix floors on icebox by
clicking on lattices. For some reason this would also cause you to waste
your matter units on this action.

## Why It's Good For The Game

fix good

## Changelog

:cl:
fix: You should be able to click on lattices on icebox with the RCD and
construct a floor now
/:cl:

* Fixes RCD'ing lattices in open space

---------

Co-authored-by: iwishforducks <65363339+iwishforducks@users.noreply.github.com>