Replaces the decloning virus clone damage with a status effect [MDB IGNORE] (#23310)
* Replaces the decloning virus clone damage with a status effect (#77824)

## About The Pull Request

Simply replaces the decloning virus's dependency on clone loss with a
status effect that operates on a strike system, very similar to how it
was handled with clone damage. The strikes are reversed if you have
mutadone in your blood, eventually removing the status effect entirely.
If the virus is cured, the status effect is also removed.

Additionally, the decloning virus was not curable before. The cure hints
suggested this shouldn't be the case, so I added it back again.

## Why It's Good For The Game

Explanation from #77569:

> Clone damage is a damage type that shouldn't exist anymore, it's a
relic left from the era of cloning and it's so specific of a damage type
that it rarely gets used as a result. It really should be a type of
affliction (wound etc) instead of its own damage counter.

What's changed since then is that I don't think clone damage should not
exist at all anymore. It only affects one department and a few odd
situations anyways.

This PR should actually change very little about the virus, as it works
with the same numbers as it did before, and since cloneloss healing is
just as easy to come by as taking mutadone is.

## Changelog
:cl: distributivgesetz
add: Replaces decloning clone damage with a strike system that you can
replenish with Mutadone.
fix: The decloning virus is curable with rezadone again.
/:cl:

* Replaces the decloning virus clone damage with a status effect

---------

Co-authored-by: distributivgesetz <distributivgesetz93@gmail.com>