Less spam from stabilized yellow extracts [MDB IGNORE] (#23349)
* Less spam from stabilized yellow extracts (#77874)

## About The Pull Request

Removes the chat log message from yellow stabilized extracts that
normally writes every pulse.

Balloon alerts and examine text are left intact.

## Why It's Good For The Game

![image](https://github.com/tgstation/tgstation/assets/86855173/dc76f589-752b-4d7c-9fdc-e84c86df2a88)

While this does reduce the visibility of the effect, there is still a
balloon alert for when a player loses it, and a line in examine text
while it is active. I don't think knowing the exact time of every pulse
is significant information and so it would be better to not have
constant spam when using this item.

(But if I'm wrong and people like it more than they're annoyed by it,
I'll close this)

## Changelog

:cl:
qol: Stabilized yellow slime extracts no longer spam a message every
charge pulse (but examining characters will still show that it's
active).
/:cl:

* Less spam from stabilized yellow extracts

---------

Co-authored-by: YehnBeep <86855173+YehnBeep@users.noreply.github.com>