Makes it easier to place tiles on multi-z holes [MDB IGNORE] (#23381)
* Makes it easier to place tiles on multi-z holes (#77935)

## About The Pull Request
Adds the "open space click handler" to tile stacks, which makes it so
you don't have to pixel hunt for a turf on the BELOW z level in order to
fix a hole.
This exists on rods and rpds since they're often used to fix holes.
But wasn't added to tiles when they were made to be able to fix holes
directly, without rods.

Additionally, closes #77540 by having the open space click handler loop
up z levels so that it works if you're clicking on items from multiple z
levels away.
## Why It's Good For The Game
The current behavior can be very frustrating to work around, and appears
to not be intended.
## Changelog
:cl:
fix: Made it easier to place tiles on multi z level holes
/:cl:

* Makes it easier to place tiles on multi-z holes

---------

Co-authored-by: FlufflesTheDog <piecopresident@gmail.com>