removes the nukie tool parcel [MDB IGNORE] (#23421)
* removes the nukie tool parcel (#77986)

## About The Pull Request
removes the tool parcel and places tools directly in their box

## Why It's Good For The Game
the step of having to pick up your tools invalidates why we put em there
in the first place - so nukies can use them in an emergency. its
annoying to put them back in your box.
the parcel was added because the tools could make the box go above 7
slots during the extended box station trait, but i fixed that in the pr
that made it so that trait increases the box size of that box

## Changelog
:cl:
qol: removes the nukie tool parcel and places the tools directly in
their box
/:cl:

* removes the nukie tool parcel

---------

Co-authored-by: Fikou <23585223+Fikou@users.noreply.github.com>