Soup is food according to the smartfridge [MDB IGNORE] (#23466)
* Soup is food according to the smartfridge (#78055)

## About The Pull Request

Bowls are no longer allowed in drink showcases, and Bowls are now
allowed within Kitchen Smartfridges
## Why It's Good For The Game

Generally, it makes sense that soup, a food, can be placed in food
smartfridge and not the drink smartfridge. This check is based on the
container. Bowls of, for example, sulfuric acid are accepted by the food
smartfridge. Empty bowls are not accepted. However, the drink
smartfridge has analogous behavior, and so this will keep consistency
between the 2 smartfridge versions

Fixes #77334
## Changelog
:cl:
fix: Soups are accepted by Kitchen Smartfridges
fix: Soups are not accepted by drink showcases
/:cl:

* Soup is food according to the smartfridge

---------

Co-authored-by: blueDev2 <89804215+blueDev2@users.noreply.github.com>