Allow bag dumping into mortars and millstones (#23390)
* Add bag dumping functionality

Allow dumping bags directly into millstones as can be done with biogenerators and the like. For quality of life and consistency's sake.

* Add bag dumping functionality

Allow dumping bags directly into large mortars as can be done with biogenerators and the like. For quality of life and consistency's sake.

* Apply suggestions from code review

Co-authored-by: GoldenAlpharex <58045821+GoldenAlpharex@users.noreply.github.com>

* Apply suggestions from second code review

Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>

* Strip surplus balloon text

This won't ever pop up and even if it did it would stack on top of balloon text from a successful transfer as the process is instantaneous.

* Update big_mortar.dm with suggestions from second code review

Update to match millstone sans surplus balloon text

* Update big_mortar.dm with suggestions from second code review

Update to match millstone sans surplus balloon text

* Strip surplus balloon text

This won't ever pop up and even if it did it would stack on top of balloon text from a successful transfer as the process is instantaneous.

* return ..()

* return ..()

---------

Co-authored-by: GoldenAlpharex <58045821+GoldenAlpharex@users.noreply.github.com>
Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>