Fixes mortars grinding and juicing [MDB IGNORE] (#23617)
* Fixes mortars grinding and juicing (#78212)

## About The Pull Request

Mortal and pestle failed to grind items, as the proc passed over the
item itself, instead of the reagent holder. This PR fixes that.

## Why It's Good For The Game

Its good to have a way to grind some chemicals and fruits without an
All-in One Grinder.
Fixes #78205

## Changelog

:cl:
fix: Mortar and pestle can grind stuff again
/:cl:

* Fixes mortars grinding and juicing

---------

Co-authored-by: Profakos <profakos@gmail.com>