Adds lazylist documentation [MDB IGNORE] (#23637)
* Adds lazylist documentation (#78184)

## About The Pull Request

I realize that we don't really explain anywhere what a lazylist actually
is, and being how important of a tool they are I don't think we should
rely on word of mouth for them.

So I just added a block comment above the macros that people can read
about.

(We could totally do `#define LAZYLIST list` and then use it in
definitions as `var/LAZYLIST/mylist`, so people can control-click the
lazylist part and then see the documentation, but I dunno, whatever.)

Feel free to suggest changes of wording, add information, or correct any
misinformation.

* Adds lazylist documentation

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>