Ninjas can get credit for placing their bombs [MDB IGNORE] (#23636)
* [no gbp] Ninjas can get credit for placing their bombs (#78228)

## About The Pull Request

I made a mistake while refactoring this item.
We check if the weakref exists, then blow up the bomb. Destroy nulls our
weakref, then we try to resolve it. This obviously doesn't work and so
it never succeeds the Ninja's objective.
I swapped the order of operations around so now it works.

## Changelog

:cl:
fix: Ninjas should be correctly credited for using their spider bombs
/:cl:

* [no gbp] Ninjas can get credit for placing their bombs

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>