Fixes a fruit wine runtime and fixes runtime when trying to add inedible plant seeds to the seed extractor [MDB IGNORE] (#23707)
* Fixes a fruit wine runtime and fixes runtime when trying to add inedible plant seeds to the seed extractor (#78319)

Fixes https://github.com/tgstation/tgstation/issues/78315
Fixes https://github.com/tgstation/tgstation/issues/69282 (same issue)
Fixes https://github.com/Skyrat-SS13/Skyrat-tg/issues/23586 (same issue)

## About The Pull Request

@ Jacquerel end me please.

This fixes the linked issue and the other issue that was mentioned in
the comments. They are unrelated, but this is a two for one.

First bug is caused by fruit wine trying to blend `filling_color`, which
is actually unset for the majority of plants. As such it will runtime
when trying to make wine from just about every plant in the game. It
could be a nice project for someone to set those for all the plants so
they have more accurate colors when smashed or made into fruit wine...

But for now the bug is fixed by just having it use the reagent's `color`
for the blending when the `filling_color` var is unset.