Replace wall/indestructible/destructible typepath with mapping helper [MDB IGNORE] (#23735)
* Replace wall/indestructible/destructible typepath with mapping helper (#78365)

## About The Pull Request

#78239 is a fun mapping add but engineered in a way that
1. Creates a real eyesore of a typepath
2. Would further proliferate a hundred subtypes if it became commonly
used

Instead of using subtypes for this I put the behaviour in a component
and made a mapping helper to apply the component.
Now you can just put the mapping helper on top of any turf you want to
make into a zelda bomb wall and it will be so, rather than having to
make different subtypes for walls with different icons.

## Why It's Good For The Game

Cleaner, more maintainable.

## Changelog

not player facing

---------

Co-authored-by: san7890 <the@ san7890.com>

* Replace wall/indestructible/destructible typepath with mapping helper

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>
Co-authored-by: san7890 <the@ san7890.com>