Holo para refund fix [MDB IGNORE] (#23823)
* Holo para refund fix (#78431)

## About The Pull Request
Holoparasite injector can be refunded if its been used, this is because
the uplink_reimburse element is not removed from it when its used. Added
some signal code to it so that this is fixed.

## Why It's Good For The Game
Fixes a bug.

## Changelog

Fixes a bug allowing holopara injectors to be refundable when used.

:cl:
fix: Fixes a bug allowing holopara injectors to be refundable when used.
/:cl:

* Holo para refund fix

---------

Co-authored-by: Coconutwarrior97 <40315842+Coconutwarrior97@users.noreply.github.com>