Allows admins to overrule God [MDB IGNORE] (#23864)
* Allows admins to overrule God (#78429)

## About The Pull Request

Adds a "manage religious sect" verb to the "game" menu of the admin
panel.
It can be used to assign the chaplain's sect if they haven't picked one
yet, or reassign it (to a different one, or to nothing) if they already
have.
This is likely mostly going to be used for ahelps where someone
misclicks or suddenly logs off and wants to be replaced by a different
chaplain with different ideas.

## Why It's Good For The Game

Admins asked me to make it

## Changelog

:cl:
admin: Admins can now reset or modify the chaplain's sect from a UI
panel
/:cl:

* Allows admins to overrule God

* Updates modular to use the reset signal for nulling the altars instead

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>
Co-authored-by: Giz <13398309+vinylspiders@users.noreply.github.com>