Adds a frog pAI holoform [MDB IGNORE] (#23896)
* Adds a frog pAI holoform (#78537)

## About The Pull Request
It has been over a year since we have got a new pAI holoform, so I
present to you, the fwog.

![image](https://github.com/tgstation/tgstation/assets/20053168/3e1a77f0-e7b6-4aed-8401-c510afbcd2c5)
More importantly, it can be worn on the head like a normal frog.

![image](https://github.com/tgstation/tgstation/assets/20053168/795dcfd2-d576-4722-ab6f-edbed5b8fb36)

All of the new icons

![image](https://github.com/tgstation/tgstation/assets/20053168/b84ceca5-ff27-4a3a-a023-1d350762148b)

## Why It's Good For The Game
More holoform variety for pAIs is nice, especially one that can be worn
as a hat.
## Changelog
:cl:
image: adds a frog holoform for pAIs
/:cl:

* Adds a frog pAI holoform

---------

Co-authored-by: Alexis <catmc8565@gmail.com>