Advanced Plastic Surgery 2.0, It's rare now. [MDB IGNORE] (#23903)
* Advanced Plastic Surgery 2.0, It's rare now. (#78504)

## About The Pull Request
A revival of #76508 which got closed due to it eating up genetics niche.
This PR fixess that issue by seperating the surgery into two. the normal
roundstart plastic surgery and advanced plastic surgery. both works
identically with the one exception being that the advanced one allows
the feature that the last PR tried to add. which is to allow surgery to
mimic people in photographs.
the regular one remains as-is.

The disk containing the surgery can be found in a lot of places, a
role-restricted item to doctors and roboticists for 1TC, as a rare maint
loot and BEPIS technode reward
## Why It's Good For The Game
Opportunity for more fun gimmicks, ammo for BB's and paranoia, staple of
SS13 gameplay.
This also retroactively buffs forensic scanners and records since it
really only changes your name and voice. You can change your hair all
you want but anything more than that is a trip to genetics. overall
allows for more engaging gameplay while not detracting from genetics
niche with how niche and rare it would appear in rounds.
## Changelog
:cl:
add: Adds an advanced plastic surgery procedure, allowing you to imitate
people in pictures. Simply hold a picture in your offhand of the person
you wish to imitate as while conducting the surgery! Remember, it's not
foolproof, it only changes your name and voice!
adds: You can obtain the disk containing the afromentioned surgery. as a
role-restricted item to doctors and roboticists for 1TC, as a rare maint
loot and BEPIS technode reward
/:cl:

---------

Co-authored-by: ATH1909 <42606352+ATH1909@ users.noreply.github.com>
Co-authored-by: MrMelbert <51863163+MrMelbert@ users.noreply.github.com>

* Advanced Plastic Surgery 2.0, It's rare now.

---------

Co-authored-by: Singul0 <127663818+Singul0@users.noreply.github.com>
Co-authored-by: ATH1909 <42606352+ATH1909@ users.noreply.github.com>
Co-authored-by: MrMelbert <51863163+MrMelbert@ users.noreply.github.com>