Fixes map votes of only one choice [MDB IGNORE] (#23910)
* Fixes map votes of only one choice (#78509)

## About The Pull Request

Unfortunately, back in #70340 (4085e792ab5716485d379ce5f3dcbbae1c3d0522)
which was declared to fix the issue of having only one map to vote for,
I cocked up a part of it. For some reason, the part where it states that
only one map could be voted for would only play on the emergency shuttle
leaving. I think I did this because of the fact that the mapvote would
auto-call, and this didn't really respect the fact that people would
also call their own votes with only one map to vote for. Thus, the vote
would fire with one map to choose for and there wouldn't be the message
on there being only one map to vote for that I added in the
aforementioned PR. This brings it back to the intended working
condition.

* Fixes map votes of only one choice

---------

Co-authored-by: san7890 <the@san7890.com>