Gives maint drones random designations [MDB IGNORE] (#23944)
* Gives maint drones random designations (#78573)

## About The Pull Request
Right now, drone ghost spawners give all drones the name "drone." This
makes it difficult to differentiate drones, both as a drone in drone
chat and as a non-drone. This gives all drones a unique identifier like
posibrains.

Also changes a couple of 1s to TRUEs

## Why It's Good For The Game
It's good to be able to identify drones, and for drones to be able to
identify each other.

## Changelog
:cl: Tattle
qol: drones now have individual names, instead of just "drone"
/:cl:

Co-authored-by: tattle <article.disaster@ gmail.com>

* Gives maint drones random designations

---------

Co-authored-by: tattle <66640614+dragomagol@users.noreply.github.com>
Co-authored-by: tattle <article.disaster@ gmail.com>