Fix: Make death sandwich safe to examine [MDB IGNORE] (#23969)
* Fix: Make death sandwich safe to examine (#78630)

## About The Pull Request

The death sandwich currently has a humorous bug which causes the "death
sandwich poisoning" disease to be applied upon examining the item. The
bug is caused by the usage of the `check_liked` callback, which is part
of the edible component, and is called when a player examines the food.
Flavor text such as "You find this meal edible" is expected, but side
effects are applied instead.

To fix the bug, I swapped over to the `after_eat` callback which is
called for each bite of food taken.

In addition to fixing the bug, I also added a new override for the death
sandwich `examine` proc which causes the examiner to become
disgusted/nauseated if they "look at it wrong".

## Why It's Good For The Game

I had the pleasure of watching this bug happen live on Skyrat, it was
hilarious! Someone dumped a pile of death sandwiches on the evac shuttle
for everyone to look at.

Testing the fix in-game was successful and I was able to examine the
sandwich without getting the disease. I was able to eat it "wrong" and
get the message/disease as expected. I also tested eating it "right"
with the gelled back hair and cooking jorts.

## Changelog

:cl: A.C.M.O.
fix: Fixes the death sandwich, making it safe to examine.
/:cl:

* Fix: Make death sandwich safe to examine

---------

Co-authored-by: Dani Glore <fantasticdragons@gmail.com>