Only drop wall-mounted objects on turf change if it changes into a floor [MDB IGNORE] (#23998)
* Only drop wall-mounted objects on turf change if it changes into a floor (#78651)

## About The Pull Request

Fixes #78643
It was clearly the intent of this feature to drop stuff onto the ground
when the wall is destroyed.
If we turn a wall into a _different_ wall then there isn't any need to
do this.

If there's some side effect to this that I haven't thought about, I
haven't thought about it.

## Why It's Good For The Game

Dimensional anomalies aren't supposed to destroy lights, fire alarms,
APCs, and air alarms.

## Changelog

:cl:
fix: Dimensional Anomalies no longer destroy wall-mounted equipment.
/:cl:

* Only drop wall-mounted objects on turf change if it changes into a floor

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>