Adds add/remove mob ability to VV dropdown menu [MDB IGNORE] (#24009)
* Adds add/remove mob ability to VV dropdown menu (#78652)

## About The Pull Request

You add spells to mobs via a dropdown in VV but mob abilities via
marking the mob and pressing a button in the admin status panel.
I like opening the VV menu more than I like marking mobs (I am usually
going to need to open it anyway) so I added an alternate route in the VV
dropdown.

## Why It's Good For The Game

It's better for my personal workflow.

## Changelog

:cl:
admin: Mob abilities can be granted to arbitrary mobs via the VV menu in
a similar way to spells.
/:cl:

* Adds add/remove mob ability to VV dropdown menu

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>