Adds warden to default required enemies for dynamic rulesets [MDB IGNORE] (#24039)
* Adds warden to default required enemies for dynamic rulesets (#78610)

## About The Pull Request

The parent/default for rulesets had all the other metaprotected roles
except warden. Adds warden to the list as I suspect this was an
oversight.

## Why It's Good For The Game

It's an edge case, but if you're using the default required enemies for
some reason, you probably should count the other guy with access to all
the guns too.

## Changelog

:cl:
fix: Added warden to list of default required enemies for rulesets.
/:cl:

* Adds warden to default required enemies for dynamic rulesets

---------

Co-authored-by: Higgin <cdonny11@yahoo.com>