Security Officers can now Download the Plexagon Crew Manifest PDA App [MDB IGNORE] (#24049)
* Security Officers can now Download the Plexagon Crew Manifest PDA App (#78678)

## About The Pull Request

Despite starting with the Plexagon app, security officers did not have
sufficient access to actually download the app onto new machines. Now
they do.
## Why It's Good For The Game

Its weird and unintuitive to start with an app but not be able to
download it if you have to get a new PDA.
## Changelog
:cl:
fix: Security officers can now download the crew manifest PDA app that
they start with.
/:cl:

* Security Officers can now Download the Plexagon Crew Manifest PDA App

---------

Co-authored-by: GPeckman <21979502+GPeckman@users.noreply.github.com>