Paper bins behave properly when dumping their contents. [MDB IGNORE] (#24086)
* Paper bins behave properly when dumping their contents. (#78702)

## About The Pull Request

Fixes #78696

`dump_contents()` for paper bins was never updated to match the refactor
in #69586. This meant that it didn't work at all, dumping only
initialized sheets of paper and not removing references to them from the
bin itself. Most tangibly, this means that cutting open a natural bundle
of paper deletes all of the paper along with the object.

The paper bundle problem is now fixed. I don't know if there's actually
any way for a regular paper bin to dump its contents, but it works
properly regardless.
## Why It's Good For The Game

Dumping the contents out of something should dump the contents out of
that thing.
## Changelog
:cl:
fix: Cutting open a hand-pressed paper bundle no longer deletes all of
the paper.
/:cl:

* Paper bins behave properly when dumping their contents.

---------

Co-authored-by: lizardqueenlexi <105025397+lizardqueenlexi@users.noreply.github.com>