Fixes an issue w/ antiglow. [MDB IGNORE] (#24244)
* Fixes an issue w/ antiglow. (#78865)

## About The Pull Request

Antiglow's ``glow`` var used to be how strong it would glow, but it was
renamed to ``glow_power`` when
``/obj/effect/dummy/lighting_obj/moblight`` got added to it, which took
its previous name of ``glow``. Antiglow was never updated for this, so
they set the moblight as -1.5, which is supposed to be its strength.

I also removed some unused defines dw about that

## Why It's Good For The Game

Fixes an unintentional change that maybe breaks something idk.

## Changelog

:cl:
fix: Antiglow now probably has negative glow power.
/:cl:

* Fixes an issue w/ antiglow.

---------

Co-authored-by: John Willard <53777086+JohnFulpWillard@users.noreply.github.com>