Relocates a garbage spawner on Tramstation that was causing rng CI failures [MDB IGNORE] (#24256)
* Relocates a garbage spawner on Tramstation that was causing rng CI failures (#78883)

## About The Pull Request

This garbage spawner was in a spot where depending on which module was
loaded it would end up spawning in space, which led to it causing CI to
fail whenever it randomly chose
/obj/effect/spawner/random/trash/cigbutt.

This is because that spawner includes an /obj/effect/decal/cleanable/ash
as its garbage to spawn, which aren't supposed to spawn in groundless
turfs.

![firefox_8EQzJ52tBl](https://github.com/tgstation/tgstation/assets/13398309/50362bfe-ef3c-4f4f-acf3-b9d6a600f977)

![StrongDMM_866qdoxrLD](https://github.com/tgstation/tgstation/assets/13398309/b6c60748-5e8b-4f0c-9444-bce8c4211fa1)

<details><summary>Module that caused the issue whenever it was
loaded---spawner would spawn at the cursor's position, in
space.</summary>

![StrongDMM_8A7P5eoG4W](https://github.com/tgstation/tgstation/assets/13398309/fdb38b08-4379-4774-bb79-4e5de779b7b1)

![image](https://github.com/tgstation/tgstation/assets/13398309/7772e904-f057-4fe0-9062-59d93e182a4b)

</details>

I have moved the problem spawner one tile to the left which should fix
the issue of it causing variable runtimes.

## Why It's Good For The Game

Less CI failure rng

## Changelog

:cl:
fix: moved a garbage spawner on Tramstation that was causing random
runtimes due to sometimes spawning in space depending on which module
got loaded
/:cl:

* Relocates a garbage spawner on Tramstation that was causing rng CI failures

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>