Summoning affix fixes [MDB IGNORE] (#24369)
* Summoning affix fixes (#78983)

## About The Pull Request

The fantasy "summon x" affix blacklist doesn't work, and probably hasn't
worked for a significant amount of time.
This is because it generated a typecache with true/false values
representing whether we should be able to spawn a mob... and then
performed a `pick` on it. Pick doesn't care if the value is true or
false, so everything in the blacklist was explicitly whitelisted.

For some reason the list was also containing subtypes of a datum? Then
passing this to a component which expected typepaths of mobs it could
spawn? That doesn't work either.

We _also_ never added basic mobs to this list, so it would never spawn
those and they're an increasing number of our mobs total.

While I was there I also just did some general code tidying. I moved the
list of "specific subtypes to remove" to a global list because I suspect
something else either will need it in the future or already does.

## Changelog

:cl:
fix: Megafauna, lavaland elites, and abstract mobs now correctly cannot
be spawned by a toolbox of ash drake summoning
/:cl:

* Summoning affix fixes

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>