Fixes playing card hands displaying the wrong card at the top when there's more than five cards in the hand [MDB IGNORE] (#24377)
* Fixes playing card hands displaying the wrong card at the top when there's more than five cards in the hand (#78954)

## About The Pull Request
Title says it all, really.

It wasn't working right, because BYOND indexes start at 1 and not 0, so
when you had 6 cards, `cards.len - 5 = 1`, which is not what it was
meant to do.

## Why It's Good For The Game
Now hands of card can be used much more effectively for stuff like
discard piles where the top card is important.

(example of what placing down a full, unshuffled deck looks like. It
ends on a King, and as you can see, it displays the King now)

![image](https://github.com/tgstation/tgstation/assets/58045821/37da8d75-708a-48b6-a853-b5053632b4ca)

## Changelog

:cl: GoldenAlpharex
fix: Hands of cards will now properly display the last card added to the
hand all the time, even when there's more than five cards in that hand.
/:cl:

* Fixes playing card hands displaying the wrong card at the top when there's more than five cards in the hand

---------

Co-authored-by: GoldenAlpharex <58045821+GoldenAlpharex@users.noreply.github.com>