[MIRROR] Light dreaming refactor + new dream [MDB IGNORE] (#24499)
* Light dreaming refactor + new dream (#78996)

## About The Pull Request

Makes it so new dream types can be more easily added to the game with
effects within them beyond just some text being displayed if you want. I
added a dream that plays a random sound file available in the game files
to demonstrate how it's used. Mainly I wanted this out of the way for
another thing I'm working on.

I also made it so if the sound subsystem fails to reserve a channel it
throws a runtime since this happening is likely always going to result
in a bug, like it did while I was testing this.

## Changelog

:cl:
add: New dream that plays sound at you
/:cl:

* Light dreaming refactor + new dream

---------

Co-authored-by: Emmett Gaines <ninjanomnom@gmail.com>