[MIRROR] Precalculates special list var names and avoids incorrect vv targets [MDB IGNORE] (#24553) (#253)
* Precalculates special list var names and avoids incorrect vv targets (#79154)

## About The Pull Request

This adds a global list to precalculate lists that need to be handled
special by vv instead of manually updating such a list. @ LemonInTheDark
brought up a potential issue with handling all lists in the way special
lists are handled, that modifying a list in vv can in some cases result
in the wrong list being opened for editing. This makes only special
lists have the issue instead of blindly accessing all lists the same
way.

* Precalculates special list var names and avoids incorrect vv targets

---------

Co-authored-by: SkyratBot <59378654+SkyratBot@users.noreply.github.com>
Co-authored-by: Emmett Gaines <ninjanomnom@gmail.com>