[MIRROR] Adds a few extra params to blackbox dynamic threat logs. [MDB IGNORE] (#24635)
* Adds a few extra params to blackbox dynamic threat logs. (#79344)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds the roundstart pop count and the calculated max threat to blackbox
logging of dynamic threat.

This allows for identifying lowpop shifts and more intelligent analysis
of dynamic threat.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

I need this added to better track what's going on when analysing threat
distribution, as all threat stats will bias low with the inclusion of
lowpop shifts capping max threat.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

No player-facing changes.

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Mothblocks <35135081+Mothblocks@ users.noreply.github.com>

* Adds a few extra params to blackbox dynamic threat logs.

---------

Co-authored-by: Timberpoes <silent_insomnia_pp@hotmail.co.uk>
Co-authored-by: Mothblocks <35135081+Mothblocks@ users.noreply.github.com>