[MIRROR] New botany plant - Lanternfruit [MDB IGNORE] (#24659)
* New botany plant - Lanternfruit (#79204)

## About The Pull Request

So I noticed the code references an Ethereal plant called
"Lanternfruit," a sour pear-like fruit used as flavoring for a few foods
and drinks. However, this plant doesn't actually appear in the game,
this pr hopes to remedy this while also bringing some new hybrid
potential to the table for botany.

In essence, the plant contains sugar, liquid electricity, and sulfur;
making it the easiest way to obtain the latter two chems, allowing
botanists to experiment more with potential chemical recipes, and in
theme with the plant itself, make flash powder. The justification for
containing sulfur being that grapefruit gets its signature smell from a
sulfur compound.

The plant can also be fermented into voltaic wine, which has some
gimmick potential for particularly snobby Ethereals.

* New botany plant - Lanternfruit

---------

Co-authored-by: Toastgoats <63932673+Toastgoats@users.noreply.github.com>