[MIRROR] Initializes on_finish_callbacks in the JPS move loop [MDB IGNORE] (#24788) (#450)
* Initializes on_finish_callbacks in the JPS move loop (#79505)

## About The Pull Request

The on_finish_callbacks list in `/datum/move_loop/has_target/jps` was
not initialized. Usually, when you add something to an uninitialized
list, it gets initialized and the item gets added to it. However, the
`CALLBACK` wrapper is around `new /datum/callback`, this fails.

This meant on_finish_callback was not a list, therefore at the end of
pathfinding, its contents could not be iterated and invoked. This PR
fixes this problem by initializing the list.

## Why It's Good For The Game

Closes #79383
Blob minions rally and punch again.

## Changelog

:cl:
fix: Basic mobs using JPS can move again
/:cl:

* Initializes on_finish_callbacks in the JPS move loop

---------

Co-authored-by: SkyratBot <59378654+SkyratBot@users.noreply.github.com>
Co-authored-by: Profakos <profakos@gmail.com>