[MIRROR] Fixes a small issue with disposal outlets. [MDB IGNORE] (#25018)
* Fixes a small issue with disposal outlets. (#79706)

## About The Pull Request
Disposal outlets eject the contents of the holder with a 2 seconds
delay. Delete the outlet before the delay is over and the contents will
stay stuck inside the holder.

Also merging disposal holders doesn't properly register a signal for
mobs transfered to the new holder.

## Why It's Good For The Game
Bugfixing. Disposal code is a mess. Perhaps related to #79629 but I
couldn't exactly reproduce it.

## Changelog

:cl:
fix: Fixed a small issue with disposal outlets leaving contents about to
be ejected stuck inside the pipe beneath it if deleted.
/:cl:

* Fixes a small issue with disposal outlets.

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>