[MIRROR] [NO GBP]RCD can build directional windows on top of existing grills [MDB IGNORE] (#25206)
* [NO GBP]RCD can build directional windows on top of existing grills (#79864)

## About The Pull Request
- Fixes https://github.com/Skyrat-SS13/Skyrat-tg/issues/23788

So after #77858 was merged an unintentional side effect was rcd could no
longer build directional windows on grills if they already existed on
the turf. Sure rcd won't build a grill and then build a directional
window on top of that but if a grill already exists on the turf the rcd
should still be able to build directional windows on that grill as it
helps in repairing mapped in directional windows and such.

If you want to build a grill first set the mode to full tile window,
build the grill and then switch to directional window to build a window
on top of that grill. Or just select directional window mode and build
the windows directly without a grill, choice is yours

## Changelog
:cl:
fix: RCD can build directional windows on top of existing grills &
without them.
/:cl:

* [NO GBP]RCD can build directional windows on top of existing grills

---------

Co-authored-by: SyncIt21 <110812394+SyncIt21@users.noreply.github.com>