[MIRROR] Fixes fullupgrade chem dispenser not spawning with all their chemicals [MDB IGNORE] (#25180) (#771)
* Fixes fullupgrade chem dispenser not spawning with all their chemicals (#79858)

## About The Pull Request

Fixes https://github.com/Skyrat-SS13/Skyrat-tg/issues/24911

Tin. They were spawning with only the emagged reagents. This was not an
issue for the drink dispensers because they call the parent after
adjusting the reagent lists. It needs to be done that way so that
`RefreshParts()` gets called after adjusting the lists.

<details>
<summary>Before/after</summary>

![dreamseeker_H0sicTrScp](https://github.com/tgstation/tgstation/assets/13398309/649ab998-8308-4efa-be25-f4942ecab175)

![image](https://github.com/tgstation/tgstation/assets/13398309/9709dfa3-bf64-40f8-ab24-8508255d17ae)

</details>

## Why It's Good For The Game

Fixes a bug

## Changelog

:cl:
fix: fullupgrade chem dispensers will now spawn with all their chems
/:cl:

* Fixes fullupgrade chem dispenser not spawning with all their chemicals

---------

Co-authored-by: SkyratBot <59378654+SkyratBot@users.noreply.github.com>
Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>