Makes xenobotany plants more interesting. (#1002)
* Makes xenobotany plants more interesting.

* vaporsac spelling fix

There was a spelling error

* Jurlmah change

Replaces the hecuri in Jurlmah with fake beer, which should knock you out for the cryoxadone and frost oil to actually work.

* More honey gelthi

it's all people are gonna be using this plant for there's twice the honey now

* thaadra sansufentanyl

puts sansufentanyl, instead of super melatonin. Super melatonin is going to go into jurlmah instead

* jurlman super melatonin

replaces the fake beer in jurlmah with super melatonin. less overpowered knockout chem and also it will boost the healing

Co-authored-by: Milk <Catpunter2@gmail.com>