[MIRROR] [NO GBP] Fixes runtime in mat container chat display [MDB IGNORE] (#25617)
* [NO GBP] Fixes runtime in mat container chat display (#80321)

Forgot to format the chat in the right format when items are rejected by
the mat container. Rather than displaying the item is rejected it would
simply runtime & not display anything. That's fixed now

:cl:
fix: items that are rejected by the mat container will display the chat
message saying that.
/:cl:

* [NO GBP] Fixes runtime in mat container chat display

---------

Co-authored-by: SyncIt21 <110812394+SyncIt21@users.noreply.github.com>