[MIRROR] qm uses cargo tech icons [MDB IGNORE] (#25707) (#1206)
* qm uses cargo tech icons (#80411)

## About The Pull Request
qm now has the "x-in-training" intern title instead of "intern
quartermaster", similar to the other heads of staff
qm's hud and id card icons are now the cargo tech's icon (but blue),
following the other heads of staff which have blue versions of their
dept (blue wrench, blue cross, blue... sec badge?? blue exclamation)

## Why It's Good For The Game
makes the qm more visually consistent with the other heads, makes it
clearer that the qm is the head of the cargo department
the current icon is kinda hard to decipher too? you can tell the cargo
tech's is a crate, but the qm's is just like a weird head dude (which
the hop already has, making those ids hard to distinguish since theyre
both blue)? or a stamp?

## Changelog
:cl:
fix: qm's intern id is now "quartermaster-in-training"
image: qm's id and hud icons now use the cargo tech icons (but blue) to
be consistent with the other heads of staff
/:cl:

* qm uses cargo tech icons

* Removes the icon from modular as well

* Revert "Removes the icon from modular as well"

This reverts commit ac6a674887ca74a1231a8433373d55aa8df912e8.

* Updates the modular icons with the new version

---------

Co-authored-by: SkyratBot <59378654+SkyratBot@users.noreply.github.com>
Co-authored-by: Fikou <23585223+Fikou@users.noreply.github.com>
Co-authored-by: Giz <13398309+vinylspiders@users.noreply.github.com>