Fixes radar issues [MDB IGNORE] (#25785)
* Fixes radar issues (#80453)

## About The Pull Request
Went through tgui:sonar to search for bugs. Some of these are valid -
they've been addressed. Others, not so. Telling me I need to make icon
names and css classes into constants is a nit at best. Radar seems to be
bugged in certain areas: Classnames ok to dupe, but not ones with
conditionals. Most of the function duplication errors rely on hooks.

I think this should close #79815 and we should remove radar shortly
after. See my comment
[here](https://github.com/tgstation/tgstation/issues/79815#issuecomment-1859286784)
## Why It's Good For The Game
Fixes #79815
## Changelog
N/A none of this was player facing

* Fixes radar issues

---------

Co-authored-by: Jeremiah <42397676+jlsnow301@users.noreply.github.com>