Allows ai's to eject from mech domination (#80357)
## About The Pull Request
Lets ais eject from dominated mechs

If this is too strong at 30 I'd be happy to bump it or apply some other
restrictions
## Why It's Good For The Game
They have a big eject button which is essentially bait at the moment, it
does nothing.

I think its fine to let combat upgraded ai's and malfunctional ai's to
use this mech domination ability more freely, because how it usually
goes is malf ai dominates mech, everyone sees occupants of mech get
tossed out and the ai is promptly killed. I don't think this is fun or
healthy for an ability which is pretty expensive and the only other way
I see it used is an ai just trying to live for a few moments more by
shunting into a cargo mech, from which they can't do a whole lot and
simply exist until the crew figure them out.

TLDR: For a very expensive ability it does not provide a lot of benefit
and is easily dispatched by the crew.
## Changelog
:cl:
balance: Ai's using the mech domination ability can now eject from the
mech
/:cl:

Co-authored-by: John Willard <53777086+JohnFulpWillard@ users.noreply.github.com>