[MIRROR] Remove eye surgery fluff line [MDB IGNORE] (#25867)
* Remove eye surgery fluff line (#80565)

## About The Pull Request
Removes a side effect in can_start which MUST be pure. This causes the
surgery tgui to spam this message. It's pointless because you can't do
the surgery to begin with if this returns false, so you only ever see
this where you shouldn't.

## Changelog

:cl:
fix: Fixed the surgery menu spamming chat messages when on the eyes
section of a player with no eyes.
/:cl:

* Remove eye surgery fluff line

---------

Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>