[MIRROR] Makes immerse use weakrefs [MDB IGNORE] (#25876)
* Makes immerse use weakrefs (#80594)

## About The Pull Request
Immerse was causing harddels due to it having references to mobs. Makes
it use weakrefs for mobs instead.

## Why It's Good For The Game
Immerse would cause harddels if a mob was deleted while it was in it's
list. It could probably also happen if a turf was deleted too, but doing
that here would be much harder.

no CL since nothing playerfacing

* Makes immerse use weakrefs

---------

Co-authored-by: Arturlang <24881678+Arturlang@users.noreply.github.com>