fix a hard delete in basic statues (#243)
* fix a hard delete in basic statues (#80696)

## About The Pull Request
if the creator of the statue got deleted, it might cause a hard delete.
also i noticed this variable wasnt being used so i removed it and
instead directly added the ref of the creator to the mob's faction list.
also i noticed the proc that was setting it had alot of nested
typechecks so i split it into child procs

## Why It's Good For The Game
fixes a hard delete

## Changelog
not player facing

* fix a hard delete in basic statues

---------

Co-authored-by: Ben10Omintrix <138636438+Ben10Omintrix@users.noreply.github.com>
Co-authored-by: NovaBot <novasector@gmail.com>