Map export admin verb/buildmode (#267)
* Map export admin verb/buildmode (#80326)

## About The Pull Request

The base for the PR was taken from
https://github.com/shiptest-ss13/Shiptest/pull/206 and thank them for
that.

The point of this verb is to save pieces of the map to your computer for
further use. It's not that necessary, but rarely, it can be useful.

[Video](https://i.imgur.com/M6mdDTC.mp4)

## Why It's Good For The Game

Transferring buildings from one round to another, preserving the decor
made in the game.

## Changelog
:cl: Vishenka0704
admin: The ability to export a part(or z-level) of the map has been
added.
/:cl:

* Map export admin verb/buildmode

---------

Co-authored-by: Yaroslav Nurkov <78199449+AnywayFarus@users.noreply.github.com>
Co-authored-by: NovaBot <novasector@gmail.com>