Updates Pirate/Sailor Costume sprites (And spacesuit) (#249)
* Updates Pirate/Sailor Costume sprites (And spacesuit) (#80550)

## About The Pull Request

![image](https://github.com/tgstation/tgstation/assets/76465278/70ce71ad-4138-4f1f-acf3-dc9e1e3fd7be)

Not much else to it.

If you can't tell what's happened or the image doesn't load though...
Resprites a total of 4 items:
- Pirate Costume
- Sailor Costume
- Pirate Jacket
- Pirate Spacesuit (since it had the jacket on it)
## Why It's Good For The Game
More consistent sprites - the jacket was showing its age, and the
shading on the costumes was inconsistent and scared me :(
Otherwise, the sprites haven't changed. Mostly just color and shading!
## Changelog
:cl:
image: resprited the Pirate/Sailor costumes, the pirate jacket, and the
pirate spacesuit
/:cl:

Co-authored-by: Ghom <42542238+Ghommie@ users.noreply.github.com>

* Updates Pirate/Sailor Costume sprites (And spacesuit)

---------

Co-authored-by: OrionTheFox <76465278+OrionTheFox@users.noreply.github.com>
Co-authored-by: Ghom <42542238+Ghommie@ users.noreply.github.com>
Co-authored-by: NovaBot <novasector@gmail.com>