Organ movement refactor *Un-nullspaces your organs* [MDB IGNORE] #25530 (#42)
* Organ movement refactor *Un-nullspaces your organs*

* Fix conflicts

I checked the conflicts on the two weird conflicts and no previous TG pr touches them i assume its just github being github because those shoulden't be conflicts *shrug

* Fix #1 uhh...this is going to be a long one

* Fix #2 Modular Movement Flags

* Fix #3 It builds now

* Fix #4 Oh god it builds now, I missed some things

* Fix #5 No more Runtimesplosion Now time for Synths

* Fix #6 Shut up CI Where did you even come from

* Fix #7 Didn't I fix you already?

* Update language.dm

* Fix Opfor jank

* Revert "Fix Opfor jank"

This reverts commit 6af2fd71d85e6be58890f27f4f32bf92cfcb00fb.

* Revert "Revert "Fix Opfor jank""

This reverts commit cbd72277afd05f30a0b6f499a8162412fe3663ea.

* Update module_template.md

* Revert "Revert "Revert "Fix Opfor jank"""

This reverts commit 71ae4c10d8e9876c97f67278c55caa994b26a7dc.

* Fixing a runtime with H.A.R.S. (#80772)

The brain gets moved into the chest with H.A.R.S. now, but the bodypart
insertion and removal procs for it still asume it can only be found in
the head. This should fix it. For the sake of preventing similar issues
in the future, I've also updated the pretty-much-identical versions for
ears, eyes and tongue.

I've checked, and the brain var is only used by the head for visuals and
examine strings.

* Modular updates

* Fix synths not dying after being debrained

* Update brain.dm

* Update cortical_borer.dm

* Revert "Fixing a runtime with H.A.R.S. (#80772)"

This reverts commit 225035c2101d0e8946f7ef8cb375b68f2d25533a.

* Update dismemberment.dm

---------

Co-authored-by: Time-Green <7501474+Time-Green@users.noreply.github.com>
Co-authored-by: SomeRandomOwl <somerandomowl@ratchtnet.com>
Co-authored-by: SomeRandomOwl <2568378+SomeRandomOwl@users.noreply.github.com>
Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>