Icebox cytology fix (#80997)
## About The Pull Request

Fixes cytology in icebox's perma

This fix is slightly soulful (or the most efficient if you hate object
editing).

## Why It's Good For The Game

This has been broken for a while, and while it can be mildly interesting
to make improvised tools or convince security to fix it a few times, it
gets tiring to do every time - moreover, that it's broken is not obvious
nor intuitive to new players. Or anyone unfamiliar with troubleshooting
broken plumbing. So I think it is better if it's not broken in its base
state.

## Changelog

:cl:
fix: Icebox perma's cytology lab is now useable without outside repairs.
/:cl: