[MIRROR] The construction console drone becomes visible again while its in use (#677)
* The construction console drone becomes visible again while its in use (#81153)

## About The Pull Request

The construction console (in game only used as the aux base construction
console) summons a little drone while its active. This drone's
invisibility is turned off, and when the user logs out, its reset to
default. However, after the invisibility refactor, it was the console's
invisibility that was being removed and reapplied. This PR fixes that,
and properly applies it to the drone.

## Why It's Good For The Game

I want to know where the RCD is about to place the wall, so I can place
the items where I want them to go.

## Changelog


:cl:
fix: The construction console drone becomes visible again while its in
use
/:cl:

* The construction console drone becomes visible again while its in use

---------

Co-authored-by: Profakos <profakos@gmail.com>