[MIRROR] Fixes Alien Hunter's Pounce Visuals (#829)
* Fixes Alien Hunter's Pounce Visuals (#81322)

## About The Pull Request

Currently on live, when an alien hunter goes for a pounce, once the
pounce is completed the hunter will rise quite a distance slowly from
the floor as setting body_position_pixel_y_offset now is done with
animation as opposed to being instant. This looked extremely stupid, so
I went and fixed it. Initially, I was going to add some form of trait to
cancel out the smoothing effect from a shifting y offset, until I looked
at the sprite for alien pouncing and realized it was 96x96 for no
reason. For reference, 96x96 is the size of Bubblegum's sprites.

So, I just made the sprites 48x32, got rid of the unused sprites
(sentinel and drone both had pounce sprites but neither ever had that
ability iirc) and adjusted some of the pounce ability's internal offset
stuff to compensate and it looks perfectly fine now. Somehow made the
image file larger by 243 bytes according to Github but I guess it is
what it is.

## Why It's Good For The Game

The alien hunter ascending from 2 tiles below where it actually was
whenever it pounced was incredibly jarring. This fixes the problem.

## Changelog
:cl:
fix: Alien Hunter's pounce ability no longer has any weird offset issues
once the pounce is completed.
/:cl:

* Fixes Alien Hunter's Pounce Visuals

---------

Co-authored-by: IndieanaJones <47086570+IndieanaJones@users.noreply.github.com>