[MIRROR] Nerfs probability that a rat decides to bite a cable (#1914)
* Nerfs probability that a rat decides to bite a cable  (#81364)

## About The Pull Request

Rats are 5x less likely to decide to bite a cable

## Why It's Good For The Game

Way back when I converted rats to basic mobs, *something* went wrong and
rats bite cables wayyyy too often now - it's not uncommon to see a rat
has de-cabled an entire section of maint due to some good luck.

Funny but not how it functioned originally. I always intended to tone it
back down and just never got around to it.

## Changelog

:cl: Melbert
balance: Rats are now 5x less likely to decide to eat a cable when
idling. (1%, down from 5%)
/:cl:

* Nerfs probability that a rat decides to bite a cable

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>