[MIRROR] Air alarms display location/trigger information (#2067)
* Air alarms display location/trigger information (#81436)

## About The Pull Request

Adds additional information to triggered fire alarms, viewable on the
air alarm. Displays the method of alarm trigger and the source of the
alarm.


![image](https://github.com/tgstation/tgstation/assets/83487515/2b8bb334-e641-451b-9d6d-97e650e74032)

## Why It's Good For The Game

Useful for engineers to track down the source of why the fire
alarm/firedoors are triggering. Fire alarms that are manually triggered
will never reset without intervention, so this also identifies these
cases instead of chasing down an issue that may not exist.

## Changelog

:cl: LT3
qol: Air alarms now display the source of triggered fire
alarms/firedoors
/:cl:

* Air alarms display location/trigger information

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>