[MIRROR] Makes Ethereal Charging Loop Just About Everywhere (#1254)
* Makes Ethereal Charging Loop Just About Everywhere (#81775)

## About The Pull Request
Basically,


For APCs.

![dreamseeker_bfZeNmkOPh](https://github.com/tgstation/tgstation/assets/12636964/39eddc7b-7389-4586-a392-4a8d2437e93b)


For recharging APCs.

![dreamseeker_mlKuDR2JJW](https://github.com/tgstation/tgstation/assets/12636964/1f28850c-57a7-43b2-9663-81b2f3b34ce9)


For recharging from cells. The Ethereal offscreen was blue, you'll just
have to trust me that this was a color-coded PR.

![dreamseeker_bcP0126NlF](https://github.com/tgstation/tgstation/assets/12636964/c20eff96-3112-4b9b-a733-c320528adb5e)

## Why It's Good For The Game
It's kind of CBT as-is right now to have to click this stuff over and
over, especially when light tubes have it that much more convenient than
everything else. I figure some while loops would make the situation
better for them.

## Changelog
:cl:
qol: Ethereal charging now loops when they're charging (from) APCs or
from power cells!
/:cl:

* Makes Ethereal Charging Loop Just About Everywhere

---------

Co-authored-by: Nerevar <12636964+Nerev4r@users.noreply.github.com>