[MISSED MIRROR] Resprites Mi-Go to not be plagiarized from a CDDA tileset + adds special shiny variant of the Mi-Go (#81851) (#1387)
Resprites Mi-Go to not be plagiarized from a CDDA tileset + adds special shiny variant of the Mi-Go (#81851)

A lot of the xenobiology mobs are really, _really_ crusty. The mi-go is,
in my opinion one of the worst offenders (DMI file shown below), and
even if it wasn't as old as it is - and it is very old, I think it
predates this github - it's also just straight up stolen from a CDDA
tileset. This pr fixes that.
Old mi-go:

![image](https://github.com/tgstation/tgstation/assets/69398298/61ccbde7-d36e-40d3-8bef-24f52f11bbb0)

New mi-go:
![Go migo
Go](https://github.com/tgstation/tgstation/assets/69398298/6427f767-33a2-434f-8469-68f458466a7f)

"But what about the shiny variant you added?" Well, let me introduce you
to...


![MigoH1](https://github.com/tgstation/tgstation/assets/69398298/b80ec0a6-9ab6-488e-867b-4cececcee0aa)
_**Hatsune Mi-go.**_

...I'll see myself out. 

In honesty, this PR started as solely a shitpost made in an hour of
photoshopping a miku wig onto a mi-go, but ended up willing itself into
quality.
## Why It's Good For The Game

Besides the spritework obviously having aged, having stolen/traced art
ingame is not a good look. Below is the DMI of the mi-go, put along with
a screenshot of the CDDA tileset. It is, admittedly, _slightly_ altered,
but thats because the CDDA version wouldn't fit in a dmi's 16x16
tilesize, and so parts are just chopped off to make it work, and some
coloring was changed.


![a-migo](https://github.com/tgstation/tgstation/assets/69398298/9bd28073-2519-4049-a62c-6cfa46409484)

As for the mi-go/miku pun - if we have the 1 in 100 among us firelock, a
1 in 25 chance for a migo (a rare mob) to have an easter egg/pun seems
fine. I like the tiny chance things like that in this game.

## Changelog
:cl:
add: Shiny joke mi-go variant (not xenobio spawnable)
del: Removed unused mi-go static sprites
image: Resprited mi-gos to not use plagarized art from CDDA
image: Also allowed mi-gos to have directional facing instead of always
facing east
/:cl:

Co-authored-by: KingkumaArt <69398298+KingkumaArt@users.noreply.github.com>