[MIRROR] Space dragons are no longer acceptable dna for the omnitrixbelt (#1650) (#2578)
* Space dragons are no longer acceptable dna for the omnitrixbelt (#82157)

## About The Pull Request

We've had people able to transform into space dragons with no
limitations for quite a while now. Whenever dragons are killed multiple
people will rush the polymorph belt and become space dragons. this
version of the space dragon faces none of the limitations of the regular
antagonist, such as not needing to place portals.

## Why It's Good For The Game

Crewmembers should not be able to become space dragons en masse after
doing a single anom core test. This isn't healthy and leads to headaches
for admins and players who aren't a part of the dragon spam.

## Changelog

:cl:
balance: The polymorph belt will no longer register space dragons as
acceptable DNA.
/:cl:

* Space dragons are no longer acceptable dna for the omnitrixbelt

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: carshalash <carshalash@gmail.com>