[MIRROR] Mech mob deletion fix (#2797)
* Mech mob deletion fix (#82415)

## About The Pull Request

If the occupant of a mech has their mob changed -- mutation, polymorph,
admin fuckery -- or if a mob somehow enters a mech without becoming an
occupant -- the above, or a chestburster in said mech pilot -- then they
don't get added to the occupants, only the contents. All contents are
hard-deleted if a mech is destroyed. This fixes that, and ejects any
mobs inside mechs who are technically inside the mech but not an
occupant.
## Why It's Good For The Game

Fixes #82395 and any similar case
## Changelog

Being inside a mech when it's destroyed should no longer delete you,
even if you ended up in that mech in an odd way.
:cl: Bisar
fix: Mechs will no longer delete their living non-occupant contents when
destroyed.
/:cl:

* Mech mob deletion fix

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: Joshua Kidder <49173900+Metekillot@users.noreply.github.com>