[MIRROR] Fix playing card memory (#2156)
* Fix playing card memory (#82834)

## About The Pull Request

`/datum/memory/playing_cards` accepts a list of other players, which it
uses to format into an english list.

But where it is applied, it takes a list of players, formats it into an
english list, and passes it in. Oops.

Also fixed some verbage for how the memories look in-story-form. 

## Changelog

:cl: Melbert
fix: Playing Cards memory now reads better
/:cl:

* Fix playing card memory

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>