General maintenance of the Animal Traits quirks (#2241)
* Replace a ..()

Ought to be a '. = ..()'

* Remove verb_sing from resets.

The quirks don't change these!

* Make canidae and avian trait quirks normal quirks

No reason for these to be item_quirks. Just makes the code more confusing.

* Oops.

* Unify the names in code for readability

Just having them be called canine and avian is a bit... Let's add an _aspect those.

* Make the animal traits quirks clean up after themselves

There is no reason these quirks shouldn't give you your default species tongue when they're removed.

* Give each animal trait quirk gain/lose text

Giving quirk feedback to players is important. Especially for Changelings.

* Add Avian Traits to mimicable_quirks_list

A quarter of the reason I started doing this is so Changelings could copy these quirks without having any issues. I do not think this is out of the scope here.