[MIRROR] A birdshot atmos improvement, and fixes the engineering microwave. (#3125)
* A birdshot atmos improvement, and fixes the engineering microwave. (#82924)

## About The Pull Request
I have changed the layout of the pumps going into atmos from the tanks
to make it easier to use.

![image](https://github.com/tgstation/tgstation/assets/64122807/ec95eab6-b7ac-4927-b3d9-70f7c58e5dc8)

![image](https://github.com/tgstation/tgstation/assets/64122807/571bb2f3-49c8-4b4b-9d9d-405a8d5a06eb)
I also fixed the engineering microwave.
I also adjusted the name for the pumps and vents in the AI chamber. 


This is my first PR!
## Why It's Good For The Game

This change allow people to easily set up a pump to fill up a can with
the gas they want. Before this it was impossible to fill up an n2o can
and a plasma can without making a big mess or having them mix together.

The microwave in engi being able to work is also a good quality of life
change for the poor engineers that want to cook a donk pocket in their
department.

The change of names for the vents and pumps in the AI chamber is for
consistency sake. n2o is not helium!
## Changelog
:cl:
qol: Made birdshot atmos easier to use.
fix: The Birdshot engineering department is no longer using a second
hand broken microwave.
/:cl:

* A birdshot atmos improvement, and fixes the engineering microwave.

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: Constellado <64122807+Constellado@users.noreply.github.com>