[MIRROR] civilian bounty control terminal displays bounty description and payout of new bounty options (#2275) (#3146)
* civilian bounty control terminal displays bounty description and payout of new bounty options (#82987)

## About The Pull Request
The green buttons for selecting one of three bounties now also display
the description and payout.


https://github.com/tgstation/tgstation/assets/94711066/53d7fb02-28be-4138-b3d8-9f7a4dd6a356


Fixes #81867
Sorta brings it back to how it used to look, I think? The payout was
still in the code, just not being displayed. But I looked through the
file history and could find NO EVIDENCE of the bounty console displaying
the description in the past so I guess I hallucinated that.
## Why It's Good For The Game
It's nice to be able to see the payout before you're locked into a
bounty for 5 minutes. Also more chances to read the flavortext and it
makes the buttons feel less barren. With some bounties it also gives a
better idea of how to even get the thing before committing to it.
## Changelog
:cl:
fix: the civilian bounty control terminal displays the payout of new
bounty options again
qol: the civilian bounty control terminal displays the description of
new bounty options
/:cl:

* civilian bounty control terminal displays bounty description and payout of new bounty options

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: kawoppi <94711066+kawoppi@users.noreply.github.com>