[MIRROR] Removes a loose power cable that fails tests in Icebox Maint (#2345)
* Removes a loose power cable that fails tests in Icebox Maint (#83059)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
In the Electrical Maintenance area on Ice Box, there is a pair of SMESs
that are used for emergency power. One of the SMESs has a single, loose
power cable underneath it that does not connect to anything. It causes a
test to fail on the CI suite, so I made a PR to remove it.

The test in question:

![image](https://github.com/tgstation/tgstation/assets/110273561/9fac3846-46fe-474e-8d94-dfdbf82e31b8)

proof of testing:

![image](https://github.com/tgstation/tgstation/assets/110273561/48f76bde-1278-4b30-ad1d-3db0262dbbf4)
(Before)

![image](https://github.com/tgstation/tgstation/assets/110273561/74c199c5-9ca0-4fb9-878f-f38ed0317a1b)
(After)
It's GONE.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I am annoyed by this single mapping error, and by resolving it, I have
fixed every issue with ice box and we will never need to touch Ice Box
again. (Looks at the camera)
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: ReturnToZender (hitting delete on a power cable)
del: Single loose power cable in icebox maintenance
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

* Removes a loose power cable that fails tests in Icebox Maint

---------

Co-authored-by: Return <donwest947@gmail.com>