[MIRROR] Adds animation to objects sliding into or out of morgue trays (#2402) (#3248)
* Adds animation to objects sliding into or out of morgue trays (#83127)

## About The Pull Request


https://github.com/tgstation/tgstation/assets/51863163/b0662ee7-8f67-42d3-95ca-31012556f4ee

This effect is purely visual, objects are still "instantly" put into or
out of the morgue itself. (Meaning you can't grab an object as it is
being shifted in or taken out, nor can you move off the tray while it is
sliding in or out.)

Also applies to crematoriums. 

## Why It's Good For The Game

Adds a tiny bit of visual flair to help "sell" the effect that the tray
belongs to the morgue container, rather than just appearing and
disappearing.

## Changelog

:cl: Melbert
qol: Morgue trays (and the contents inside) are now animated on open and
close
/:cl:

* Adds animation to objects sliding into or out of morgue trays

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>