[Mirrors] A batch of missed mirrors #2442 (#3291)
* Coroner, Chief Engineer, Bitrunner are no longer exempt from certain Traitor objectives (#83165)

While reading traitor objective code for a project, I noticed that both
the Coroner and CE are not listed as potential targets for the 'Destroy
Heirloom' objectives. Looking into it, I noticed that the Coroner was
also not listed for the kidnapping objective. This fixes that.
Coroner should not be immune to traitory
And the CE should not be exempt from having their heirloom potentially
destroyed
:cl:
fix: The Coroner and Bitrunner can now be selected as a target for
kidnapping and heirloom destruction objectives.
fix: The Chief Engineer is now a valid target for heirloom destruction
objectives.
/:cl:

* Automatic changelog for PR #83165 [ci skip]

* Automatic changelog compile [ci skip]

* fix glass style /hooch being duped (#83169)

## About The Pull Request
/datum/glass_style/drinking_glass/hooch was being added 2 times, so I
just replaced the one with "aleglass" sprite to be /ale

* Automatic changelog for PR #83169 [ci skip]

* fixes spy bug, thank you melbert (#83153)

## About The Pull Request

~~removes protect objectives from spies,~~ actually just finds
objectives* fixing the 'nothing' or 'free objective' bug previously
present.

## Why It's Good For The Game

~~fix good until I can actually fix whatever's making protect objectives
not work. they aren't used anywhere else right now and work when
manually added after roundstart or with other active mobs around - I
suspect spies being a roundstart is the cause of this.~~

* Automatic changelog for PR #83153 [ci skip]

* Adds some autodoc to sound.dm (#83148)

Adds some autodoc to sound.dm

* Reimplements node compatibility mode (#83141)

## About The Pull Request
Puts a switch inside our build tools that will download the appropriate
node version based on your OS.

#82334

## Why It's Good For The Game
Closes #83076
Allows players to play the game

* Tgui: More common components in ts (#83098)

## About The Pull Request
Converts more of common into typescript, with some tests
## Why It's Good For The Game
Typescript,,,,

---------

Co-authored-by: norsvenska <73006946+norsvenska@users.noreply.github.com>
Co-authored-by: orange man <61334995+comfyorange@users.noreply.github.com>
Co-authored-by: Changelogs <action@github.com>
Co-authored-by: Xander3359 <66163761+Xander3359@users.noreply.github.com>
Co-authored-by: Higgin <cdonny11@yahoo.com>
Co-authored-by: Jeremiah <42397676+jlsnow301@users.noreply.github.com>