[MIRROR] Fixes a reversion to the greyscale plate armor (#2463)
* Fixes a reversion to the greyscale plate armor (#83214)

## About The Pull Request

https://github.com/tgstation/tgstation/pull/82288 reverted these sprites
to a prior, worse state. This restores the sprites.

## Why It's Good For The Game

Please check your icon diffs when you make a pr. Particularly if the PR
has been up for a really long time and things are being merged at the
same time. Do NOT simply merge from master and overwrite dmi's when a
conflict occurs on a .dmi file.

We slowly lose sprites over time due to this, and I might be the only
person who actually checks for these reversions.

## Changelog
:cl:
fix: Fixes an accidental reversion to greyscale plate armor.
/:cl:

* Fixes a reversion to the greyscale plate armor

---------

Co-authored-by: necromanceranne <40847847+necromanceranne@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>