[MIRROR] Protects space spawned humans from radiation nebula (#2477)
* Protects space spawned humans from radiation nebula (#83202)

Fixes #80864 

When a mob is spawned in space in a nebula, they get 1 minute of
radiation immunity. Ninja is the main one, but it'll also apply to other
space spawned mobs (lone nukie, space changeling?)

This only works for dynamic spawned antags. If an admin forces ninja
through an event, it doesnt work


![image](https://github.com/tgstation/tgstation/assets/7501474/98fbb499-86b3-43bc-bf63-403d7fa8df39)

:cl:
fix: Space Ninja and other space spawned antags get a 1 minute radiation
shield in the radioactive nebula
/:cl:

* Protects space spawned humans from radiation nebula

---------

Co-authored-by: Time-Green <7501474+Time-Green@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>