[MIRROR] Unreverts the fixes from #82713; the mining station once again no longer rapidly loses power (#2567)
* Unreverts the fixes from #82713; the mining station once again no longer rapidly loses power (#83319)

## About The Pull Request

#82537 unfortunately reverted my fixes in #82713 because of how old that
PR was when it was merged. So this restores those fixes.

## Why It's Good For The Game

The bane of map prs that stay up for a really long time is that this
happens a lot more than you would think.

## Changelog
:cl:
fix: Miners can actually access and fix their engineering issues on the
lavaland base via the engineering section of the base. AGAIN.
fix: The gulag SMES unit is no longer needlessly draining the entire
power grid of the main mining base. AGAIN.
/:cl:

* Unreverts the fixes from #82713; the mining station once again no longer rapidly loses power

---------

Co-authored-by: necromanceranne <40847847+necromanceranne@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>