[MIRROR] Fix issues resulting from an elevated object being created inside of a non-turf atom (#3526)
* Fix issues resulting from an elevated object being created inside of a non-turf atom (#83498)

## About The Pull Request

If an elevated object is initialized inside of a non-turf atom, it'll
still make the turf it is on elevated. Permanently. Which is weird.

## Why It's Good For The Game

Randomly elevated turfs are bad. Bugs bad.

## Changelog
:cl:
fix: Fix a rare issue where a turf would remain permanently "elevated"
if an elevated object was initialized inside of a non-turf object.
/:cl:

* Fix issues resulting from an elevated object being created inside of a non-turf atom

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: Lucy <lucy@absolucy.moe>
Co-authored-by: NovaBot13 <novasector13@gmail.com>