[MIRROR] Fixes Using 'ESC' Key To Unbind Keybindings + Adds Function to Check It Elsewhere [MDB IGNORE] (#3659)
* Fixes Using 'ESC' Key To Unbind Keybindings + Adds Function to Check It Elsewhere (#83710)

## About The Pull Request

Consequence of #80335 (d3554b3902384071415bb09e408ecebd9396fe5a)


![image](https://github.com/tgstation/tgstation/assets/34697715/cce22c30-e4d0-4f92-b8af-de908fbcce28)

Likely some weird IE shit, we get "Esc" instead of "Escape".

I added it as a function so this is a bit hardier to potentially
breaking so it doesn't go haywire silently again when we migrate to
Webview2. It's also likely that a weird combo of OS and IE and BYOND
Version and whatever will spit out "Esc" instead of "Escape", so let's
just have a helper function to access both so we don't have to worry
about it any more.

✅: Works on my machine
## Why It's Good For The Game

Players should be allowed to unbind whatever keys they want, regressions
are bad. New helper function that can check both cases (since we're
unsure of why either case pops up and it's unlikely that we have control
over it) was implemented in every spot where we were checking
`KEY.Escape` in this PR as well.

## Changelog
:cl:
fix: Using the 'ESC' key on your keyboard to unbind a key in the
keybindings preferences menu should now work as expected. This should
also be fixed for people in a variety of other spots too.
/:cl:

My javascript is a bit bad let me know if I'm doing something wack

* Fixes Using 'ESC' Key To Unbind Keybindings + Adds Function to Check It Elsewhere

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>