[MIRROR] Nerfs moon smile [MDB IGNORE] (#3008)
* Nerfs moon smile (#83248)

## About The Pull Request
This pull request makes the Moon Smile spell a worse alpha strike by
lowering the minimum duration of most of the effects and making the
knockdown only occur if the targets sanity is low enough. Currently its
knocking down at less than 40 sanity though this threshold can be
lowered if desired.
## Why It's Good For The Game
In its current state this spell is wayyyy too strong for how quickly you
get it and I've heard a lot of frustration expressed about it. This PR
attempts to tackle this by decreasing the minimum duration of its
effects and locking the knockdown to lower sanity levels making it
better in extended combat like the intent of the ability always was.
## Changelog
:cl:
balance: Moon Smile only does a knockdown if the targets sanity is low
enough and the minimum duration of its abilities have been decreased
/:cl:

* Nerfs moon smile

---------

Co-authored-by: Rex9001 <120136161+Rex9001@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>
Co-authored-by: SomeRandomOwl <2568378+SomeRandomOwl@users.noreply.github.com>