[MIRROR] RDS-havers now get Gibbering [MDB IGNORE] (#3771)
* RDS-havers now get Gibbering (#83850)

## About The Pull Request

Player characters afflicted with the RDS brain trauma now get the
gibbering language in addition to the ones they already know. the PR
looks like a webedit (it is) because i cannot clone repos because my
connection SUCKS but rest assured i tested it by just downloading it all
normally

## Why It's Good For The Game

someone who is insane may not always be able to articulate their ideas
in a way we might consider "language," either by accident or on purpose
just to show those stupid sheep. the greatest minds of the station need
to be able to speak to one another without persecution for their ideas

## Changelog
:cl: PowerfulAtom111
add: after a quick galactic meeting, insane people around the universe
can now speak gibbering to express their ideas free from the prying ears
of the walls
add: as an insane person, taking your meds deafens your ears to the holy
tongue all the other insane people are speaking
spellcheck: added a bit of text to the RDS quirk and the RDS medical
record text to highlight the insane's new abilities
/:cl:

* RDS-havers now get Gibbering

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: PowerfulAtom111 <owensalls@outlook.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>