[MIRROR] Misha - the bear [MDB IGNORE] (#3857)
* Misha - the bear (#83949)

## About The Pull Request

![image](https://github.com/tgstation/tgstation/assets/96586172/beeee166-f838-4bce-ab52-667ed566e25c)
Adds Misha the bear to icebox HoS office, a snow bear with the same
health stats as Sgt. Araneus.
## Why It's Good For The Game
With how easy it is to break into the HoS office on icebox (you don't
need any space protection) it never made sense to keep a slow pet in the
office. The bear is fast and deals 15 damage, while also resistant to
cold.
I think it adds more depth that the HoS tamed a bear on icebox and
trained it. Also the bear fits the map.
## Changelog
:cl: grungussuss
add: Added Misha the bear to the HoS office on icebox.
/:cl:

* Misha - the bear

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>