[MIRROR] Changelings are (once again) weak to fire [MDB IGNORE] (#3078)
* Changelings are (once again) weak to fire (#83953)

## About The Pull Request

Being on fire will make the ling unable to use any ability that isn't
for getting out of danger or reviving themselves (Adrenals, Fakedeath,
Strained Muscles, Last Resort, Shriek, Defib Grasp and Biodegrade). And
no, there is no armblade nerf, it is fine as it is.

## Why It's Good For The Game

Changelings are really busted, as they are effectively immortal until
completely removed (which means finding a cremator/gibber/shuttle, and
by the time you get there most of the time they are already ready to
revive) which really limits counterplay. Goof and Jac's approach to
nerfing lings was a bit too much since it nerfed the armblade massively
as well, but I found that their approach about lings being weak to fire
was really good counterplay for lings (besides BZ (who even uses it?)).
We shouldnt just leave the guys that are almost immortal and spawn
multiple in a round with one single counterplay that relies on a
knowledge check.

## Changelog

:cl:
balance: Certain changeling abilities won't work while on fire.
/:cl:

* Changelings are (once again) weak to fire

---------

Co-authored-by: JackEnoff <166665535+JackEnoff@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>