[MIRROR] Allows fire extinguishers to refill from plumbed tanks [MDB IGNORE] (#3876)
* Allows fire extinguishers to refill from plumbed tanks (#84004)

## About The Pull Request
Changes `tanktype` to be a list and `AttemptRefill` to properly check
for a list. Also adds a `has_reagent` check to `AttemptRefill` to
prevent people from converting a fuel tank to plumbed to refill a fire
extinguisher with welding fuel.

## Why It's Good For The Game
People be allowed to refill fire extinguishers from the plumbed variant
of tanks. Also I thought it would be funny if we could refill
extinguishers from water coolers.
fix: #68174

## Changelog
:cl: Goat
qol: fire extinguishers can now be filled via stationary tanks. (and
water coolers)
/:cl:

* Allows fire extinguishers to refill from plumbed tanks

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: Goat <126099705+Goat-Real@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>