[MIRROR] Mining borgs can now get multiple copies of a single modkit again [MDB IGNORE] (#3889)
* Mining borgs can now get multiple copies of a single modkit again (#84052)

## About The Pull Request

A previous pull request cleaned borg upgrade code up a lot, and included
a check to prevent duplicate upgrades from being installed. Usually this
is a good thing, but there are some upgrades that are intended to allow
duplicates. Namely, PKA modkits in mining borgs. This PR just fixes
that, so mining borgs can have multiple modkits again. This does NOT
allow borgs to go above the 90% complexity limit on their PKAs, as a
separate piece of code handles that.
## Why It's Good For The Game

Bugs are bad.
## Changelog
:cl:
fix: Mining borgs can get multiple modkits of the same type installed
again.
/:cl:

* Mining borgs can now get multiple copies of a single modkit again

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: GPeckman <21979502+GPeckman@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>