[MIRROR] Transcendent Olfaction fix: Pesky closets [MDB IGNORE] (#3893)
* Transcendent Olfaction fix: Pesky closets (#83993)

## About The Pull Request

Change to using the more robust get_turf method to get locations, as all
pinpointers currently do.

## Why It's Good For The Game

The transcendent olfaction mutation is currently broken. When the target
is inside a container, its x y z becomes zero, so it will always say the
target is below you when comparing your z level to it, and it won't even
point in the right direction. This PR fixes that.

## Changelog

:cl:
fix: Transcendent Olfaction mutation now works properly
/:cl:

* Transcendent Olfaction fix: Pesky closets

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: Kaostico <g-a-f92@hotmail.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>