[MIRROR] Adds boards for the book binder and scanner [MDB IGNORE] (#4033)
* Adds boards for the book binder and scanner (#83863)

## About The Pull Request
Adds boards for the book binder and scanner. The scanner requires one
scanning module while the binder requires one servo to make. Research
for the board was added to the computer tech research node as that's
where the other book related research was. Boards are available to print
at the service techfab.

![libraryboards](https://github.com/tgstation/tgstation/assets/126099705/ce6f4174-9939-4020-a7d4-c0c6593479f3)

## Why It's Good For The Game
Being able to (re)make the library is good. Irreparable machine is bad.
Writing down your work and being unable to scan it into the archive is
not good. Also allows the machine to be moved if needed.

* Adds boards for the book binder and scanner

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: Goat <126099705+Goat-Real@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>