Fixes NTR getting Syndicate Gloves (#4197)
* Fixes NTR getting Syndicate Gloves

The NTR's had been getting Syndicate branded combat gloves, which allowed for early Illegal tech. Made it so they only get the CC gloves they were already getting in their garment bag and removed the Syndicate ones. If someone wants the black ones back because of their look they are welcome to make a child of those and assign the  syndicate ones icons.

* Added CC black gloves

Made a child of the CC gloves and applied the syndicate gloves skin on them

* Reverts admins.txt

stupid github desktop.

Co-authored-by: OrbisAnima <orbisanima@gmail.com>