[MIRROR] [MDB IGNORE] Gives telescreens mapping directionals + adds a map lint for them [MDB IGNORE] (#3484)
* [MDB IGNORE] Gives telescreens mapping directionals + adds a map lint for them (#84261)

Title.

The telescreens that are used primarily in head offices are.. a mess.
The parent of all telescreens is also a gigantic mess.

*I've opted AGAINST making an updatepaths here and decided to hand fix
all of these. There's not a lot, but I can't make an edge case for
everything. Some of their nudges are a bit wonky too, and I want to make
sure noting looks off. Theres also a handful of prefabs on the parent
telescreen that I want to individually look at.*

Updatepaths or not, this is going to be a "smaller" one of "those" PRs
that I've done thusfar, or so I think.

Prevents more mapping sin like this (THERE ARE SUBTYPES WHY ARE YOU
MAKING "NEW" ONES ON THE PARENT?? STOPPPP)

![image](https://github.com/tgstation/tgstation/assets/70232195/020a2e03-85ce-43f8-8eee-ad3764655c82)

The sprite edits are necessary to make sure the mapping helper macro
works, otherwise they'll be inverted.

N/A

* erm

* our maps + new telescreens that were needed

* well fuck me i guess

* mfw