[MIRROR] Adds a breathing sound while using internals of any kind [MDB IGNORE] (#4265)
* Adds a breathing sound while using internals of any kind (#84529)

## About The Pull Request

Spess men will now hear their own breathing while using internals, most
noticeable while on space. It's low enough that it shouldn't affect
anything normally, and only the person with internals can hear it.
If you have misophonia or for some other reason dislike the sound, it
can also be toggled off in the game settings.

## Why It's Good For The Game

It truly makes you FEEL like you're breathing via a mask connected with
an oxygen tank. It also, potentially, can remind you that your internals
are on when it's not needed.

## Changelog

:cl: DATA_, with sounds made by Beebblie
sound: You will now be able to hear your muffled breathing while using
internals. This can be turned off in the game settings.
/:cl:

---------



* Adds a breathing sound while using internals of any kind

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: DATAxPUNGED <44149906+DATA-xPUNGED@users.noreply.github.com>
Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>