[MANUAL MIRRORS] Assorted missed workflow PRs (#4300)
* Fixes Rust Caching for Run Linters (#80168)

## About The Pull Request

Ripgrep got updated some time ago, and that means that our cache went
out of date as we had to keep doing more work to install the new
dependencies that it has. This means we were eating up another **50
seconds** every time we ran Run Linters. Unfortunately, github caches
will only hash to a file that exists on the repository, not anything
that might exist after a tool is installed.

So, in order to do this, we create a new bash file `install_ripgrep.sh`
and also pin the ripgrep version. If we need to reap some new benefits
or it's substantially faster (which might not be the case as ripgrep is
already blazing fast), we just need to update `ci_dependencies.sh` and
the cache should work for future versions without us needing to purge
rust caches manually through github.

You can view a successful run here
https://github.com/san7890/bruhstation/actions/runs/7123029395/job/19394997831

I also did something similar for SpacemanDMM while i noticed it.

* Fix alternate tests erroring when there are no alternate versions (#80287)

It wasn't exiting gracefully because of a typo? here

* Fix flaky wizard screenshot test failure, and try to make screenshot tests run more consistently (#80631)

From https://github.com/tgstation/tgstation/pull/80542

Not exactly sure why this only fails once. My hint is that it happens
when the alternate tests are skipped from this:

![image](https://github.com/tgstation/tgstation/assets/35135081/67c261c0-e771-4696-add2-f41bb800ad65)

...though this should still be *consistent* at least.

* Improves OpenDream Linting (Annotations and More) (#82029)

## About The Pull Request

This PR does a few things.

The main goal was to have it annotate the code like we do for `Run
Dreamchecker`, which it now does. The code for this was mostly
shamelessly ripped from the very same process I just mentioned, but
stripped down a lot more for the task. Since I had to write a lot more
code to handle the way we get warnings and the like from OpenDream, I
didn't fold both into one program.


![image](https://github.com/tgstation/tgstation/assets/34697715/f4a05b59-5407-412e-a73f-5c069fdb4b02)


![image](https://github.com/tgstation/tgstation/assets/34697715/0d3335aa-0da1-45ed-af19-1d17a7fc4174)

It works! Ta-da.

Anyways, in order to facilitate the code annotation, we need to use a
Python script. This Python script means that we need to use the
bootstrap, and the bootstrap is best used with the bootstrap cache. I
then realized that we should just really fold this special CI runner
into the Run Linters workflow anyways, instead of having it on its own
runner (since it'll cut into the number of runners we have available at
a time anyways).

I think it's a good way to save precious CPU cycles on our dying Earth,
and the issues are now visible enough to the average coder (in the
`Files Changed` tab) to where I feel pretty good about doing this.

There's a bit of weirdness and jank in the Python program, but it's all
done for a reason I assure you (it's always silly when stuff breaks
silently and no one knows to fix it until years later). You can look an
example of a failing run here:
https://github.com/san7890/bruhstation/actions/runs/8303759645/job/22728427623#step:17:14

* Fixes CI not checking warnings. (#83015)

Yeah.

* Cancel CI on new commit (#83255)

This sets concurrency at the workflow level, which should provide a more
concise way to eject if a new commit is entered

https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#concurrency

I've also added timeouts that are between 5-15 minutes per test. Default
is 360 minutes without this

* Lowers frequency of Docker Publish and Generate Documentation changes to lower throughput (#71294)

Docker Publish is now a nightly.

Generate Documentation now runs every 6 hours.

Both used to run every commit, and our CI is getting heavily
bottlenecked. Need to ease some stress.

---------

Co-authored-by: FlufflesTheDog <piecopresident@gmail.com>
Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: vvvv-vvvv <136390975+vvvv-vvvv@users.noreply.github.com>
Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>
Co-authored-by: AnturK <AnturK@users.noreply.github.com>
Co-authored-by: Jeremiah <42397676+jlsnow301@users.noreply.github.com>