[MIRROR] Adds the Holy Skateboard, a holy version of the hoverboard. [MDB IGNORE] (#3580)
* Adds the Holy Skateboard, a holy version of the hoverboard. (#84514)

## About The Pull Request
This is my first PR ever, and I have tested it. It just adds an
additional nullrod variant that could be helpful for people that wish to
explore space as a chaplain, or chaplains that wish to look sick and
tubular.
## Why It's Good For The Game
More nullrods are better. Also, it's a fun vehicle that i feel allows
for more gimmicks as the 540 cr cost for a regular hoverboard is a bit
steep. It's not super good, and it's not bad, so I feel it doesn't
affect balance too much.
## Changelog
:cl:
add: There's a previously undiscovered variant of a nullrod, recently
revealed from the depths of the skating rink..
/:cl:
<!-- You can use multiple of the same prefix (they're only u


https://github.com/tgstation/tgstation/assets/126758839/5c94391d-8a01-40fd-93ec-ea3617ba6a59

sed for the icon ingame) and delete the unneeded ones. Despite some of
the tags, changelogs should generally represent how a player might be
affected by the changes rather than a summary of the PR's contents. -->

* Adds the Holy Skateboard, a holy version of the hoverboard.

---------

Co-authored-by: Loafin34 <126758839+Loafin34@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>