[MIRROR] Limits the methods through which you can be stunned by a shove. Shoves disarm people who are knocked down. [MDB IGNORE] (#4321)
* Limits the methods through which you can be stunned by a shove. Shoves disarm people who are knocked down. (#84640)

For the sake of this PR's language, we'll be defining two different
stuns;

First: **wall stuns** are any stuns related to the shoving mechanic
where you are pushing people into objects and other people to cause a
knockdown. And, while they are in that specific knockdown, you shove
them again to stun them.

Second, a **shove stun** is any instance where you would be currently
able to apply this stun from a shove even if the conditions above have
not been met.

This PR limits down the potential stuns from a shove to; **wall stuns**,
and **shove stuns** following a telescoping baton knockdown or being
touched by the Mansus Grasp.

To know when you can stun someone with a shove, they will have this
animated visual cue.

![image](https://github.com/tgstation/tgstation/assets/40847847/123a47f2-93eb-4c25-a8e7-a7ae62e33c7f)

Successfully stunning, paralyzing or immobilizing the victim will remove
the opportunity for a stun. You cannot apply another stun from a shove
on someone for at least 3.5 seconds after the last stun was applied. The
stun lasts for 3 seconds, so there is a 0.5 gap after the stun expires
before another stun from a shove can be applied. This does not prevent
further knockdowns or staggers. This part was added mostly to help
clarify better when someone is or is not vulnerable, as well as cut down
on chain stuns without some additional attacks.

To make up for the loss of this advantage, shoving someone while they
are in any kind of knockdown forces them to drop their held items.

You may have remembered the removal of shove stuns testmerge a bit ago.
[You can read a bit about it here, but I'll link specifically to my
conclusions about that
testmerge.](https://tgstation13.org/phpBB/viewtopic.php?p=726119#p726119)

At the moment, shove stuns are very oppressive. Generally, it makes a
few engagements against a variety of tools too dangerous to be worth
attempting, and can often mean that a fight has already been won not
from the intelligent employment of a tool, but the mere existence of one
taking advantage of a very specific combination of actions. The most
egregious part is that it enables an individual who knows the patterns
to chain these stuns together to deny someone a chance at fighting back.

Wall stunning, however, is a good equalizer that punishes people for not
considering their environment, and rewards players who are. It allows
the clever player to maneuver themselves just right to score a powerful
reversal against even better equipped opponents, or a strong opener to
allow them a chance at a successful attack. And it is a strong example
of our environmental combat mechanics. Something that shove stuns in
isolation is not necessarily (though it can be in some cases).

Shove stuns are not exclusive to these scenarios, however. They exist as
a ubiquitous means of turning any given knockdown effect into a chain of
stun effects. Better equipped opponents, particularly ones with gear
able to apply knockdowns, often find themselves able to readily match
any opponent without having to worry about the aforementioned
environmental aspect of shoving. It is all the advantages with none of
the effort put in, and does play into any kind of struggle once the
chain effects start to take place.

We want to enable players to be able to have their big win moments
against the odds. But we also don't want to be rewarding players for
already being in an advantage position beyond what their gear affords
them.

Edit: With the new visual cue, I've allowed in a few shove stuns for
equipment that somewhat needs it in order to function well. Heads of
staff utilizing their telebatons are meant to be reasonably intimidating
opponents, contractors absolutely need the advantage in order to
successfully kidnap, and Heretics were one of the biggest sufferers
during the testmerge and clearly also hinge quite a bit on the effect.

:cl:
balance: You can only stun someone with a shove if they have been shoved
into objects or people, or have been hit by telescopic batons,
contractor batons or the Mansus Grasp.
balance: If someone is knocked to the floor, however, you can still
disarm them of any object with a shove.
/:cl:

---------



* Limits the methods through which you can be stunned by a shove. Shoves disarm people who are knocked down.

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: necromanceranne <40847847+necromanceranne@users.noreply.github.com>
Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>