[MIRROR] Fixes grounding rods [MDB IGNORE] (#4424)
* Fixes grounding rods (#84839)

## About The Pull Request

Grounding rods didn't have a circuit properly defined. This meant that
mapped-in grounding rods (and presumably admin spawned ones) could not
be upgraded and did not yield any parts or a circuit board when
deconstructed. Now they do have a circuit defined.

Fixes #84588 
## Changelog
:cl:
fix: Mapped in grounding rods can be upgraded again (not that upgraded
parts do anything for it) and don't disappear when deconstructed
anymore.
/:cl:

* Fixes grounding rods

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: GPeckman <21979502+GPeckman@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>