[MIRROR] Fixes the Metastation cargo ripley not being able to charge. [MDB IGNORE] (#4443)
* Fixes the Metastation cargo ripley not being able to charge. (#84863)

## About The Pull Request
Adds a mechbay power console to the warehouse so that the cargo ripley
is actually able to charge with the roundstart setup. I had to rearrange
the setup a bit to fit things nicely.

## Why It's Good For The Game
the cargo ripley is possibly one of the coolest things and it should
have a somewhat functional charging setup by default instead of half of
one. The old version of Metastation cargo had the console so it was lost
in the cargo rework.

## Changelog
:cl:
fix: Added a Mechbay power console to the Metastation cargo warehouse so
that the cargo ripley is able to be charged.

/:cl:

* Fixes the Metastation cargo ripley not being able to charge.

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: aaaa1023 <74441292+aaaa1023@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>