[MIRROR] You can implant yourself with pathfinder MOD implant while you're wearing the suit [MDB IGNORE] (#4477)
* You can implant yourself with pathfinder MOD implant while you're wearing the suit (#84927)

## About The Pull Request
Pathfinder MODule now starts as a usable module which can be triggered
in order to implant MOD's wearer.

## Why It's Good For The Game

Its a very cool module which unfortunately is severely hampered by
having to open the suit and pull it out to eject. This doesn't have any
impact on balance, but does make people more likely to use it in-game as
its now not as bothersome to set up. Who even knew that captain's
MODsuit has it?

## Changelog
:cl:
qol: Pathfinder MODule can now be triggered while wearing the MODsuit to
implant yourself without having to pull it out of the suit.
/:cl:

* You can implant yourself with pathfinder MOD implant while you're wearing the suit

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: SmArtKar <44720187+SmArtKar@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>