[MIRROR] Synth and space heater energy fixes/adjustments (#4564)
* Synth and space heater energy fixes/adjustments (#3707)

* fix space heaters

* adjusts synth values

less magic numbers

* force drain, allows for fully depleting a cell

* code cleanup, decrease rate, allow full charge

* use LEVEL_FULL, don't slow down charge rate

Switching everything else to use LEVEL_FULL too, consistency.
ALMOST_FULL is supposed to be for auto-regenning nutrition, this'll
widen the range before mood debuffs

also now that we have a a reasonable cell capacity (equivilant to a t2
cell) we don't need to artificially limit charge rate. It'll be pretty
much instant with a fully upgraded charger, but it would be for a borg
too. Scale by JOULES_PER_NUTRITION instead

* persocom uses LEVEL_FULL

* Move charge level defines, remove a silly one

* add a little bit of a buffer for charging, re-ordering

protects against floating point weirdness, and gives people a few
while where you'll still be "fully charged"- but not enough that you
can't stay at a satisfying level.

* better feedback

* cell update appearance

* just use normal nutrition here, cap to be safe

* borg charger hack

one day we should refactor synths to use an actual cell, but today is
not that day.

* boosts borg chargers

* [MIRROR] Synth and space heater energy fixes/adjustments

---------

Co-authored-by: FlufflesTheDog <piecopresident@gmail.com>
Co-authored-by: StealsThePRs <StealsThe@Email.com>