[MIRROR] You can now fish with explosives [MDB IGNORE] (#3933)
* You can now fish with explosives (#85118)

## About The Pull Request
Iansdoor suggested this to me earlier (but I already knew it was a thing
on Goon), and I said ok. Basically, this PR adds an interaction between
explosions and fishing.

At the cost of potential collateral damage, you can too now use a TTV to
catch a lot of fishes (and other stuff). Fish spawned this way will be
pretty much dead however.

This PR also adds the examine tips already presents in the fishing spot
component to the lazy fishing spot element.

## Why It's Good For The Game
This PR adds an interaction between fishing and other mechanics of the
game.

## Changelog

:cl:
add: You can now fish with explosives.
fix: Fixed an inconsistency with examining fishing spots with
sufficiently high fishing skill (or skillchip).
/:cl:

* You can now fish with explosives

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>