[MIRROR] Surgery steps no longer fail silently when missing chemicals  (#4744)
* [MIRROR] Surgery steps no longer fail silently when missing chemicals [MDB IGNORE] (#4038)

* Surgery steps no longer fail silently when missing chemicals (#85260)

## About The Pull Request

Fixes surgery steps that require a chemical from failing silently if
that chemical is missing. Player receives a bubble notification and the
surgery is not performed if the chemical is missing. Operating computer
UI tells you if the required chem is present or not.


![dreamseeker_RMO06bKPcV](https://github.com/user-attachments/assets/c57e5615-0786-4e8c-a2ef-d268f1c82f7d)

## Why It's Good For The Game

Having to guess, without feedback, if the chemical was in the target's
system at the time surgery completes is bad.

## Changelog

:cl: LT3
fix: Players now receive a notification when trying to perform surgery
steps that involve chemicals
/:cl:

* Surgery steps no longer fail silently when missing chemicals

---------

Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>

* [MIRROR] Surgery steps no longer fail silently when missing chemicals

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>
Co-authored-by: StealsThePRs <StealsThe@Email.com>