[MIRROR] Fixes exploit allowing players to open Bag of Holding rifts in the Heretic dimension  (#4783)
* [MIRROR] Fixes exploit allowing players to open Bag of Holding rifts in the Heretic dimension [MDB IGNORE] (#4065)

* Fixes exploit allowing players to open Bag of Holding rifts in the Heretic dimension (#85330)

## About The Pull Request

This PR prevents players from opening a BoH tear in the Heretic
dimension areas. It also introduces a new area flag, "NO_BOH", which can
be added to areas to replicate this behavior.

## Why It's Good For The Game

This prevents the (very unlikely) edge case that someone manages to open
a BoH tear in one of the funny hand rooms, which is bad not only because
it fucks over anyone else who gets sacrificed, but could also interfere
with other things on that z-level (shuttles, other Lazy Template areas,
etc).

Messing with some new code here, so if there's better ways to handle
this, by all means, chime in. This is a much more elegant method than
just banning anything but the station.
## Changelog
:cl: Vekter
fix: Fixes an exploit that allowed players to open a Bag of Holding rift
in the Heretic dimension.
/:cl:

---------

Co-authored-by: Kapu1178 <75460809+Kapu1178@users.noreply.github.com>

* Fixes exploit allowing players to open Bag of Holding rifts in the Heretic dimension

---------

Co-authored-by: Vekter <TheVekter@users.noreply.github.com>
Co-authored-by: Kapu1178 <75460809+Kapu1178@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>

* [MIRROR] Fixes exploit allowing players to open Bag of Holding rifts in the Heretic dimension

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: Vekter <TheVekter@users.noreply.github.com>
Co-authored-by: Kapu1178 <75460809+Kapu1178@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>
Co-authored-by: StealsThePRs <StealsThe@Email.com>