[MIRROR] Added a config to override the def gateway delay for specific start points.  (#4791)
* [MIRROR] Added a config to override the def gateway delay for specific start points. [MDB IGNORE] (#4084)

* Added a config to override the def gateway delay for specific start points. (#85341)

## About The Pull Request
I've a few gripes with start points and the gateway delay. First of all,
there's no way to discriminate peaceful away locations that do not need
a with a 30 minutes timegate from the rest. Places like the beach and
the museum hardly have anything OP that could tip the scales.

Second, none of the awaystart landmarks have identifiers of their own,
which means all awaystart landmarks from all away missions are linked
under the same destination point datum. This is hardly an issue in the
current state where only one map is ever loaded and all maps have only
one way in that directs you to one of several locations at least until
the gateways are linked, but it's nevertheless something that I have to
take care of, since the config requires it.

## Why It's Good For The Game
See above.

## Changelog

:cl:
config: Added a config for specific gateway delays so locations like the
beach and the museum don't have to take 30 minutes to become available
like the rest.
/:cl:

* Added a config to override the def gateway delay for specific start points.

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>

* [MIRROR] Added a config to override the def gateway delay for specific start points.

---------

Co-authored-by: NovaBot <154629622+NovaBot13@users.noreply.github.com>
Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>
Co-authored-by: StealsThePRs <StealsThe@Email.com>