replaces the unusable energy cannon in mystery boxes to the new smartgun (#85529)
## About The Pull Request

see title
## Why It's Good For The Game

It's been a really long time since anyone other than a turret could
shoot the energy cannon (free tasers for everyone!!) so it's an entirely
useless weapon roll. I'm replacing it with the new smartgun because the
smartgun is cool and more importantly you can actually shoot it.
## Changelog
:cl:
add: The new smartgun has replaced the unusable energy cannon in mystery
boxes
/:cl: