Reintroduces custom pies and cakes (#86013)
## About The Pull Request

Fixes (Or so I'd guess) part of the #59312 and #81480 issues.

I'll do kebabs and sandwiches in their own atomized PRs as they actually
need some codework

## Why It's Good For The Game

Long before time had a name custom pies and cake were created by the
First Chef Master by using the four Weapons of Chefing: The Knife of
Quakes, the Microwave of Lightning, the Freezer of Ice, and the Oven of
Fire. Weapons so powerful no one can handle all of their power at once.

When he passed away his two sons swore to protect them, but the oldest
was consumed by darkness and destroyed the custom recipes forever with a
rework. A battle between brothers broke out and the oldest was struck
down and banished to the Underworld with the power of this PR.

![image](https://github.com/user-attachments/assets/7f3c377d-d80f-4ba4-8cc3-523b4d576d98)

![image](https://github.com/user-attachments/assets/d3030053-9a05-4bf1-971f-396c09fd1005)

## Changelog

:cl:
fix: custom pies and cakes are craftable again
/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>