Fixes the text of the instructions paper for lures. (#86537)
## About The Pull Request
I'm not a good writer, there're a couple instances that could've been
worded better, and a typo or two, however the biggest problem were two
bold tags not being properly closed on the last paragraph.

## Why It's Good For The Game
Spellchecking.

## Changelog

:cl:
spellcheck: fixed a few wording and markdown issues on the instructions
paper for fishing lures.
/:cl: