fixes leaper polymorph outbreak (#86559)
## About The Pull Request
fixes a bug where if people ride animals that share abilities with their
rider, and this animal happens to be polymorphed, players would be able
to turn into it by clicking that ability. this pr makes it so riders are
unable to use the ridden mob's polymorph ability, or if they are, it
will only transform the ridden.

## Why It's Good For The Game
closes #85114 , closes #85122

## Changelog
:cl:
fix: fixes being able to transform into polymorphed mobs by riding them
/:cl: