Slims down the aquarium content component, moving most of its code to the items it's attached to. (#86104)
## About The Pull Request
If I ever want to convert aquariums to component so that we get portable
fish tanks, I'll have to remove the garbage code from aquarium contents
first, because it's a bit confusing to look at, and most of it is
hard-coded or about fish animations, which is annoying because it'll
also prevent me from making different animations for different fish in
the long run. This PR should replace some of the mess with a few signals
so that we can have slightly more freedom around the visuals of aquarium
contents.

#86090 should be merged first by the way to avoid issues with missing
visuals.

## Why It's Good For The Game
See the previous section.

## Changelog
N/A