Fixing fishing code not caring whether a reward is limited or not. (#86125)
## About The Pull Request
At some point, I must have removed the check for limited items from
`get_modified_fish_table()` and never replaced the `fish_table.Copy()`
call with `get_fish_table()` (which has it), causing limited rewards to
be yet `again` unlimited. This extends to explosive and lobstrosities
(clientless mobs with the profound fisher comp) once the associated
value of the reward in the `fish_counts` list (the list of keeps track
of these limited rewards) reaches a negative value, because the code
only checks if the value is equal to zero where it frankly it should
check if the value is equal or less than zero.

Also I've set `explosion_malus` to true for a few fish sources that
cover multiple turfs, revisited a few comments and a redundant
UnregisterSignal call in lazy_fishing_spot

## Why It's Good For The Game
Fixing an oversight.

## Changelog

:cl:
fix: Fixing fishing not caring whether a reward is limited or not.
/:cl: