Fix up the information on some void heretic abilities; fixes a typo and makes another more descriptive (#86611)
## About The Pull Request

Fixes a typo in one part of the description for Void Conduit, and
updates a menu description for it to be more informative of what it
does.
## Why It's Good For The Game

Good descriptions may one day make our players want to read good and
maybe do other stuff good too
## Changelog
:cl: Bisar
fix: Fixed up a couple of Void Heretic rework descriptions.
/:cl: