ethereal cell is now stored in contents instead of nullspace (#86586)
## About The Pull Request

ethereal cell is now stored in contents instead of nullspace

## Why It's Good For The Game

i mostly made this because i realized that the recharging function of
relics doesn't actually work on these, and i found out it's because this
isnt in contents but nullspace

seems kind of lame for a cool interaction i hoped for (ethereal charging
via relic) and possibly other interactions in the future as well

it might cause a disaster somehow. but let's yolo it and revert if it
causes anything awful

## Changelog

:cl:
fix: The recharging relics now work on ethereal cells.
balance: EMP now affects ethereal's hunger.
/:cl:

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>