The fishing line and float are now actually offset to match the fishing spot (#86659)
## About The Pull Request
I've made a PR before so that beam datum (and therefore the fishing
line) would take pixel_w/z into account, but what I didn't consider is
that the fishing line is actually attached to the float (as it should
be), so it did basically nothing on the other end of fishing lines while
fishing. This PR fixes that.

## Why It's Good For The Game
Fixing a small visual inconsistency.

## Changelog
N/A