Fixing a runtime with examining fishing spots while wielding a rod (#86689)
## About The Pull Request
Well, it seems examining a fishing spot (with sufficient skill) while
wielding a rod is borked because of a bad line. It should be `init_name`
and not `reward`. Thanks to vinylspider for informing me about it.

## Why It's Good For The Game
Fixing stuff I made and then broke.

## Changelog

:cl:
fix: examining fishing spots while wielding a rod (with sufficient
skill) now works.
/:cl: