Fixes some nonhuman surgery oversights (#86749)
## About The Pull Request

Added a null check to mood, because it caused runtimes. Nonhumans have
mood, but they never initialize it, yet surgery mood assumes it is.

Removed unnecessary ishuman check in the organizer, allowing it to work
on xenomorphs and other theoretical carbons.
## Why It's Good For The Game

> Added a null check to mood, because it caused runtimes. Nonhumans have
mood, but they never initialize it, yet surgery mood assumes it is.

Bug bad

> Removed unnecessary ishuman check in the organizer, allowing it to
work on xenomorphs and other theoretical carbons.

Oversight bad
## Changelog
:cl:
fix: Added a null check to mood, because it caused runtimes. Nonhumans
have mood, but they never initialize it, yet surgery mood assumes it is.
fix: Removed unnecessary ishuman check in the organizer, allowing it to
work on xenomorphs and other theoretical carbons.
/:cl: