Resprites Short Bangs 2 and Double Bun (#86780)
## About The Pull Request
Resprites Short Bangs 2 and Double Bun to look cleaner and makes the
hair bangs look more consistent when looking at a side, as well as
making the sidelocks show up on the back sprite.

Short bangs 2 on the top, double buns on the bottom.

![dreamseeker_6GNR5LuDah](https://github.com/user-attachments/assets/d8f81bc2-dfc1-417e-81bc-11f729f22956)

![dreamseeker_CVEL2yVrxS](https://github.com/user-attachments/assets/861faf87-5867-4082-b143-3fd604fbe4f8)

![dreamseeker_cjTiuzpMw7](https://github.com/user-attachments/assets/9f4989b5-079f-450e-8f6d-4b0b701c7914)

## Why It's Good For The Game
Makes the hair look cleaner and fit the modern spriting standards.
The older version's palette was also.. interesting, it needed to be
reduced.

![image](https://github.com/user-attachments/assets/759d2e7a-0866-4bd4-a5ff-fd0dabb555c3)
## Changelog
:cl: Hardly3D
image: Resprited Short Bangs 2 & Double Buns
/:cl: