Account job changes account for curator patronage, including VVedit job changes (#86744)
## About The Pull Request

In #86711 we made it so adjusting the trim of an id card also updated
the job on the associated bank account to match, but we forgot to update
`bank_accounts_by_job` to match this.
This made it so Curator painting patronage cuts wouldn't get updated to
match the trim, and would risk issues with removing it from
`bank_accounts_by_job` if the bank account were to get deleted.

In this pr we add a `update_account_job_lists(...)` proc that updated
`bank_accounts_by_job` to match.
This fixes our issues.

Additionally, we update `vv_edit_var(...)` to also run this when
changing related values, such that it actually removes and adds the
values when needed.
## Why It's Good For The Game

Fixes a bug, helps with VVediting bank account jobs.
## Changelog
:cl:
fix: Changing a bank account's job to or from Curator actually changes
whether they get a cut from painting patronage.
admin: VVediting a bank account's account_job actually updates what job
the account is associated with. Currently only matters for Curators.
admin: VVediting a bank account's add_to_accounts actually removes it
from or adds it to the job to account associations. Currently only
matters for Curators.
/:cl: