Fix air alarms warning message to use pressure & temp settings (#86917)
## About The Pull Request
Air alarms had temp/pressure that were hardcoded to be dangerous
temps/pressure for humans. This fixes it so that the temp/pressure is
based on the settings used in the air alarm. So if a air alarm is
ignoring pressure/temp then it doesn't give a warning message.

## Why It's Good For The Game
If you want to ignore atmos checks for air alarms in certain areas, then
it shouldn't be giving warning messages.

## Changelog
:cl:
fix: Fix air alarms warning message to use pressure & temp settings
/:cl: