Birdshot Cargo Changes (#86769)
## About The Pull Request

Some QoL changes for cargo, mostly just with the goal of cutting down on
the sheer amount of airlocks in cargo. I also added a small cargo lobby
to aid in pushing the ORM a tad closer to the miners and cutting down on
the airlocks in and out of cargo.
## Why It's Good For The Game

Less claustrophobic, more space. The changes are very minor, just small
changes to the layout without completely overhauling it.

You can see the changes below.

![image](https://github.com/tgstation/tgstation/assets/168238458/fd087ff0-71b7-4860-a2e2-11df4873c517)

re-added dnd room:

![image](https://github.com/tgstation/tgstation/assets/168238458/27a29d84-6604-4d87-a45c-d74ab2fc8761)

new cargo lobby area:

![image](https://github.com/tgstation/tgstation/assets/168238458/3db44063-d8b9-46f7-a72a-f515149a7e57)

the freed up space:

![image](https://github.com/tgstation/tgstation/assets/168238458/5d42cb0a-07db-4554-8e2d-09eee02c5110)

## Changelog
:cl:
add: Birdshot Cargo now has a lobby and is slightly more open overall.
add: Re-added the DnD room  to Birdshot
add: re-added the skill station to the library.

/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>