Better ghost tray scan (#87143)
## About The Pull Request

Partially improves:
* #46002

Previously, the ghost's "t-ray view" acted as a toggle, even though the
scan only occurred once. This PR changes it to function as an action
rather than a toggle, removing the need to press the button twice.

While the ideal solution would be for the t-ray scan to work on the
process tick, my attempt to give ghosts their own t-ray scanner and to
toggle it on upon using the verb did not work as expected because the
scanner didn't recognize the ghost as its holder (`mob/viewer` in the
code).

This update serves as an improvement to the current behaviour, but in
the future, the t-ray scan verb should function like a normal t-ray
scanner, instead of just a one-time scan.
## Why It's Good For The Game

Improves the t-ray verb, before it acted as a toggle even though it only
worked once upon toggling on. Improves user experience using this verb.
## Changelog
:cl:
fix: the ghost "t-ray view" no longer requires a double press to
activate a second time.
qol: renamed "t-ray view" to "t-ray scan" for clarity.
/:cl: