Fixes oversight of the Springlock MOD module and its interaction with water vapor (#87169)
## About The Pull Request
Adds the interaction between the Springlock MOD module and the Gas Water
Vapor so that it snaps shut upon being exposed to the Gas.
Fixes #85666.
Specifically the oversight mentioned in the comments.
## Why It's Good For The Game
Adds consistency to the Springlock MOD module so it properly interacts
with water as intended.
## Changelog
:cl:
fix: Springlock MOD module properly interacts with Water Vapor
/:cl: