syndicate sleeper agents can forge custom objectives (#87180)
## About The Pull Request

The removal of the ability for midround traitors to take progression
objectives had the run-on effect of preventing them from using their
uplink to view their primary objectives, which also means they cant
forge custom objectives. This PR changes that, and provides additional
logic in uplink ui code to account for uplinks belonging to traitors
with one type of objective but not the other.

## Why It's Good For The Game

Fixes #86763