Adds Better Bombanana Logging (#87249)
## About The Pull Request

Fixes #87214

Better bombanana logging that should account for all the edgecases
mentioned in the above issue report, giving more awareness on why a
bombanana peel spawned, as well as renaming the bombanana syndicate
minibomb to actually have the name "bomb". Just does a lot of
retinkering and cleaning up to ensure all the args are appropriate and
that everything should be fed into the logging procs appropriately. Big
ol' verbose message in case something wonks out though, but it should be
pretty evident in the logs now.
## Why It's Good For The Game

Better logging for admins.
## Changelog
:cl:
admin: Logging of Bombanana Peel explosions should be far better now.
/:cl: