Fix canisters and portable pumps from rounding the user's input when setting pressure value. (#87243)
## About The Pull Request
Fix canister and portable pump pressure inputs from transferring rounded
values instead of the full value inputs to their target release
pressure.

Does this by adding a round_value = FALSE argument to
tgui_input_number() in canister.dm, and removing the rounding operation
in canister.dm and pump.dm.

The UI should still round the values to display, just not rounding the
value to transfer to the canister.

### Known issues:
If the input is set to an integer with the same value as the rounded
current value, then it will consider it the same for some reason, thus
not calling ui_act(). Not sure why this happens.
## Why It's Good For The Game
Closes #87175.
## Changelog
:cl:
fix: Fixes canisters and portable pumps from not considering decimals
when transferring input to setting release pressure.
/:cl: