adds big pointer to chipped quirk (#87486)
## About The Pull Request

adds the big pointer chip that lets you recolor your large pointer to
the chipped quirk

## Why It's Good For The Game

i forgot to add this, whoops

if you are worried about spam. the 2 cost chip that literally makes
every message you say be spoken like a musical has really shocked me in
how little it has been used so far. so that is not a concern

## Changelog

:cl:
add: adds big pointer to chipped quirk
/:cl: