Add time dilation to hub status. (#87499)
If tg is gonna get new hardware might as well use it to flex on the hub.

This takes the middle number from the 3 avgs shown in the status tab and
displays it on the hub next to the current round time.

The hub status field is only updated whenever somebody joins or leaves
the server so the real-time number would be less useful.