Gives nukies beakers (#88508)
## About The Pull Request

This gives the chemical section in the nukie base some large beakers

## Why It's Good For The Game

It's a bit mad that a nukie can pay TC to get access to this area and
then they get, like, 3 small beakers? I've had to intervene as an admin
to give them more.

## Changelog
:cl:
map: In the new year's budget the syndicate have decided that chemists
need beakers to do their job properly.
/:cl:

Co-authored-by: Josh Powell <josh.powell@softwire.com>