IceBoxStation.dmm - Added Fitness equipment to fitness room (#88588)
## About The Pull Request

When playing on Icebox station with another player, we went to the
fitness room to workout to only realize they don't have any actual
fitness equipment in the room. Without totally tearing apart the layout,
I added two lifting benches so that people can still get their muscles
swole and still give the room the same ease of flow for people to move
through it.

## Why It's Good For The Game

This adds equipment to the map that needs to be there so that players
can build up their fitness levels besides using the boxing ring and
gives players more variety/flavor for roleplay.

![Screenshot 2024-12-18
125040](https://github.com/user-attachments/assets/17345960-ee53-43f1-b86d-86ccd287ead6)

## Changelog

:cl: Sparex

map: Added lifting benches/workout benches to the fitness room for
IceboxStation.dmm

/:cl: