Restore Spear's Ability to Destroy Lockers (#88377)
## About The Pull Request
This change make the lockers "electronics" weaker to the spear,
restoring its ability to destroy lockers. While this functionality was
not originally intended (throwing spears on lockers) ( #88141 ), it has
been present for so long that it has effectively become a feature. This
pull request rework that feature by making the spear do damage by
hitting lockers in melee.

https://github.com/user-attachments/assets/67058342-57e2-4670-9c68-df4c3b6d6193

## Why It's Good For The Game
Restoring the spear's ability to destroy lockers maintains its previous
utility and aligns with player expectations, preserving and reworking
feature that has become a staple of gameplay and lower the burden on
antagonist to open security or command lockers.

## Changelog
:cl: UnokiAs
add: Make spears able to break open lockers in melee.
/:cl:
# Conflicts:
#	strings/tips.txt