Adds goodie purchasable pet mice and code update to pet carriers. (#89134)
## About The Pull Request

Nothing super crazy here, I've added a new goodie item (For **75
credits** as a private purchase), a smaller variant of the pet carrier
that is pre-loaded with a single mouse already. So that should be good
to g-

Okay so, it turns out that pet carriers have some open issues, such as
that pet carriers at base were made to be GAGS, but subtypes of pet
carriers as a result can generate empty icon_states due to not all
following the new GAGS template.

And as a result the exodrone loot carrier was also broken, so we added
some new states to make that work.

And then I updated a bunch of comments within the pet carrier file
because they needed auto-docs.

And then I added contextual screentips, because we're living in 2025
now, and I love me some useful screentips since I had no gosh darn idea
how to use pet carriers in the first place!

![image](https://github.com/user-attachments/assets/cd5f2e56-6a0c-47e1-a1f7-1114fd783559)

## Why It's Good For The Game

Overall, a nice code cleanup for pet carriers. Screentips aid
readability for use in-game, comments improve maintainability, and we
resolved a niche bug on a reward item.

As for why adding mice to the goodies menu is useful, there are multiple
ways to use mice in-game, including breeding more mice, regal rat
farming, cytology, cooking ingredients, etc. At a certain population,
it's very easy to actually run *out* of mice through chewing on wires or
similar, and this is a simple addition to prevent that. Plus, economy
content.

## Changelog

:cl:
add: You can now purchase pet mice as a cargo goodie.
qol: Pet carriers now have contextual screentips.
fix: The biopod pet carrier now doesn't randomly turn invisible.
/:cl: