Fix premium & contraband "hidden" discounts (#89189)
## About The Pull Request
Fixes #89186
Fixes probably an old bug. Discount premium and contraband products in
vending machines, was only controlled on the UI side the whole time.
That is, a person with enough money to make the buy button active could
buy a premium/contraband product at a discount. I'm sure it wasn't
intended that way.

I'm not sure if this is the best fix, but it's the single cheapest and
easiest one I've found
All because premium and contraband goods, exist only on the interface
side, and in the corresponding list of any vendor

## Why It's Good For The Game
Bugs bad

## Changelog

:cl:
fix: Premium and contraband products in vending machines, now really has
no discount
/:cl: