Fixes another fishy CI runtime (#89248)
## About The Pull Request

![image](https://github.com/user-attachments/assets/07e97f86-fc15-4d0b-8b36-ffa100525efc)

Fish trophies from persistence are meant to be dusted and thus drop no
fillet. So we always need to be checking that we actually have a
`fillet_type` before attempting to create one. Just fixes one spot that
wasn't doing that.

## Why It's Good For The Game

Annoying CI failures for no reason begone.

## Changelog

:cl:
fix: due to a number of complaints of fish explosions, chrystarfish will
no longer be chosen as the random starting fish for the fish mount.
/:cl: