Supply Consoles restyle & order content preview (#89198)
## About The Pull Request
In a nutshell, everything will be visible in the `Demo` section.
But if you want some details...

- Filling of static_data was divided into procs, for better
understanding and possibility to use the same actions in two order
consoles
- Added an image of the first item in crate at each order (almost)
- Added the ability to see what is inside an order, the information is
not always accurate due to the nature of some orders such as random hats
- Slightly remaked layout, making everything look more cohesive and
eliminating the empty space at the bottom of the screen
- When buying anonymously, the original price does not disappear, but is
crossed out, and the new price is written below it

## Demo
<details><summary> Screenshots </summary>

Supply Console
| Before | After |
| - | - |
|
![image](https://github.com/user-attachments/assets/6067fc76-de5a-4d32-929a-239074514767)
|
![image](https://github.com/user-attachments/assets/81b3ab2a-f6d6-4f7a-8fff-0d5f459bcef9)
|

Express Console
| Before | After |
| - | - |
|
![image](https://github.com/user-attachments/assets/934554e5-ec42-422f-b6ce-f66a137bd280)
|
![image](https://github.com/user-attachments/assets/0d00c364-eece-4aa4-8f7d-f40ab7b9bb27)
|

Other
| Content preview | Express login |
| - | - |
|
![image](https://github.com/user-attachments/assets/62b3374e-5f46-4113-9da9-f481aa25739c)
|
![image](https://github.com/user-attachments/assets/8468f191-a5da-4de4-86f1-e0aca10256b3)
|

</details>

<details><summary> Video </summary>

https://github.com/user-attachments/assets/c7e16206-39f5-413c-9eea-c0284350c4a3

</details>

## Why It's Good For The Game
More convenience plus a bit prettier.
The ability to view the contents of an order, even if some madman didn't
describe its contents in the description (those who did this are really
crazy)
Section with tabs scrolls on both 515 and 516
And little cleaner data receiving code

## Changelog

:cl:
qol: Supply console got little redesign, and ability to see what can be
inside a order.
code: Supply console and express supply console, now use a unified
method of acquiring data
/:cl:

---------

Co-authored-by: Ivory <distributivgesetz93@gmail.com>
Co-authored-by: Jeremiah <42397676+jlsnow301@users.noreply.github.com>