[MIRROR] Type 213 Submachine gun (#5391)
* Type 213 Submachine gun (#4860)

* Adds a new pistol, the Zashch 10mm

Adds the 'Zashch' 10mm pistol to the game, a slow firing pistol with a time to crit on a human of 4 seconds, but boasts 18 rounds in a magazine, giving it the highest damage-per-reload of any handgun crew acquirable in the game.

It is sold by NRI surplus for the same price as the Bobr revolver. It is chambered for 10mm to give life to a dead calibre and to give a crew-gun that fires the bullet, to give the calibre a chance to have a life outside of being solely an evil syndicate calibre that is for some reason, orderable from cargo.

Magazines, in-hand sprites, and mob sprites are all tested and functioning correctly, as if the flashlight function.

* Update modular_nova/modules/modular_weapons/code/company_and_or_faction_based/szot_dynamica/pistol.dm

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>

* Update modular_nova/modules/modular_weapons/code/company_and_or_faction_based/szot_dynamica/pistol.dm

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>

* Update modular_nova/modules/modular_weapons/code/company_and_or_faction_based/szot_dynamica/magazines.dm

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>

* Update modular_nova/modules/modular_weapons/code/company_and_or_faction_based/szot_dynamica/pistol.dm

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>

* Update modular_nova/modules/modular_weapons/code/company_and_or_faction_based/szot_dynamica/pistol.dm

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>

* Adds the Type 213 Submachine Gun

Adds a new sub machine gun to the game, a companion piece to the Type 207 kinetic pistol.  The weapon uses the same ammo, different magazines that hold 24 rounds, and fires in 3 round bursts with high spread. Made to make the kinetic ball ammo more viable. Downs in 6 hits to the chest, which is two full bursts.

* Update to the sprite of the Type 213

As per Jungle's suggestion, changes the finish of the type 213's base sprite to a more darker blued finish, contrasting better with the red-ish furniture, white barrel and bolt, and the green accenting stripe.

* I forgot the magazines.

I forgot them.

* Update modular_nova/modules/modular_weapons/code/company_and_or_faction_based/bolt_fabrications/longarm.dm

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>

* [MIRROR] Type 213 Submachine gun

---------

Co-authored-by: Trelus <140990718+Trelus@users.noreply.github.com>
Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>
Co-authored-by: StealsThePRs <StealsThe@Email.com>