Fix illiterate people writing messages with crayons and spraycans (#89738)
## About The Pull Request
Right now people can use crayons and spraycans to write messages on the
floor to communicate while being illiterate. While this is kinda funny,
it does defeat the purpose of this quirk which is to disable nonverbal
communication methods.

## Why It's Good For The Game
Consistency.

## Changelog
:cl:
fix: Fix illiterate people writing messages with crayons and spraycans.
/:cl: