Fixes research hanging the MC (#89855)
## About The Pull Request

So design disks and the techweb have this neat little thing called
```copy_research_to``` which chains three **highly expensive** procs
together which only sleep during CHECK_TICK. So it piles all the bodies
up in the corner and the Master Controller explodes trying to process a
fuck load of procs and the speed of my autoclicker

The entire research chain probally needs to be refactored to be less
expensive, but for now a 5 second cooldown will hopefully give it plenty
of time. You only need to click these buttons once. So the cooldown can
easily be increased without much gameplay effect.


![image](https://github.com/user-attachments/assets/5e0c2b5a-0fe5-4afc-b0bd-afa19e528ce6)

I didn't lag out on my local server using the debug disk so I think this
is a fix.

fixes: #89827
## Why It's Good For The Game

I'm sorry mr admin i crashed the server

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: Design disks now have a cooldown when uploading to no longer crash
the server.
/:cl:

---------

Co-authored-by: Lucy <lucy@absolucy.moe>