Return stereo for music/ambient (#89470)
## About The Pull Request
Returns stereo sound for music and ambient, which was changed here =>
https://github.com/tgstation/tgstation/pull/55333
Due to the override environment (probably) absolutely ALL music was
downmixed to mono, such as music via global sound, direct sound to mob
and megafauna themes.
In addition to this, the sound from musical instruments at certain
positioning was sent exclusively to one ear, which made it extremely
uncomfortable to listen to with headphones.

Commenting of 183-189 lines was done to make proc local_sound work
correctly, otherwise an error occurs.

Video from the tests is attached, old - what we have at the moment; new
- what we get when commenting on the lines.
https://drive.google.com/file/d/1wSZ2I5XE74nePOTfMmaj-xqmonf3g58O/view
(too much weight, had to upload it on google-drive, I apologize for
that.)

## Why It's Good For The Game
Fixes the bug? Based on the PR with override environment, downmixing
music in mono was not planned.
Makes working with music and listening to it more comfortable

## Changelog
:cl:
fix: remove downmix stereo=>mono where it is not required
/:cl:
# Conflicts:
#	code/game/sound.dm