Moves the crew pinpointer from the premium vendor section to the regular section. (#90062)
## About The Pull Request
Moves the crew pinpointer from the premium vendor tab to the regular tab
while adjusting its price slightly to ensure it is still relatively
expensive for non medical crew. Currently with this change it costs 60cr
for medical staff and 300cr for everyone else. Ideally I wanted it to be
a bit more expensive for medical staff (more around 100-150cr) but
vendors are a bit rigid in how their pricing seems to work.
## Why It's Good For The Game
Lifeline on the PDA/Tablet/Whatever they're called is more or less
objectively superior to these old things in all situations but one, when
the comms go out. The problem is that these being 300cr even for medical
staff is quite steep and discourages people from buying them at all.
With this change I hope to encourage use of these as a more reliable but
less efficient backup tool in the event that shit has hit the fan and
comms are likely to be out for an extended period of time.
## Changelog
:cl:
balance: Crew pinpointers have been moved to the regular section of the
medivend, and as such now benefit from the department discount.
/:cl: