Fixes aquarium fish overlays (#90009)
## About The Pull Request
I don't know if it's 516 or some of the latter updates 515 (EDIT:
definitely 516), but fish visual overlays are now rendered under other
overlays composing the aquarium when past a certain pixel offset, since
they are using the pixel x/y vars to do it. All we have to do is to
change it to pixel w/z

## Why It's Good For The Game
Small layering issue.

## Changelog

:cl:
fix: fixed a small rendering issue with fishes inside an aquarium.
/:cl: