Makes fishing rods with higher (or lower) bait speed a smidge easier to control in the minigame (#90011)
## About The Pull Request
While the bait overlaps the fish during the minigame, the change in
velocity will be further cut down or amplified for fishing rods that
have higher or lower speed multiplier than the identity (1).

## Why It's Good For The Game
Fishing rods that are supposed to be better are also made slighly more
annoying or finnicky by their speed multiplier.

## Changelog

:cl:
balance: Makes fishing rods with higher (or lower) bait speed a smidge
easier to control during the minigame
/:cl: