Simplifies fish traits inheritability a little (#90014)
## About The Pull Request
I'm thinking of removing some of the redundant or unnecessary bad design
from fishing, especially aquariums, which are definitely a bit fringe
and difficult to attend to on top of other stuff. However I don't have
all time of the world so I'd rather start small, even if it isn't the
main obstacle (which is keeping fish alive and getting them to reproduce
in the first place).

Thinking about it, the fact that some traits have a chance not to be
passed down to the offspring even if both parents have them is plain bad
and it's led situations where I couldn' tell if it was a bug or a
feature. The solution? Forgo the probability check for traits shared by
both parents.

## Why It's Good For The Game
Clearly that was bad design. 

## Changelog

:cl:
balance: Made fish traits inheritability easier. They're now guaranteed
to be passed down to offsprings if both parents have them.
/:cl: