Small autoreel (and cloaked?) fishing line(s) buff (#90032)
## About The Pull Request
The autoreel fishing line now nudges the bait up/down lightly towards
the fish position at 20% of the reeling velocity when idle, which is
most likely not enough to defeat gravity but enough to help with control
unless for maybe a couple of ultra-light material for fishing rods like
snow. May or may not trivialize the minigame a bit when tied with a
gyroscopic hook. We'll see if that's even a problem.

The cloaked fishing line is actually way more transparent now, and a
smidge better at catching fish with the anxiety trait (only the zappy
fish so far...) and not only those with the wary trait.

## Why It's Good For The Game
I feel disappointment knowing I had it called "autoreel" yet it doesn't
do that muh during the actual minigame, only affecting the portion
before it. Also cloaked lines aren't that hard to notice rn smh.

## Changelog

:cl:
balance: Autoreel fishing lines do what their name implies and nudges
the bait towards the fish a little during the minigame when you're not
pressing the mouse button. Also cloaked lines are more transparent, and
better at catching a certain fish...
/:cl:

---------

Co-authored-by: SmArtKar <44720187+SmArtKar@users.noreply.github.com>