Add rare chance of thunder for ash storms on lavaland (#90488)
## About The Pull Request
This adds `THUNDER_CHANCE_VERY_RARE` for a thunder strike to occur on
lavaland during ash storms. The chance is 1 in 50,000, which is rare
enough for it to occur a few times during an ash storm. The emberfalls
storm does not have thunder strikes applied. I also went ahead and
colored the thunder red, to go with the red color of the mining turf
colors as requested by:


![Discord_zBq1WlovrU](https://github.com/user-attachments/assets/11ea9b16-4cea-468d-adae-0607da88edd2)


## Why It's Good For The Game
Fun little rare event to make ash storms cool. 

You might wonder why we don't increase the chance? Since lavaland is
really only used by a few players during a round, it is a good idea to
minimize the amount of turf processing that is done. By making it very
rare, the calculations average about ~`1.18` turfs being processed per
tick for the entire z-level.


![Code_HdVqgjWtnN](https://github.com/user-attachments/assets/ec576348-b85f-4db5-97eb-4c6bbb5de118)


## Changelog
:cl:
add: Add rare chance of thunder strikes during ash storms on lavaland
/:cl: