[Mirror] Adds table flipping! Take 2.0 (#90156 ) (#5213)
## ORIGINAL PR: https://github.com/tgstation/tgstation/pull/90156/

This breaks this horrendously for us so it's going to its own quarantine
branch. It certainly succeeded in 'adding table flipping' in that they
have made our maintainers want to do that IRL.

## About This Pull Request

Adds table flipping! Take 2.0 (#90156)
Revives the long dead #80348. Right click a table to flip it over,
Useful for makeshift cover or ragequitting. Their integrity is equal to
that of the table you flipped. Some tables aren't able to be flipped
(reinforced tables, roller tables, etc).

I refactored it to be less snowflakey. so hopefully this passes with
only minor changes


![290775902-48414bb3-aaa9-467d-8edf-a170a98b1506](https://github.com/user-attachments/assets/3627c537-043e-4829-b38b-f68a7d382167)


![gambar](https://github.com/user-attachments/assets/ce04a6d1-e480-46f9-8913-041f86db8a9a)
I felt like we're lacking in the makeshift defense items. This is a
quicky way to plop up a shitty barricade for you to guard with. and also
really cool for roleplaying and such. also, this categorically goes
hard:


![gambar](https://github.com/user-attachments/assets/af593068-d9f3-49b0-9102-989ce2b4d3fb)
:cl:
add: You can now flip tables by right clicking them!
/:cl: