tgui: Shiny new dev tools (#90522)
Old project that I have been waiting on for some time. We can swap out
our javascript bundler with a rust version. We still have webpack to
some extent because sass-loader requests it, but it's unused.

This was originally suggested by stylemistake years ago but it depended
on tgui using a modern browser #90397 and rspack supporting yarn pnp
https://github.com/web-infra-dev/rspack/pull/7639.

I want to also replace jest with vitest for this PR since it's deep
internals and no one else is really tinkering here. Drafting for now
Both of these tools are the drop in replacement of the modern decade
Makes the dev server borderline instant

https://rspack.dev/misc/faq#what-are-the-advantages-of-rspack-compared-to-webpack--swc-loader