Fix hot spring runtime related to duplicate elements (#90555)
If something is mapped to spawn in on a hot spring tile, the tile will
add the immerse element in `/turf/open/water/Entered` before
`Initialize` runs, and when it does there's no check if the element has
already been added, unlike other spots where the immerse element is
added. This fixes that by adding a check on the call in `Initialize`.
# Conflicts:
#	code/game/turfs/open/water.dm