Detective Board UI update (#90591)
## About The Pull Request

I've updated the UI of the detective board to make it look more
detective-like. Also, it is not necessary to specify the name and
description of the evidence to be attached. If it is not specified,
these fields are taken from the item.



https://github.com/user-attachments/assets/51419478-5fa9-43d4-a903-992ab4f0c91d

(The detective smoked a cigar and is now coughing)

Before:

![old](https://github.com/user-attachments/assets/b900b488-8271-40df-ac7c-18b60c8f6bf5)
After:

![new](https://github.com/user-attachments/assets/6c9d19e7-d94b-4003-81b1-54d781fb62bc)

## Why It's Good For The Game

Better UI/UX

## Changelog

:cl: FeudeyTF
qol: Better UI/UX for detective's board
/:cl: