webmap is now in maps.txt instead of a config entry (#90821)
## About The Pull Request

Currently the button to open a map in webmap is tied to a config entry,
which would then take the name of the map being played and get the link
to that, however for cases like new/tested/admin uploaded maps where
there is no webmap available, this would still have a link to a url that
doesn't exist.
Instead, we'll make the webmap link be in ``maps.txt``, same as
feedbacklink.

## Why It's Good For The Game

Explained in the about section, this prevents webmaps appearing for maps
that aren't supposed to have a webmap link available.

## Changelog

Nothing player-facing, webmaps aren't currently used (afaik) currently
cause they're broken for tg codebases.