Adds the exosuit ablative insulation module. A module that protects a mech from energy attacks and their wires from EMPs. But makes it weaker to ballistics and melee weapons. Adds a special version for nukie mechs. (#90845)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds the Ablative Insulation. A type of mech armor.

The armor module protects the mech from energy weapons (such as thermal
pistols), and also protects the wires from EMPs (only the wires, not the
mech itself). However, it makes it weaker against melee and ballistic
weapons.

The Dark Gygax and Mauler both obtain a version of this module as part
of their default loadout that does not have any disadvantages against
melee and ballistic weapons.

## Why It's Good For The Game

This is a complimentary PR to this one.
https://github.com/tgstation/tgstation/pull/90830

Wire scrambling for mechs is an extremely irritable system, but it does
force mech pilots out of the mech to repair some of the effects
attached. or at least to a standstill.

This can often feel a bit like a knowledge check to resolve for some
pilots, since the optimal solutions for this typically involve pulsing
the wires in some way, utilizing assemblies like voice analyzers,
circuits or signalers.

While I would like a more obvious way for pilots to resolve these
problems without having to set up signal networks, I do think it should
come at some expense in other defenses so that it isn't such a catchall
solution compared to the other armor booster modules. There would
otherwise be no reason not to use this armor booster if there was no
downside, since it completely negates a weakness. Therefore, we're
aiming for a slight upside/downside exchange. You get a very good
upside, but you now need to compensate for weaker defenses elsewhere
(such as risking internal damage effects due to lower armor, see this PR
https://github.com/tgstation/tgstation/pull/90844)

We've given nuclear operatives an improved version because their mechs
are A) expensive as shit, and B) meant to be a huge all-in but have
suffered immensely from many years of refactors, rebalances and gear
availability changes that definitely have left them underperforming for
their promised value. I would much rather nuclear operative mechs feel
like they are less beholden to the weaknesses that crew mechs need to
contend with if only because they exist for what might be a fraction of
the time, if they are even spawned at all due to their price.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Adds the Exosuit Ablative Insulation armor booster module. Protects
the mech from energy-based attacks and its wires from EMPs (only the
wires, not the mech). However, the armor booster makes the mech weaker
to ballistic and melee weapons.
add: The Dark Gygax and Mauler now come pre-packaged with a special
version that has no weakness to ballistic and melee weapons.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->