optional limited mentoring books (#5449)
## About The Pull Request
adds a new variable to mentoring books that allow people to create
subtypes that will be destroyed after a certain amount of uses
## How This Contributes To The Nova Sector Roleplay Experience
this allows other contributors, both code and mapping, to make limited
mentoring books for whatever they want (such as ruin rewards, or
whatever). More flexible content allows for more variety and less of the
same gameplay loop.
## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>
  

![image](https://github.com/user-attachments/assets/257b8098-fd86-4119-a7a9-d00052710524)

</details>

## Changelog
:cl:
code: added a limited version of the mentoring book, so other
contributors (and stuff in game) can modify it for whatever reason
code: removed the examine signal and just used the default examine
proc... I should have done that before
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>