re-nerfs the SKS (#5481)
## About The Pull Request
puts the SKS back at 0.5x damage.

## How This Contributes To The Nova Sector Roleplay Experience

the SKS is INCREDIBLY strong with 1x damage, it is genuinely nukie-level
and thats with BASIC AMMO, if i used AP in the testing it would probably
be even worse, this puts it atleast better inline with other weaponry,
yeah i know that the sks is <2 spawns per round usually, but that doesnt
give it a reason to be an incredibly OP gun thats still extremely easy
to get, just uncommon. and even with 0.5x damage, AP is still really
strong, about on-par with 1x damage mult, though hopefully because the
ammo lathe changes AP isnt used often

## Proof of Testing

### 1x SKS damage (no ap rounds): 


https://github.com/user-attachments/assets/7f710947-7f77-49ab-9bc9-8df01e216a15

### 0.5x SKS damage (normal and then AP rounds): 


https://github.com/user-attachments/assets/0b1c1f8d-addd-428c-bc32-b997d85c2bb5






## Changelog
:cl:
balance: put the SKS back to 0.5x damage mult, because it turns out
nanotrasen actually hates when crew can buy a strong gun off an illegal
market.
/:cl: