Summon Cheese sometimes has a different invocation (#91053)
## About The Pull Request

When created, each instance of Summon Cheese has a 50% chance of its
invocation being changed to use the actual ref string for
`/obj/item/food/cheese/wheel`, instead of Skyrim's item id for cheese
wheels.

## Why It's Good For The Game

The invocation for Summon Cheese is itself a reference to a debug
feature from another video game. Why not add an extra layer of
meta-humor by occasionally using the actual ref string for the summoned
object's typepath in the invocation?

## Changelog

:cl:
spellcheck: The invocation for Summon Cheese is sometimes more
representative of the realities of life aboard the station.
/:cl: