Fix cable coil icon breaking with layer hubs (#91151)
## About The Pull Request

The icons for the multilayer and multi z layer cable hubs are in a
separate dmi than the cable coil, so when you tried to switch to either
of them by using cable coil in hand it'd just go invisible. This makes
sure the icon gets changed to the right one if you select those (and
gets switched back when you choose cable coil). It also moves the
setting of `novariants` to be above the `set_cable_color` call because
it's checked in that call and so it needs to be the updated value. This
also adds a check for `novariants` in `update_name` and `update_desc`
because currently the changed name and desc for the cable hubs gets
overwritten by these calls (triggered by `update_appearance` at the end
of `attack_self`)

## Why It's Good For The Game

Bug fix

## Changelog
:cl:
fix: fixed cable coil going invisible when you select the two cable hub
options
/:cl: