Renames Hospital Gown to Surgical Gown, adds 'surgical' examine tag (#5536)
## About The Pull Request
Title. Nothing more to it.

... except for repathing but that's optional. Just changes it all to
surgical_gown (and consistently snake_case)
## How This Contributes To The Nova Sector Roleplay Experience
I saw that the surgical mask has this little examine-tag and figure it's
the perfect thing to copy onto OUR surgical item!
Better for users to know what this item is and does, which it doesn't
explain very well as-is. The rename also hopefully will help with that.
## Proof of Testing

![image](https://github.com/user-attachments/assets/76413164-66e6-4e02-b90e-62474e1f0a18)

## Changelog
:cl:
qol: renamed the hospital gown to surgical gown, and added an
examine-tag: as a 'Surgical' item, it does not block surgery on covered
bodyparts!
/:cl: