Immobile shuttles will not appear launchable on flight control consoles. (#91275)
## About The Pull Request

Custom shuttles introduced a weird edge case in shuttle code where the
flight control console shows the shuttle as able to be launched, despite
a lack of engines making it unable to move. This fixes that edge case by
specifically showing the shuttle as immobile in the console.

## Why It's Good For The Game

I tried to launch a custom shuttle but I couldn't figure out why it
wasn't moving. This fix at least lets players know that it won't launch
when it can't.

## Changelog

:cl:
fix: Shuttle flight control computers will show you when the linked
shuttle cannot move.
/:cl: