Catalyst: The Catalyzing (Implementing the unused catalyst types) (#90818)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds the unused catalyst types into the game. These are:
pH catalyst, now called Ionic Modulator, which reduces the change in pH
of a solution as it reacts. If the purity is too low it can even flip
the other way and start making it change in the opposite direction.

Temperature catalyst, now called Thermic Modulator, which reduces the
change in temperature as a solution reacts. If you're too impure with
this one it can actually make an exothermic reaction endothermic and
visa versa.

Note: I think the purity reversal behavior is weird but I didn't write
the code for the catalyst effects themselves, so I'm keeping it as is
unless asked to change it.

Adds a new competitive reaction, similar to Tempomyocin formation.
(Don't tell anyone but I actually just copied and altered tempomyocin
formation to make this one.) Said reaction is between Catalysium T, and
Catalysium P, which react into thermic and ionic modulator respectively.
(T is for Temperature and P is for pH.) P is the harder one here, given
that pH management is usually a bigger problem than temperature
management, but if you're having trouble making P, you can add gold to
the competitive reaction to speed up P formation, same as how you add
bluespace dust to Prefactor A/B to make Tempomyocin faster.


TESTING IMAGES:

![image](https://github.com/user-attachments/assets/ee5363f8-55f7-4716-a925-4c38aef104e4)
50u sodium 50u chlorine without ionic modulator

50u sodium 50u chlorine with 10u 100% pure ionic modulator

![image](https://github.com/user-attachments/assets/4d4dcf3e-ab8e-4396-9f6c-d06701eda0b0)


competition functioning:

![image](https://github.com/user-attachments/assets/50832df5-ff40-467e-bded-49a6a7425107)
(not shown: the 20u gold i put in since it got consumed to boost P
formation)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

1. The code for the catalysts was already there and lying unused for a
long time, despite being perfectly servicable.
2. It's a way for less experienced players to slightly reduce the
difficulty of chemistry, at the expense of time and beaker space.
3. It will create new and interesting gameplay for more experienced
chemists, maybe reducing the one-and-done reliance on copy-pasted
chem-factories that (at least to me) seems pretty common among even
highly skilled players.
4. We get to actually use the competitive reaction system for something.
5. (Someone told me that palladium synthate catalyst was lame and I made
this out of curiosity and/or spite)


<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: New chemistry catalysts! Combine phenol, unstable mutagen, and
plasma in a 1:3:1 competitive reaction, similar to tempomyocin and
purity tester. Adding stable plasma to either of the two precursors will
create the new catalysts, one regulating pH, and the other temperature.
If you're having trouble getting the one you want, you can optionally
add gold to the beaker as the competitive reaction progresses, which
will cause the pH precursor to form more quickly.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->