[Modular] Adds a service guard locker to the code. (#5566)
## About The Pull Request

This adds a locker, as well as a garment bag, for the service guard to
the code, although not to any maps, at least not yet.

It currently uses the cabinet sprite, as i wasn't able to find anything
more fitting, although i'm not against making a new sprite for it.

While i wasn't seemingly able to find any previous attempts to add
something like this, i do apologize if there was, and it was rejected
for whatever reason.

## How This Contributes To The Nova Sector Roleplay Experience

As of now, the service guard is the only departmental guard which
doesn't have a locker of their own, which i feel like is a glaring
omission. This at least comes closer to bringing them to par with the
other departmental guards.

Even though this pull request doesn't actually add these lockers to the
maps, i still feel it's beneficial, because it lays down some groundwork
for potentially adding in more service guard outposts in the future,
such as through using automap templates on some other area (such as the
arrivals security outpost, or potentially the theater), or adding them
directly to Nova's maps. (As far as i know, Snowglobe is the only map
that currently has a service guard outpost.)

## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>
  

![Screenshot_298](https://github.com/user-attachments/assets/708d1f76-0faa-4903-bc23-a49efbb096cf)

![Screenshot_299](https://github.com/user-attachments/assets/624fe448-7e14-4ec5-90e6-64f405ff67a5)

![Screenshot_300](https://github.com/user-attachments/assets/19ec61d0-65e6-4a51-bd51-5a08b84599ec)

![Screenshot_301](https://github.com/user-attachments/assets/e09e4634-46b9-4f37-bffd-43bc3490bd1c)

![Screenshot_302](https://github.com/user-attachments/assets/df2d19fa-23a4-45c4-84ed-98ed24e36a19)

</details>

## Changelog
I don't know if i should add a changelog, since there isn't technically
anything player-facing in here.

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>