Resprites the 'Dagger' Designer Mantle (#5569)
~~Hello entirely unrelated PR to my projects!
I'm treating myself. Leave me alone.~~

I was using this item ingame and realized that the side sprites are
super inconsistent in color. While I fixed that I ended up touching the
whole thing up.

Should be _relatively_ unchanged barring a few intentional choices:
- The shoulders and back-of-neckline are lower and smoother.
_Them being so far our appeared really odd whenever you wore anything on
your neck. Suit-slot items should typically stick closer to the body's
outline..._
_Now that it won't clip it can safely be worn UNDER a jacket (w/ jacket
using the suit-neckslot-feature) to have a lot more variety in its use._
- The side-sprites are almost entirely redone.
_They should 'spin' better._

I'll be 100 with you. I was doing this with sheer tops in mind, so my
design vision was definitely influenced by that.
It still works as a mantle! But now (with loadout rename/redesc) it has
a much more varied use.

~~Sheer tops will save us.~~
Improves a modular sprite that had some inconsistencies and
peculiarities. Now it should be smoother to use and allow a touch more
variety.

It's really not that much different. It just holds more color and the
shoulders don't jab out like angry spikes now.

![image](https://github.com/user-attachments/assets/7f1b33cc-5362-4144-9eaf-1fba531b7d4d)

:cl:
image: resprited the 'Dagger' Designer Mantle
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>