Refactors mech interactions to modern attack chain handling (#91475)
## About The Pull Request

What it says on the tin. Attacks against the mech are handled as
attacks, and anything else is handled as item interactions and return
using those relevant values.

## Why It's Good For The Game

These procs are a bit of a mess and it did in fact hinder my last pr
related to mechs, so this should make things a bit easier for
development upon going forward.

## Changelog

nothing player facing