[Staff Request] [Balance] Mining Blood Jaunt rework, Mining Crushers adjustments, SEVA exploit fixed, Shotgun PKA revised, Cooldown PKA limited (#5474)
## About The Pull Request
Adds the SEVA suit to the list of things you cannot use in the neck.

Makes the Bubblegum Blood Potion grant a Mining variant of the Blood
Jaunt.

Expanded base blood jaunt to allow for a white and black list of areas
where it can be used. It cannot be used and will expel the user when not
in a whitelisted area and when in a blacklisted area.

Mining Blood Jaunt can only be used in planetary areas, with interdyne
and tribal areas excluded.

Mining Blood Jaunt has an activation time of 2 seconds, from instant.

Shotgun damage modkit nerf was reverted.

Shotgun was made heavy weight to be used in two hands rather than one,
to prevent the PB cheese with two hands.

Code support to limit modkits of the same type was added.

Cooldown modkits were limited to 3 per weapon max, as its the max
intended for upgraded pka's in tg code.

Mob AoE modkits were limited to 1 per weapon max, temporarily at least
until tg decides a proper max for them, as otherwise they can one shot
mega fauna or other difficult mobs.

Kheiral Cuffs were added to the voucher webbing option.

## How This Contributes To The Nova Sector Roleplay Experience
SEVA: It was being abused to put the seva in the neck and use an armor
underneath to be inmune to the storm. bugs abuse is bad.

Following  were Headmin request:

- Blood Jaunt: A spell with 0 cooldown that lets you jaunt anywhere,
that is, turning invisible, inmortal, untargetable, all seeing, faster
than a ghost, practically a camera less, body less super AI unfettered
by the laws of the living, to hear and see anything in the station at
any time, traveling with all the gear you have that is not in your
hands. I am surprised we are even letting it remain in the game with
such level of abuse.

<details>


![image](https://github.com/user-attachments/assets/5c6a9d15-42b7-46fb-944e-50df5ea125bd)

</details>


Requests by players:

- Shotgun nerf to damage kit was removed. Shotgun was made two handed
instead.

- Cooldown PKA modkits were limited to a max of 3. Abuse was reported by
using more than this amount.

- Kheiral Cuffs be added to the Webbings options, as its the weakest
option of the bunch and its useful to have such set early for newbie
miners.

Exploit found during these fixes:

- Using 10 aoe modkits allowed for a pka that did 100 damage on station,
400 in lavaland.

## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>
  
Seva: It just works.

Blood Crawl:

Snow Inside:

![image](https://github.com/user-attachments/assets/167140ed-0f7f-414d-b284-27b8a9c01578)
Snow Outside:

![image](https://github.com/user-attachments/assets/63e65609-445d-42a5-be20-becd14a94f54)
Snow Interdyne:

![image](https://github.com/user-attachments/assets/042f908c-34a2-49cc-bac6-00f0dfd9b37c)
Snow Heartkin:

![image](https://github.com/user-attachments/assets/4bcece60-ad8e-4f96-880d-0286d989aa83)


Lava Station: (Ignore the old text)

![image](https://github.com/user-attachments/assets/1512ef01-5632-414c-a55c-424b892d5483)
Lava Mining:

![image](https://github.com/user-attachments/assets/e5150990-1b54-4e33-be28-6b6ce89846bf)
Lava Shuttle: (Ignore the old text)

![image](https://github.com/user-attachments/assets/8532fa26-000c-4c6c-b3b5-23db9abca280)
Lava Mines:

![image](https://github.com/user-attachments/assets/06867c96-b696-43e7-bbc3-db6fdfb5d4e6)
Lava Interdyne:

![image](https://github.com/user-attachments/assets/f52ffea3-78dc-4f34-a957-a520f4a25a7f)
Lava Ashwalkers:

![image](https://github.com/user-attachments/assets/6514eac0-e23e-48eb-aefb-ba6e061a2c5f)

PKA's

Cooldown limit:


![image](https://github.com/user-attachments/assets/69f0aa9a-4f4f-4339-9b2e-c614965aa2c5)


![image](https://github.com/user-attachments/assets/44a51c9f-5941-4f8c-bdf1-e8e7d9497bf5)


![image](https://github.com/user-attachments/assets/373b9138-12c9-41ac-9898-d55fa05e42ef)

Shotgun two handed:


![image](https://github.com/user-attachments/assets/06447a05-5c59-4efc-b525-07c28f5568e4)


![image](https://github.com/user-attachments/assets/3ca314b3-7496-4b17-8691-ec866862492a)

Voucher:


![image](https://github.com/user-attachments/assets/a5f1ff02-6b9a-47cb-be22-be9db2a7dcb9)


</details>

## Changelog
:cl:
refactor: Added code to Blood Jaunt so it can use allowed and disallowed
areas to it, disallowing its casting and jaunting off when not in the
whitelisted areas or in the blacklisted ones.
balance: Bubblegums Blood Jaunt reward was changed to a Mining
subvariant. It has a 2 seconds cast time, and cannot be used in non
planetary civilized areas. This means it can only be used in planets and
ruins, but not in interdyne or tribal camps.
fix: SEVA outfit cannot be used in the neck again, as this allowed
people to be inmune to ash storms and use a different armor underneath.
balance: Shotgun Damage modkits un-nerfed, they now apply 10 damage
extra to each pellet. Shotgun needs to be used with both hands, but not
wielded.
balance: Cooldown PKA Modkits were limited to 3, to prevent cooldown
cheese.
balance: Mob AoE PKA Modkits were limited to 1, to prevent damage
cheese.
add: Kheiral Cuffs were added to the Webbing set for miners, a great
starting option in line with the rest!
/:cl:

---------

Co-authored-by: jjpark-kb <55967837+jjpark-kb@users.noreply.github.com>
Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>