Blueshield extra headwear (#5624)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Adds a Blueshield's Cap clothing item to both the Blueshield's garment
bag, and the Command clothing vendor. Supports flipping like other caps
in game. Otherwise its functionally identical to the berets.

<!-- Please add a short description of why you think these changes would
benefit the game and the roleplay atmosphere of the server. If you can't
justify it in words, it might not be worth adding. -->

<details>
<summary>Screenshots/Videos</summary>

![image](https://github.com/user-attachments/assets/cce68348-2705-4d4f-9b1e-ee4daae36e4d)

![image](https://github.com/user-attachments/assets/29e858bd-3a20-4ceb-811f-3ce213505892)

![image](https://github.com/user-attachments/assets/2c4d8752-2b35-4c02-9146-0577682730d3)

![image](https://github.com/user-attachments/assets/97ceb35e-8698-4e2c-8460-d35a36852ba4)

![image](https://github.com/user-attachments/assets/d37ba5d2-1b54-4346-a368-897deefc7691)
</details>

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Added new Blueshield cap. Stylish and flippable.
fix: fixed a couple typos in Blueshield item descriptions.
image: added sprites for the new Blueshield caps.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->