Adds an option to open a TG upstream issue from the bug report verb (#5633)
## About The Pull Request

Tin, adds a prompt where you can directly report an issue to TG if you
know it's a TG issue. It does warn to pick Nova if unsure.

We don't want people to be reporting bugs on TG if they're just from us
generally, but at the same time given the large portion of bug reports
being TG ones I think this is a useful feature to have.

## How This Contributes To The Nova Sector Roleplay Experience

Some QoL for bug reporters.

## Proof of Testing

<details>
<summary>Screenshots/Videos</summary>
  

![dreamseeker_h8mLZSz8lB](https://github.com/user-attachments/assets/fde94888-3964-4dd4-b835-26e4070c8077)

</details>

## Changelog

:cl:
qol: you can now report a bug directly to tg through the in game bug
report. please only do this if you're sure it's an upstream bug!
/:cl: