Adds an enter soulcatcher button for ghosts (#5635)
## About The Pull Request

To help acclimate people to the lack of the ghost menu, added a
temporary verb to direct people to the new ghost settings menu

Also adds a button for enter soulcatcher.

Also early mirrors my TG pr to restore the Spawners Menu and Minigames
verbs for power users.

## Proof of Testing

<details>
<summary>Screenshots/Videos</summary>
 

![dreamseeker_sysaArFAYP](https://github.com/user-attachments/assets/c51988a3-4cbf-4786-b6f8-77e0933af3d7)


![dreamseeker_RhfzDVPbgI](https://github.com/user-attachments/assets/1a151cac-a1eb-48f1-8697-2425d3b95605)
 
</details>

## Changelog

:cl:
qol: adds an 'Enter Soulcatcher' button for the ghosts in addition to
the verb. ghost tab is no more, TG removed it, you can still use the
verbs in chat though as well as the buttons on the bottom of the screen.
/:cl: