self-indulgent loadout glasses pr (#5646)
## About The Pull Request
Renames the binoclard glasses entry in loadout to `binoclard glasses`
and not `thin glasses`.
Adds a transparent variant of modern glasses as a reskin.

## How This Contributes To The Nova Sector Roleplay Experience
I like the transparent glasses.

## Changelog

:cl:
fix: Renames the binoclard glasses in loadout to be binoclard glasses
and not thin glasses.
image: Adds a transparent alternate skin of the modern glasses, which
has more transparency than the basic modern glasses. Access these via
alt-click!
/:cl:

---------

Co-authored-by: Hatterhat <Hatterhat@users.noreply.github.com>