[MIRROR] Crimes & Station Manifest logging in DB #5486  (#5582)
* Crimes are now logging in db (#90948)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Crimes are now logged in db

Logging:

![image](https://github.com/user-attachments/assets/6eee470d-863a-4f68-9836-04a0a0079aa2)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
idk
we logging citations but not crimes, and i wanted to log the crimes too
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: Rengan
server: The crimes are now logging in the database.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->
# Conflicts:
#	SQL/database_changelog.md
#	code/__DEFINES/subsystems.dm

* Automatic changelog for PR #90948 [ci skip]

* Adds logging of the station manifest to the DB (#90993)

## About The Pull Request
This adds a `manifest` table to the database, basically a databasified
version of the `manifest.txt` log.

## Why It's Good For The Game
This is less of a game feature and more of an admin & Statbus tool. We
currently do not have a good way to see who played which character in a
given round. This aims to fix that.

~~Drafted because I'm not sure on some of the typecasting for the
`special` and `latejoin` columns, I would like someone to give that a
once-over.~~ We good.
# Conflicts:
#	SQL/database_changelog.md

* Merge branch 'master' into sql

* Revert "Merge branch 'master' into sql"

This reverts commit 3e1222c09b54255fecfe960149aef95353425473.

* Fixes manifest SQL logging (#91125)

I forgor a column assignment :\

* Fix manifest sql again (#91185)

This one will work this time I swear

---------

Co-authored-by: RengaN02 <60517664+RengaN02@users.noreply.github.com>
Co-authored-by: tgstation-ci[bot] <179393467+tgstation-ci[bot]@users.noreply.github.com>
Co-authored-by: Nick <nick@nfreader.net>
Co-authored-by: FlufflesTheDog <piecopresident@gmail.com>
Co-authored-by: Kocma-san <kocmasan@gmail.com>