[Modular][Dev Suggestion] Adds Pill options (#5672)
## About The Pull Request

This PR is adding the suggested content in the dev suggestion `Add a
gummy-bear "pill" to the ChemMaster 3000` and more.

## How This Contributes To The Nova Sector Roleplay Experience

Suggestion: `Pill designs are mostly irrelevant to the recipients of
pharmaceuticals. Adding a "gummy" pill to the ChemMaster 3000 would be a
very fun inclusion for the purposes of roleplay. This would introduce
more Roleplay at the pharmacy as characters with prescriptions may
explicitly request or deny gummy pills and the Chemists may choose to
comply or not as per their own authority. Initiating more roleplay is
almost always a net positive with "isolated" roles like Chemist.`

My opinion: I think this adds a bit more to roleplay and the being a bit
more grounded in potential reality as many times with medicines some are
pilled, and some are gummies to help people with difficulties take them
down. Including I thought it would be fun to have a nova variety of
medicine pills!

## Proof of Testing

<details>
<summary>Screenshots/Videos</summary>
  

![image](https://github.com/user-attachments/assets/314a26c5-2f1c-4306-baf0-7ec7e9a4bb04)


![image](https://github.com/user-attachments/assets/69247431-2eeb-4aa2-8049-12ecf2aa9af2)

</details>

## Changelog

:cl:
image: Added new "gummy" sprites to the chem master 3000 (fish, slugcat,
bear)
/:cl: