[NO GBP] Fix: Support for a light theme for standard radio channels (#91907)
## About The Pull Request

Fix of https://github.com/tgstation/tgstation/issues/91881
What has changed? Now you can't set custom colors for channels that have
color settings in the chat based on the theme (freqspans), but the color
will be displayed correctly in different themes

It is still possible to change the channel name for all types (custom
and standard). It is also possible to change the color of your channels
(without support for the light theme)

I hope this will be a temporary solution, as I would like to return full
customization of channels based on the user's chosen theme.

## Why It's Good For The Game

Fans of the light theme will be pleased.

## Changelog

:cl: FeudeyTF
fix: added light theme support for standard radio channels
/:cl: