Adds the Ability to Toggle Whether Collars Appear Above or Below Suits (#5700)
## About The Pull Request
Adds the ability for people to toggle whether their collars appear above
or below suits (like lab coats) by alt + right clicking on them and adds
a tip to their descriptions informing of such a feature!
## How This Contributes To The Nova Sector Roleplay Experience
Just lets people who want to wear a collar choose whether they want to
be a bit more subtle or open about it!
## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>
Works and doesnt go under jumpsuits!


https://github.com/user-attachments/assets/70da54fc-78f4-4738-b937-54ac785efc60


For all varieties of collars!


https://github.com/user-attachments/assets/ed171d42-27a0-4426-a658-853f5d070bac



With a handy little tip in the description

![dreamseeker_EkwHwG1B3v](https://github.com/user-attachments/assets/724321c0-4212-46e7-b9f9-322f78557c5f)

</details>



## Changelog
:cl:
add: Collars are now able to be hidden beneath suits, coats and jackets.
Alt+Right Click to toggle!
/:cl: