`md5asfile` now uses rust-g to get a file hash (#91952)
## About The Pull Request

This removes the `md5filepath` proc, which was only used in `md5asfile`.

Instead, `md5asfile` will just directly use `rustg_hash_file`, so we
avoid having to read the file that we *just* wrote to disk right back
into BYOND to MD5 it, and we can use rust-g to hash that file instead.

in theory, this should be even more performant if
https://github.com/tgstation/rust-g/pull/228 is ~~merged~~ released

## Why It's Good For The Game

avoids an unneeded round-trip and hopefully improves performance.

## Changelog

no user-facing changes