the default linux `PreCompile.sh` now copies .so files instead of moving them (#91990)
## About The Pull Request

by changing the `mv` commands in the `PreCompile.sh` preset for rust-g
and dreamluau to `cp -f`, we can avoid having to re-link the so files
every compile if nothing has changed, meaning that this step is
practically instant if the rust-g and compiler versions were the same as
last time.

tested downstream and this does not seem to cause any issues in
production

## Why It's Good For The Game

Improves compile times.

## Changelog

no player-facing changes