Visitor to the station quirk! for assistant (#5593)
## About The Pull Request
Adds a quirk which only works for assistants, it gives you an ugly ID
and removes you from the crew manifest, only giving the records system
limited data about who you are.
The PR also adds a blacklist to the clock out app, including assistants
and prisoners.

## How This Contributes To The Nova Sector Roleplay Experience

Having players be able to join with limited records deepens the
engagement of roles who handle paperwork during the round, encouraging
beginning the records yourself. It makes jobs like head of personnel,
security, medical and silicons take different routes when engaging with
players, without the elephant in the room of non-crew equals antag.
That said, this buffs antags who get agent IDs, as they can use it to
fake out a visitor ID and blend in much easier. I think this is
appropriate, because taking on your mission covertly is apt for the
server, and the chameleon approach is simply the most treaded path.


![image](https://github.com/user-attachments/assets/45579d84-9ca9-414f-a304-f03c395a441a)

## Proof of Testing


https://github.com/user-attachments/assets/52a5e340-51ff-43d1-ae57-e6db2560e0f4


![image](https://github.com/user-attachments/assets/47a5c3a9-84c7-41f5-b4f6-70fc5ef1bd00)

## Changelog
:cl:
add: Added the visitor quirk! Join with limited records data and no
automatic manifest injection
qol: Assistants and prisoners can't clock out anymore. No more 'Off-duty
off-duty' crew
change: Records lacking data for a name or job will not appear on the
crew manifest
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>