Adds a secondary shuttle dock to snowglobe (#5715)
## About The Pull Request

I got pinged that snowglobe lacks a way for an ert/traitor/nonstandard
shuttle to land in the map so this PR adds one. It's a secondary dock
south of arrivals. This should Just Work but who knows maybe I'm getting
sloppy from how lax Solaris is to map for.

## How This Contributes To The Nova Sector Roleplay Experience

Done by request for someone designing a new antag that needs it. Also
the map should just have it period for admin events and what not.

## Proof of Testing

ingame
![Screenshot 2025-07-06
150141](https://github.com/user-attachments/assets/960a658b-10ea-4a0a-bd99-58af9dcd3f71)

## Changelog

:cl:
add: Added a secondary shuttle dock to snowglobe.
/:cl:

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>