Some BRM Balancing i meant to do when i first balanced them a year ago (#5734)
## About The Pull Request

Been meaning to do this for over a year but this just slightly lowers
their point generation.

The alternative option is to slow down their boulder production, but i
feel that'd be a bit overly impactful for the moment at least. the Vent
system is designed for quicker shifts than what we standardize as, and
while i wish this was a dynamic adjustment based on default roundtime -
that's up to TG to handle that maths.

points_held = round(points_held + (quantity *
possible_mat.points_per_unit * MINING_POINT_MACHINE_MULTIPLIER))

**Old**
100 = round ( 100 + (23 * 15 * 0.8)
376 points

**New**
100 = round ( 100 + (23 * 15 * 0.6)
307 points

_These are not actual values but rather a representation of the math,
each boulder has their own value assigned as well as quantity_

|276−207|[(276+207)2]×100
|69|[4832]×100
69241.5×100
0.285714×100
=28.5714%difference

if BYOND math was perfect i'd have it at .55 but it wont math past .5
annoyingly due to how the decimal comes into play, but this is a ~30%
decrease in points across the round. This math is the baseline of an
unupgraded BRM.

## How This Contributes To The Nova Sector Roleplay Experience

we cant really kid ourselves with how many mining points a few vents can
pump out in a long shift, dropping it from 0.8 to 0.6 on the multiplier
is what my calculator says is the balance that'd work out.

I did add a 50% increase to the points discovered/completion so the
bonus for doing them is pretty substantial now

## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>
  
Glass and Iron doin work

![image](https://github.com/user-attachments/assets/785a4844-cea6-498a-b980-be693ca41c43)


</details>

## Changelog
:cl:
balance: slightly lowers the points per boulder of the BRM
/:cl: