Extends Integer scaling range, corrects some info (#92130)
## About The Pull Request

Extends the max range for integer scaling up to 9x, corrects some
misinfo in the visuals guide

## Why It's Good For The Game

This allows proper integer scaling up to 8k resolution, which not only
is generally the highest consumer-available resolution currently, but
also the next resolution that is able to perfectly display a 480p image
on it via integer scaling.

Also corrects misinfo because misinfo is bad.

## Changelog

:cl:
qol: Increased the integer scaling range up to 9x
/:cl: