Stunsword (Real) (#5780)
## About The Pull Request
Ports the stunsword, fished up from the commits in
https://github.com/tgstation/tgstation/pull/88830
Credits to NecromancerAnne for the idea/code, SmArtKar and DrDiaysel for
the sprites, as noted in the original PR before it was removed.
## How This Contributes To The Nova Sector Roleplay Experience
funny
## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>

uh yeah dude gimme a sec  
</details>

## Changelog

:cl:
add: Heads of Security are now being issued the NT-20 "Excalibur" stun
sword (not to be confused with the NT-20 personal defense weapon, issued
to Blueshields).
/:cl:

---------

Co-authored-by: Hatterhat <Hatterhat@users.noreply.github.com>