[NO GBP] Fixes oil ignition, sanity checks blood DNA and adds some logging to it (#92221)
## About The Pull Request

Oil blood will no longer try to ignite twice when hit with a welder,
runtiming on the second try.
Also added crashes when ``add_blood_DNA`` is passed non-list or lists
with null assoc values, as both are invalid and cause crashes, because I
need to hunt down a stupid bug.

## Changelog
:cl:
fix: Trying to ignite oil with a welding tool no longer runtimes due to
it trying to get ignited twice
/:cl: