Climbing hooks; usage limit removed, can now climb up OR down, Syndicate climbing hook is much much faster than normal ones (#92191)
## About The Pull Request

This does as the title says; climbing hooks are no longer limited-use,
and have had their functionality updated to where you can use them to
climb both UP and DOWN. As well, the Syndicate climbing hook has a
climbing speed of 0.75 seconds per movement, in contrast to the 4 second
speed of the emergency climbing hook.
## Why It's Good For The Game
Climbing hooks are a cool idea for people experimenting with multi-z,
either in the game or in map design, but for some reason we gimped them
with a hard and stingy usage limit. The syndicate climbing hook having
an extremely fast speed enables you to do cat burglar activities before
scuttling away and cackling evilly.

## Changelog
:cl: Bisar
add: Climbing hooks can now be used to climb UP or DOWN, instead of only
up.
qol: Climbing hooks no longer have a limit on their number of uses.
balance: Syndicate climbing hooks now have a speed of 0.75 seconds to
climb, in comparison to the standard 4 seconds for emergency hooks.
fix: Fixed a runtime in climbing hooks when you would try to climb down.
/:cl: