Removes wetstacks from the refitted wetsuit (#5846)
## About The Pull Request
Removes the component that adds wetstacks from the refitted wetsuit
## How This Contributes To The Nova Sector Roleplay Experience
There is a specific item in the loadout selection that does the exact
same thing for any suit, this component doesn't need to be on the
refitted wetsuit.
## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>
  
</details>

## Changelog
:cl:
balance: The refitted wetsuit no longer gives wetstacks, use the
stardress hydro-vaporizer suit accessory if you want wetstacks.
/:cl: