Lets IPCs actually hallucinate with the RDS quirk (#5896)
## About The Pull Request

Ports https://github.com/IrisSS13/IrisStation/pull/538. One of our
lovely contributors noticed that RDS didn't work on synthetics. Unsure
if there's a better way of doing this but it adds a trait to the quirk

## How This Contributes To The Nova Sector Roleplay Experience

Allows synthetics to actually use a feature of the quirk intended for
them (Sensory Processing Fault).

## Proof of Testing

We tested it on Iris, worked well.

## Changelog

:cl: WhereAmO
fix: IPCs with the RDS quirk can now hallucinate again.
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>