[BUFF + QOL] Crusher Feedback With Blocking + New Proto-Kinetic Machete Changes (#5883)
## About The Pull Request

When adding the ashen bones to the kinetic crusher and its variants,
there was absolutely no feedback to blocking. It was just a visual and
that was it, but it'd probably be more satisfying if a player actually
heard that (applies to the machete as well).

The machete can now be used both one-handed and two-handed! A bugfix
that I have decided to tweak. Before, the machete dealt NO damage when
held with one hand, and only detonated marks. Now, it will do a force of
10 one handed, and a force of 15 when wielded properly. When sharpening
with a whetstone, you will not get the full benefits of it until you
wield the machete properly.

## How This Contributes To The Nova Sector Roleplay Experience

Audio feedback is good! It also stays consistent with weapons that have
block_chance and block_sound. Additionally, the machete not doing damage
when wielded in hand was really strange, and the implication that
something suddenly doesn't hurt despite being a sharp weapon is really
strange and immersion breaking.

## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>
Unwielded vs Wielded
<img width="488" height="77" alt="image"
src="https://github.com/user-attachments/assets/9f8df10c-97f8-4a74-876e-8501acb9bb04"
/>

<img width="493" height="85" alt="image"
src="https://github.com/user-attachments/assets/7b3dcbe4-74f6-45ab-b97d-01c2ab4ed41d"
/>

Sound: Trust me it works

  
</details>

## Changelog
:cl:BasilTamaya
add: The machete now has unique unwielded and wielded damage values.
qol: Machete now works properly and as intended.
balance: The machete now has a special use case that makes it unique
from the rest of the crushers, compared to everything else that needs to
be wielded in order to use them. Now properly fits its description: A
compact crusher.
sound: Added block sounds for the crushers. Whether it's the Marked One
trophy, or the machete, players now receive proper audio feedback when
blocking.
/:cl: