Burned mess from culinary disasters is once again toxic. (#92394)
## About The Pull Request
When an item is cooked, reagents are generally cleared from the
resulting object and the reagents of the source object are transferred
to it instead. This means burned mess won't have its toxins when cooked,
locking you out of one or two chemistry recipes that require it while
also making the item generally not toxic at all.
So to fix it, we simply have to replace all or most of the consumable
reagents inside the burned mess with bad food, via component signals.

## Why It's Good For The Game
Fixing an issue with food and chemistry.

## Changelog

:cl:
fix: burned mess made with a microwave, oven or griddle once again
toxins.
/:cl:

---------

Co-authored-by: _0Steven <42909981+00-Steven@users.noreply.github.com>