bounty app for PDA (#91802)
## About The Pull Request

This PR adds a brand new app to the PDA allowing you to view your
current bounties and even generate new ones.

You can even select a bounty remotely and it will seemlessly work.

The program is not without its own drawbacks. Choosing to reroll your
bounties from the app will always trigger a 5 minute cooldown, and this
cannot be reduced. Meaning that rerolling directly from the bounty
terminal (assuming it has been upgraded) is preferable as an upgraded
bounty pad will give the terminal a shorter cooldown. The cooldown is
shared between the app and the terminal, and depends on where you
initially rerolled it. So if you rerolled it from the app, it will
always be 5 minutes, but if you rerolled it from the terminal, it may be
as low as 2 minutes.

Otherwise it is identical to the way the bounty terminal functions,
minus the fact you cannot actually turn in the bounty using your PDA,
you still have to go to a terminal with connected bounty pad for that!

And in addition to that, in order for the app to work you must have your
ID inside your PDA, the app will not function without it (the top button
will be greyed out, and trying to push one of the options will just give
you a minor error sound)

Size is 5 GQ since that seemed like a good enough size for it.

## Why It's Good For The Game

It always annoyed me that I had to run all the way back to cargo to
check on what my bounties are, and this solves that. It also lets you
reroll your bounties remotely with a punishment of it always having the
default (5 min) waiting period between rerolls.

## Changelog

:cl:
add: Adds a new bounty app to the PDA allowing you to remotely check
your bounties
/:cl:

![image](https://github.com/user-attachments/assets/b0a10106-1c6d-4f9e-91b1-f95f8cc377cb)

![image](https://github.com/user-attachments/assets/c2cc8075-7314-48b1-a668-e7f8a44bc217)

![image](https://github.com/user-attachments/assets/9536b8cb-8c95-47ed-ba42-2a8d162123c7)

![image](https://github.com/user-attachments/assets/eb254331-35f0-413a-9b74-77c5f9ad6ccb)

---------

Co-authored-by: John Willard <53777086+JohnFulpWillard@users.noreply.github.com>