*jump emote will not display to runechat (#88464)
## About The Pull Request
Just like spin and flip, it will not display a runechat and chat
message.
## Why It's Good For The Game
there's plenty of indication that you jumped, like a sound and an
animation. Printing a message just spams the chat and runechat above
you, especially when you spam it.
## Changelog
:cl: grungussuss
qol: *jump emote will not print a message
/:cl: