Updates the AFK logout report to utilize configs for it's timers (#92662)
Rather than use hardcoded timers for the logout report, the report
instead utilizes a timer based on config values. The default value is
the current live value.

This timer can also be configurated to utilize randomization. Currently,
it defaults to randomizing between -3 minutes to +3 minutes for when the
report is sent.

The logout report timer is also added to the admin log.

This shouldn't be dictated by the codebase, as it is an admin tool.