Fixes an oversight with suit accessory overlays (#92687)
Just because an accessory has `above_suit` set to `TRUE` does not mean
the suit has an `accessory_overlay`.

You wouldn't wanna add a null to the overlays list to return here so
make sure we actually have one first.

(It turns out, it is actually silently doing this every time you first
attach or remove an accessory-when an accessory is added, it calls
`successful_attach()` which ends up calling `update_clothing()` which
calls `worn_overlays()`--all _before_ creating the `accessory_overlay`
-- which results in a `null` in the returned list of `worn_overlays()`)

Shifts the ordering around so that this race condition no longer occurs.