The speed of tools used in a crafting recipe now affect the time spent on it (#92577)
Title. Using better tools should now make crafting faster (or lower)
depending on the quality of the tools. This affects roughly 85% of the
crafting time, so, for every second of it, you're guaranteed to spend
0.15 seconds even if your tools are fast as light itself.

Also a nit with the main benefit of the mining skill being locked to
humans despite it being bound to the mind datum.

Tool speed previously didn't play a role in crafting. Now it does.

:cl:
balance: The speed of tools used in a crafting recipe now affect the
time spent on it. Report any recipe that takes awfully little or too
long to complete.
balance: Non-humanoid mobs can now benefit from being skilled at mining
when swinging that pickaxe.
/:cl: