Clarifies space ninja as hired by Cybersun & changes their objectives (#6023)
## About The Pull Request

The space ninja is now clearly designated as working for Cybersun to the
ninja player, though references to the spider clan as independent
faction still remains.

Their intention is to assassinate command members and **behead** them!
The cyborg hijacking objective remains, but is optional and suggests it
is to assist in the primary objective.
If there is no command available, the goon server theft objective will
still be granted.

They have lost their special C4, and the ability to fuck with security
records. Good riddance.

Lastly, removes the non-modular edits on the space ninja objective. This
is now overwritten in a modular file.

## How This Contributes To The Nova Sector Roleplay Experience

The spinja can do a lot of things that the server doesn't want. Tedious
records console button pressing, overbearing structural damage. Ugh.
Players who find themselves in the ninja's shoes often need to avoid
some objectives or request extra permission. If we change the objectives
to something better. Something that is more likely to create RP, then
perhaps you'll see the antag be enjoyed by more of the community. The
removed content can be gained by forging custom objectives/talking to
admins after all.

[Spider clan](https://wiki.novasector13.com/wiki/Lore:Spider_Clan) and
Cybersun aren't the same yet the former is a faction hired by the
latter. I reworded the flavortext of the space ninja to be less "I am
teh ninja!1!" and more informative to the player about who they actually
are helping.

## Proof of Testing


https://github.com/user-attachments/assets/f11e75a7-1989-486c-ae47-711fd90dbd7f

## Changelog

:cl:
add: The space ninja has gotten a primary objective, brutally
assassinating command and spreading fear among the upper ranks.
fix: The space ninja's explosive implant is now *actually* gone
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>