Readds medical sprays to the nanodrug plus (#92785)
## About The Pull Request

Reverts a bit of https://github.com/tgstation/tgstation/pull/92134 - the
change was not documented, and the creator expressed neutrality at it's
return in discord.
## Why It's Good For The Game

Diluted aiuri and libital sprays are good stopgaps for when chemists
aren't making chems, or at roundstart, when your options for chems are
the limited medkits. They're not exactly good at their job, they're
impure, limited, but they get the job done.

The change is also a bit out of scope for that PR (it wanted to change
emergency vendors and not the nanodrug) and - again - undocumented.

Feel free to close if this is actually bad for the game, I just really
like using these as a doctor when better chems aren't available.
## Changelog
:cl:
add: 2 aiuri and 2 libital cans to nanodrug plus
/:cl: