Recolourable Skirt Overalls (#6089)
now with 100% less sweater.

In all seriousness I saw the skirt overalls added as part of #5835 and
really wanted to use them, but I didn't want to have to use a turtleneck
so using their sprite I threw together a version to match the regular
overalls that takes the suit slot.

More clothes more better right?
<details>
<summary>Screenshots/Videos</summary>

https://github.com/user-attachments/assets/59554e9a-6786-4d77-a0c5-586aa41d7cb3

https://github.com/user-attachments/assets/7f32e9f0-5268-4507-a5c9-1986ab16e309

</details>

:cl:
add: A colourable skirted overall in the suit section of your loadout.
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>