[QoL] [Staff Request] Make the Shuttle transit times consistent (15 minutes), regardless of Alert (#5331)
## About The Pull Request
Changed the time mod on all the alert levels on a modular file to be
1.5, which makes it go to 15 minutes given the current tg config. (Is
not wise to force it on a higher level, given the shuttle code.)

## How This Contributes To The Nova Sector Roleplay Experience
Requested by Deadmon.

It's also I believe a very good change, since it gives consistency with
the shuttle, doesnt force HoS and other command to change the alert
levels based on the round end, but based on the real danger level. It
allows everyone to know that roughly if everything goes well a round
will last the same, from finish to end, barring round delays.

## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>
  

![image](https://github.com/user-attachments/assets/74958641-0e0a-4dcb-be2a-435d3dd4a742)


</details>

## Changelog
:cl:
qol: In compliance with SolFed safety regulations, we have disabled the
overdrive from our escape shuttles, forcing us to replace the
hyperdrives with TX2 models. These are roughly 30% faster to prepare,
but safety meassures cannot be skipped from the startup, this means
shuttles will take 12 minutes from 17 minutes to prepare, and will
arrive 3 minutes after preparation, setting the time for our shuttles to
arrive from CC to the station from 20 to 15 minutes, in all alert
levels.
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>