Adds several holopads to wawastation (#93131)
## About The Pull Request

Adds several holopads to wawastation, in places precedented by other
maps

## Why It's Good For The Game

There's not enough holopads on wawastation. There's currently one in all
of science, and it's in the RD's office. Not even one in robotics.
There's five in engi, but four of them are in locations rarely visited.
There isn't a holopad in the HOP's, CE's, or HOS's offices. Only two
holopads in security. Holopads are fun, and the coverage is lacking.

## Changelog
:cl:

map: increases holopad coverage on wawastation

/:cl: