Fix for stacks transforming into other stacks when removed from the Protolathe/Techfab. (#93260)
## About The Pull Request

I *think* this should be all that's required to fix this, it looks like
it's just the variable that iterates over all available stacks on the
dropturf needed to compare against the iterator's type as opposed to the
stack we're about to create otherwise.

Fixes #93256 

## Why It's Good For The Game

Prevents a tiiiiny little material transformation/duplication bug.

## Changelog

:cl:
fix: Sheets removed from the protolathe/techfab will properly remove the
correct quantity and merge into the right material.
/:cl: