Gets rid of the oranges meow again (#6206)
## About The Pull Request
Gets rid of the 1% oranges meow again, it was re-added by #5863. Just
removes the line actually adding the 1% chance for the oranges meow.
## How This Contributes To The Nova Sector Roleplay Experience
It was gone before and should have been for the last two months.

The sound sucks big time, too.
## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>
  
</details>

## Changelog
:cl:
fix: [NOVA] The 1% chance for the Oranges meow sound to play is gone
again.
/:cl: