Turns the chaplain's energy swords into...energy swords. (#93394)
## About The Pull Request

The chaplain's energy swords were weird fake energy swords that couldn't
be turned on or off. This is largely because, in the past, we couldn't
easily convert other objects into null rods conveniently. Though we did
get the transforming component, it never was added to the chaplain
weapons. Maybe because there would be a lot of copypasted code.

Now, with the ``nullrod_core`` element, we can easily turn an object
into a null rod. So I did that. It's now an energy sword AND a null rod.
The future is now.

For clarity, the stats on the null rods are identical to before. The
only difference being that it has the ability to switch on and off, and
it now emits light. That's it really. If there are any extra stats I
missed, let me know.

Also adds the carp'sie plushie to the list of wearable objects for
chaplain suit storage.

## Why It's Good For The Game

This has bugged me like, for a really, really long time. And we now have
a lot of the bits and pieces to make more of the null rods actually
closer to the objects they're ripping off.

## Changelog
:cl:
qol: The chaplain energy swords now work like energy swords. Turn them
on and off! Light your way with the blade. Light a cigarette with your
blade like a badass. Feel bad that it isn't a REAL energy sword.
fix: The carp'sie plushie now fits in chaplain suit storage.
/:cl: