knuckle duster brand new detective gear and it's traitor version (#6232)
Adds:
beacon to the detective
overrides regular batong. instead allows to choice 1) unique (very)
piece of gear
The knuckleduster that gives evil boxing and allows to punch lethally or
not lethally with damage slightly bigger than from punch!

and it's tartor version to uplink which is slighlty more evil and
beefier
https://youtu.be/PtvqZtWv_M8

:cl:
add: Knuckleduster and it's variant to uplink!
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>