Fixes PKC mark underlays not respecting mob offsets and pixel_w/z (#93561)
## About The Pull Request
Mobs should offset using ``pixel_w/z`` (not always for y-z), not
``pixel_x/y``, and crushers weren't respecting that. They also checked
for current offsets rather than base, meaning that the underlay always
stayed in the same place at the center of the tile, or offset to the
side if the mob used w/z offsets

## Changelog
:cl:
fix: Fixed PKC mark effect being offset to the side when applied to some
mobs.
/:cl: