True Strike is now also an armour piercing munition (#93642)
## About The Pull Request
So it dawned on me for awhile, we did get true strike and I got to use
it way more, not necessarily after it is available as speedloader but
I've used it for atleast a bit, realized just buying the match mag was
better anyway.

so, true strike sucks for actually bouncing and killing people. and
while 15 damage is nothing to scoff at because any damage is better than
doing no damage at all. It's already well outclassed by it's actual
roundstart counterpart. This fix that, the truestrike ammo is researched
fairly late into the round. Giving it AP make it more useful
## Why It's Good For The Game

I think it make senses we'd want a researched ammo type more effective
than something you can just. Buy? It's not like we don't have tech laser
that are more effective than buckshot for raw damage. This is similar to
that idea
## Changelog
:cl:
add: True Strike is also now AP, making detective more scary and BR-38
much more effective against highly armoured target
/:cl: