Fixes tactical maid headband (#6352)
## About The Pull Request

Fixes half of https://github.com/NovaSector/NovaSector/issues/6338


https://github.com/NovaSector/NovaSector/commit/1f50a0e639f593a48bcd7e4c49befb50fc108dc7
seems to have repathed this item to be a subtype of the tg version,
making it inherit the greyscale configs of the parent type. But it's not
a greyscale item and uses completely different non GAGS icons from the
TG one.

## How This Contributes To The Nova Sector Roleplay Experience

Bugfix

## Proof of Testing

<details>
<summary>Screenshots/Videos</summary>
  
</details>

## Changelog

:cl:
fix: fixes the tactical maid headband having messed up icons
/:cl: