Heretic gateway overhaul (#5650)
This PR completely fixes and overhauls the heretic away mission which
was a proof of concept and will be overhauled to actually be
comprehensive and not a dogshit map. THANK YOU TO BUBBER FOR HAVING MADE
MODULAR TRAMS BEFORE HELPING ME MAKE MY MODULAR TRAMS SPECIFICALLY LT3

Adds more to the gateways

I’ve tested it on a local host it’s okay and it works as intended.

<details>
<summary>Screenshots/Videos</summary>

</details>

:cl:
map: Reworked the heretic gateway to be good and not terrible.
added: CBRN armor found in the heretic gateway (no sprite)
added: a P90 and other weapons to be found in the heretic gateway.
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>