Fixes Bitrunners Not Having ORM Access on Blueshift (#6403)
## About The Pull Request

swaps the mining to general cargo access for the room with the ORM.

## How This Contributes To The Nova Sector Roleplay Experience

bitrunners need to drop their stuff off too, apparently. nerds. they
don't get mining access unless skeleton crew, and having the room gated
off is a bit odd, so we're swapping those around a bit.

## Proof of Testing

n/a

## Changelog
:cl:
fix: bitrunners can now get to the orm on blueshift.
/:cl: