Fixes "N OVA" edit (#6417)
# https://github.com/tgstation/tgstation/pull/93877

> *Adds `volume_preference` to `sound_to_playing_players` and uses it to
tie security level looping sounds to the ambience volume*

So this can be test merged or something down here while waiting for that
PR to make it in? I dunno. This is a result of a dev-suggest, but this
problem could appear on TG if they ever get a security level with
looping sounds, so I'm trying to upstream it.

## How This Contributes To The Nova Sector Roleplay Experience
Our sounds for this get crazy annoying. I don't think they should
continue existing due to their quality, but let's start with this.
## Proof of Testing
Works, but it's hard to send a clip of it working due to GitHub's file
size limits. Tested with an ambience volume of 0 and then one of 50.
<details>
<summary>Screenshots/Videos</summary>
  
</details>

## Changelog
### The relevant TG PR has a changelog, and this PR shouldn't be merged.

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>