Small fixes/oversights to the maintsrooms gateway map. (#5914)
## About The Pull Request

This PR adds nothing to the maintsrooms besides making a few area's
secrets more obvious so they're found along with halving the number of
mobs that spawn, I mean it runs fine as is I've seen it played 3 times
now on the main server but I couldn't live with myself if I didn't fix
these issues that I missed in debug mode.

:cl:
fix: halved the numbers of mobs in the maintsrooms.
fix: A myriad of smaller oversights with the maintsrooms, reworked the
materials
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>