Fixes Pun Pun qdel on Monkey Day (#94463) (#6657)
## ORIGINAL PR:
https://github.com/NovaSector/NovaSector/commit/221dc61a01c3c8bfd323d032c76571d0e22e39c4

Fixes https://github.com/tgstation/tgstation/issues/88502

## About The Pull Request

The issue with the current code is that it was returning
`INITIALIZE_HINT_QDEL` without going through the whole carbon initialize
chain, making it such that we were trying to qdel() the typepath of the
`ai_controller` or the `bodypart`s instead of actual datums since those
variables pull double duty. I don't think pun pun code should be
lackadaisical when it comes to the carbon initialize chain, so I just
shifted around a bunch of stuff to retain the same functionality while
also making it that we do a full intialize. The tech debt of one monkey
shall be infinite.
## Why It's Good For The Game

Less errors on monkey day = ideal.
## Changelog

I don't think this affected the holiday itself so not really
player-facing

Co-authored-by: san7890 <the@san7890.com>