Readds soul to shocks (#94615)
Replaces the `Paralyze` after being shocked with a `Stun`.
(For those unaware the only difference between these effects is
`Paralyze` forces you to the floor, `Stun` does not.)
You still get `Paralyze`d after a few seconds, which forces you to the
floor.

If you play with RDS, you may be familiar with the shock hallucination
that keeps you upright for a bit before flooring you.
That's what normal shocking looks like now.

I also added some extra logic to revert back to using a `Paralyze` if
you are already stunned, so two shocks in quick succession floors you
faster.

Functionally it is largely identical, so this ultimately isn't a major
change.
I think this is how shocking was supposed to look, before the PR that
added `Paralyze()` and friends?
Either way I just think it looks more fun.

:cl: Melbert
balance: When you are shocked, you remain standing for a short period of
time before being forced to the ground. You are still stunned for the
duration of this effect, that remains unchanged.
/:cl: