AI deploy to shell verb does not take an input (#94664)
## About The Pull Request

Fixes #94663 
It seems that using the "Deploy to Shell" verb as an AI will launch a
popup asking which mob in view you want to trigger the verb as, which is
obviously unintended behaviour.
I've resolved this by moving all of the behaviour to a proc which the
verb calls, so existing code which passes in a mob should still function
as it did previously.

There might be a different way to do this but I'll be honest I try not
to ever add verbs so I don't know what it is.

## Why It's Good For The Game

When you use a verb from the verb panel you implicitly expect to be the
mob using the verb, not to have to select yourself from a list first.

## Changelog

:cl:
fix: AI shell deployment verb no longer requires you to select yourself
from a list first
/:cl: